% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ram.R
\name{A}
\alias{A}
\title{A Matrix}
\usage{
A(taudot, beta, alpha)
}
\arguments{
\item{taudot}{Numeric.
Slope of path from \code{x} to \code{y} \eqn{\left( \dot{\tau} \right)}.}

\item{beta}{Numeric.
Slope of path from \code{m} to \code{y} \eqn{\left( \beta \right)} .}

\item{alpha}{Numeric.
Slope of path from \code{x} to \code{m} \eqn{\left( \alpha \right)} .}
}
\description{
Matrix of regression slopes from the simple mediation model.
}
\details{
The simple mediation model is given by
\deqn{
    y_i = \delta_y + \dot{\tau} x_i + \beta m_i + \varepsilon_{y_{i}}
  }

\deqn{
    m_i = \delta_m + \alpha x_i + \varepsilon_{m_{i}}
  }

The parameters for the mean structure are
\deqn{
    \boldsymbol{\theta}_{\text{mean structure}} = \left\{ \mu_x, \delta_m, \delta_y \right\} .
  }

The parameters for the covariance structure are
\deqn{
    \boldsymbol{\theta}_{\text{covariance structure}} = \left\{ \dot{\tau}, \beta, \alpha, \sigma_{x}^{2},
    \sigma_{\varepsilon_{m}}^{2}, \sigma_{\varepsilon_{y}}^{2} \right\} .
  }
}
\examples{
A(
  taudot = 0.207648,
  beta = 0.451039,
  alpha = 0.338593
)
}
\seealso{
Other reticular action model functions: 
\code{\link{A.std}()},
\code{\link{Mfrommu}()},
\code{\link{M}()},
\code{\link{S.std}()},
\code{\link{Sfromsigma2}()},
\code{\link{Sigmatheta.std}()},
\code{\link{Sigmathetafromsigma2}()},
\code{\link{Sigmatheta}()},
\code{\link{S}()},
\code{\link{mutheta}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{reticular action model functions}
\keyword{ram}
