% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ram.R
\name{A.std}
\alias{A.std}
\title{Standardized A Matrix}
\usage{
A.std(taudotprime, betaprime, alphaprime, lambdax, lambdam, lambday)
}
\arguments{
\item{taudotprime}{Numeric.
Standardized slope of path from \code{x} to \code{y} \eqn{\left( \dot{\tau}^{\prime} \right)}.}

\item{betaprime}{Numeric.
Standardized slope of path from \code{m} to \code{y} \eqn{\left( \beta^{\prime} \right)} .}

\item{alphaprime}{Numeric.
Standardized slope of path from \code{x} to \code{m} \eqn{\left( \alpha^{\prime} \right)} .}

\item{lambdax}{Numeric.
Factor loading \code{xlatent ~ x} \eqn{ \left( \lambda_x \right)} .
Numerically equivalent to the standard deviation of \code{x}.}

\item{lambdam}{Numeric.
Factor loading \code{mlatent ~ m} \eqn{ \left( \lambda_m \right)} .
Numerically equivalent to the standard deviation of \code{m}.}

\item{lambday}{Numeric.
Factor loading \code{ylatent ~ y} \eqn{ \left( \lambda_y \right)} .
Numerically equivalent to the standard deviation of \code{y}.}
}
\description{
Matrix of regression slopes from the standardized simple mediation model.
}
\examples{
A.std(
  taudotprime = 0.2080748,
  betaprime = 0.4126006,
  alphaprime = 0.3708979,
  lambdax = 1.137308,
  lambdam = 1.038248,
  lambday = 1.134973
)
}
\seealso{
Other reticular action model functions: 
\code{\link{A}()},
\code{\link{Mfrommu}()},
\code{\link{M}()},
\code{\link{S.std}()},
\code{\link{Sfromsigma2}()},
\code{\link{Sigmatheta.std}()},
\code{\link{Sigmathetafromsigma2}()},
\code{\link{Sigmatheta}()},
\code{\link{S}()},
\code{\link{mutheta}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{reticular action model functions}
\keyword{ram}
