% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci.R
\name{bcaci}
\alias{bcaci}
\title{Bias-Corrected and Accelerated Confidence Intervals}
\usage{
bcaci(
  thetahatstar,
  thetahatstarjack = NULL,
  thetahat,
  theta = NULL,
  data,
  std = FALSE,
  alpha = c(0.001, 0.01, 0.05),
  par = TRUE,
  ncores = NULL,
  blas_threads = TRUE,
  mc = TRUE,
  lb = FALSE
)
}
\arguments{
\item{thetahatstar}{Numeric vector.
Sampling distribution of thetahat.}

\item{thetahatstarjack}{Numeric vector.
Jackknife vector of parameter estimates.
If \code{thetahatstarjack = NULL},
\code{thetahatstarjack} is calculated using \code{\link[=.jack]{.jack()}}.}

\item{thetahat}{Numeric.
Parameter estimate.}

\item{theta}{Numeric.
Parameter.
Optional argument.}

\item{data}{\code{n} by 3 matrix or data frame.
\code{data[, 1]} correspond to values for \code{x}.
\code{data[, 2]} correspond to values for \code{m}.
\code{data[, 3]} correspond to values for \code{y}.}

\item{std}{Logical.
Standardize the indirect effect
\eqn{
    \hat{\alpha}^{\prime} \hat{\beta}^{\prime}
    = \hat{\alpha} \hat{\beta} \frac{\hat{\sigma}_x}{\hat{\sigma}_y}}.}

\item{alpha}{Numeric vector.
Alpha level.
By default \code{alpha = c(0.001, 0.01, 0.05)}.}

\item{par}{Logical.
If \code{TRUE}, use multiple cores.
If \code{FALSE}, use \code{\link[=lapply]{lapply()}}.}

\item{ncores}{Integer.
Number of cores to use if \code{par = TRUE}.
If unspecified, defaults to \code{detectCores() - 1}.}

\item{blas_threads}{Logical.
If \code{TRUE}, set BLAS threads using \code{blas_set_num_threads(threads = blas_get_num_procs())}.
If \code{FALSE}, set BLAS threads using \code{blas_set_num_threads(threads = 1)}.
If \code{par = TRUE}, \code{blas_threads} is automatically set to \code{FALSE}
to prevent conflicts in parallel processing.
This argument is useful when \code{FUN} can handle implicit parallelism
when \code{par = FALSE}, for example linear algebra operations.}

\item{mc}{Logical.
If \code{TRUE}, use \code{\link[parallel:mclapply]{parallel::mclapply()}}.
If \code{FALSE}, use \code{\link[parallel:clusterApply]{parallel::parLapply()}} or \code{\link[parallel:clusterApply]{parallel::parLapplyLB()}}.
Ignored if \code{par = FALSE}.}

\item{lb}{Logical.
If \code{TRUE} use \code{\link[parallel:clusterApply]{parallel::parLapplyLB()}}.
If \code{FALSE}, use \code{\link[parallel:clusterApply]{parallel::parLapply()}}.
Ignored if \code{par = FALSE} and \code{mc = TRUE}.}
}
\description{
Bias-Corrected and Accelerated Confidence Intervals
}
\examples{
B <- 5000
data <- jeksterslabRdatarepo::thirst
thetahat <- .fit(data, minimal = TRUE)
thetahatstar <- .pbmvn(data = data, B = 5000, par = FALSE)
bcaci(
  thetahatstar = thetahatstar,
  thetahat = thetahat,
  theta = 0.15, # assuming that the true indirect effect is 0.15
  data = data,
  par = FALSE
)
thetahat <- .fit(data, minimal = TRUE, std = TRUE)
thetahatstar <- .pbmvn(data = data, std = TRUE, B = 5000, par = FALSE)
bcaci(
  thetahatstar = thetahatstar,
  thetahat = thetahat,
  theta = 0.15, # assuming that the true indirect effect is 0.15
  data = data,
  std = TRUE,
  par = FALSE
)
}
\seealso{
Other confidence intervals functions: 
\code{\link{bcci}()},
\code{\link{evalci}()},
\code{\link{len}()},
\code{\link{pcci}()},
\code{\link{shape}()},
\code{\link{theta_hit}()},
\code{\link{zero_hit}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{confidence intervals functions}
\keyword{ci}
