% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_complete_dat.R
\name{beta_dat}
\alias{beta_dat}
\title{Generate Data with Beta X
following a Simple Mediation Model}
\usage{
beta_dat(taskid)
}
\arguments{
\item{taskid}{Numeric.
Task ID.}
}
\description{
Generate Data with Beta X
following a Simple Mediation Model
}
\examples{
data <- beta_dat(taskid = 1)
fit.ols(data, minimal = TRUE)
fit.ols(data, minimal = TRUE, std = TRUE)
fit.ols(data, minimal = FALSE)
}
\seealso{
Other beta X data functions: 
\code{\link{beta_dat_simulation}()},
\code{\link{beta_dat_task}()},
\code{\link{genbeta}()},
\code{\link{setparamsbeta}()},
\code{\link{useparamsbeta}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{beta X data functions}
\keyword{beta_dat}
