% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_complete_dat.R
\name{beta_dat_task}
\alias{beta_dat_task}
\title{Generate Data with Beta X
following a Simple Mediation Model
(Single Task)}
\usage{
beta_dat_task(taskid, dir = getwd(), overwrite = FALSE)
}
\arguments{
\item{taskid}{Numeric.
Task ID.}

\item{dir}{Character string.
Directory where results of \verb{beta_dat_*} are stored.}

\item{overwrite}{Logical.
Overwrite existing results.}
}
\description{
Generate Data with Beta X
following a Simple Mediation Model
(Single Task)
}
\seealso{
Other beta X data functions: 
\code{\link{beta_dat_simulation}()},
\code{\link{beta_dat}()},
\code{\link{genbeta}()},
\code{\link{setparamsbeta}()},
\code{\link{useparamsbeta}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{beta X data functions}
\keyword{beta_dat}
