% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{.fit}
\alias{.fit}
\title{Fit Simple Mediation Model - Minimal}
\usage{
.fit(data, minimal = TRUE, std = FALSE)
}
\arguments{
\item{data}{\code{n} by 3 matrix or data frame.
\code{data[, 1]} correspond to values for \code{x}.
\code{data[, 2]} correspond to values for \code{m}.
\code{data[, 3]} correspond to values for \code{y}.}

\item{minimal}{Logical.
If \code{TRUE}, only returns the estimate of the indirect effect
\eqn{\left( \hat{\alpha} \hat{\beta} \right)}.
If \code{FALSE}, returns more information.}

\item{std}{Logical.
Standardize the indirect effect
\eqn{
    \hat{\alpha}^{\prime} \hat{\beta}^{\prime}
    = \hat{\alpha} \hat{\beta} \frac{\hat{\sigma}_x}{\hat{\sigma}_y}}.}
}
\description{
Fits the simple mediation model and returns the indirect effect.
}
\details{
The fitted simple mediation model is given by
\deqn{
    y_i
    =
     \hat{\delta}_{y} + \hat{\dot{\tau}} x_i + \hat{\beta} m_i +
     \hat{\varepsilon}_{y_{i}}
  }

\deqn{
    m_i = \hat{\delta}_{m} + \hat{\alpha} x_i + \hat{\varepsilon}_{m_{i}}
  }

The estimated parameters for the mean structure are
\deqn{
    \boldsymbol{\hat{\theta}}_{\text{mean structure}}
    =
     \left\{ \hat{\mu}_{x}, \hat{\delta}_{m}, \hat{\delta}_{y} \right\} .
  }

The estimated parameters for the covariance structure are
\deqn{
    \boldsymbol{\hat{\theta}}_{\text{covariance structure}}
    =
     \left\{ \hat{\dot{\tau}}, \hat{\beta}, \hat{\alpha},
     \hat{\sigma}_{x}^{2}, \hat{\sigma}_{\hat{\varepsilon}_{m}}^{2},
     \hat{\sigma}_{\hat{\varepsilon}_{y}}^{2} \right\} .
  }
}
\examples{
data <- jeksterslabRdatarepo::thirst
.fit(data, minimal = TRUE)
.fit(data, minimal = TRUE, std = TRUE)
.fit(data, minimal = FALSE)
}
\seealso{
Other model fit functions: 
\code{\link{.fitsem}()},
\code{\link{fit}()},
\code{\link{mvn_fit_simulation_summary}()},
\code{\link{mvn_fit_simulation}()},
\code{\link{mvn_fit_task_summary}()},
\code{\link{mvn_fit_task}()},
\code{\link{mvn_fit}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{model fit functions}
\keyword{fit}
