% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.fiml.R
\name{.fitfiml}
\alias{.fitfiml}
\title{Fit Simple Mediation Model - Structural Equation Modeling with FIML}
\usage{
.fitfiml(data, std = FALSE)
}
\arguments{
\item{data}{\code{n} by 3 matrix or data frame.
\code{data[, 1]} correspond to values for \code{x}.
\code{data[, 2]} correspond to values for \code{m}.
\code{data[, 3]} correspond to values for \code{y}.}

\item{std}{Logical.
If \code{TRUE}, estimate standardized simple mediation model
using latent variables and nonlinear constraints.}
}
\description{
Fit Simple Mediation Model - Structural Equation Modeling with FIML
}
\seealso{
Other model fit functions: 
\code{\link{.fitsem}()},
\code{\link{.fit}()},
\code{\link{fit}()},
\code{\link{mvn_fitfiml_mar_simulation}()},
\code{\link{mvn_fitfiml_mar_task}()},
\code{\link{mvn_fitfiml_mar}()},
\code{\link{mvn_fitfiml_mcar_simulation}()},
\code{\link{mvn_fitfiml_mcar_task}()},
\code{\link{mvn_fitfiml_mcar}()},
\code{\link{mvn_fitols_simulation_summary}()},
\code{\link{mvn_fitols_simulation}()},
\code{\link{mvn_fitols_task_summary}()},
\code{\link{mvn_fitols_task}()},
\code{\link{mvn_fitols}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{model fit functions}
\keyword{fit}
