% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jack.R
\name{.jack}
\alias{.jack}
\title{Jackknife Estimates of Indirect Effect for a Simple Mediation Model}
\usage{
.jack(
  data,
  std = FALSE,
  complete = TRUE,
  par = TRUE,
  ncores = NULL,
  blas_threads = TRUE,
  mc = TRUE,
  lb = FALSE
)
}
\arguments{
\item{data}{\code{n} by 3 matrix or data frame.
\code{data[, 1]} correspond to values for \code{x}.
\code{data[, 2]} correspond to values for \code{m}.
\code{data[, 3]} correspond to values for \code{y}.}

\item{std}{Logical.
Standardize the indirect effect
\eqn{
    \hat{\alpha}^{\prime} \hat{\beta}^{\prime}
    = \hat{\alpha} \hat{\beta} \frac{\hat{\sigma}_x}{\hat{\sigma}_y}}.}

\item{complete}{Logical.
Listwise deletion to generate data with complete cases.}

\item{par}{Logical.
If \code{TRUE}, use multiple cores.
If \code{FALSE}, use \code{\link[=lapply]{lapply()}}.}

\item{ncores}{Integer.
Number of cores to use if \code{par = TRUE}.
If unspecified, defaults to \code{detectCores() - 1}.}

\item{blas_threads}{Logical.
If \code{TRUE}, set BLAS threads using \code{blas_set_num_threads(threads = blas_get_num_procs())}.
If \code{FALSE}, set BLAS threads using \code{blas_set_num_threads(threads = 1)}.
If \code{par = TRUE}, \code{blas_threads} is automatically set to \code{FALSE}
to prevent conflicts in parallel processing.
This argument is useful when \code{FUN} can handle implicit parallelism
when \code{par = FALSE}, for example linear algebra operations.}

\item{mc}{Logical.
If \code{TRUE}, use \code{\link[parallel:mclapply]{parallel::mclapply()}}.
If \code{FALSE}, use \code{\link[parallel:clusterApply]{parallel::parLapply()}} or \code{\link[parallel:clusterApply]{parallel::parLapplyLB()}}.
Ignored if \code{par = FALSE}.}

\item{lb}{Logical.
If \code{TRUE} use \code{\link[parallel:clusterApply]{parallel::parLapplyLB()}}.
If \code{FALSE}, use \code{\link[parallel:clusterApply]{parallel::parLapply()}}.
Ignored if \code{par = FALSE} and \code{mc = TRUE}.}
}
\description{
Jackknife Estimates of Indirect Effect for a Simple Mediation Model
}
\examples{
data <- jeksterslabRdatarepo::thirst
.jack(data, par = FALSE)
.jack(data, std = TRUE, par = FALSE)
}
\seealso{
Other jackknife functions: 
\code{\link{mvn_jack_mar_simulation}()},
\code{\link{mvn_jack_mar_task}()},
\code{\link{mvn_jack_mar}()},
\code{\link{mvn_jack_mcar_simulation}()},
\code{\link{mvn_jack_mcar_task}()},
\code{\link{mvn_jack_mcar}()},
\code{\link{mvn_jack_simulation}()},
\code{\link{mvn_jack_task}()},
\code{\link{mvn_jack}()},
\code{\link{mvn_std_jack_simulation}()},
\code{\link{mvn_std_jack_task}()},
\code{\link{mvn_std_jack}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{jackknife functions}
\keyword{jack}
