% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_complete_dat.R
\name{exp_dat}
\alias{exp_dat}
\title{Generate Data with Exponential X
following a Simple Mediation Model}
\usage{
exp_dat(taskid)
}
\arguments{
\item{taskid}{Numeric.
Task ID.}
}
\description{
Generate Data with Exponential X
following a Simple Mediation Model
}
\examples{
data <- exp_dat(taskid = 1)
fit.ols(data, minimal = TRUE)
fit.ols(data, minimal = TRUE, std = TRUE)
fit.ols(data, minimal = FALSE)
}
\seealso{
Other exponential X data functions: 
\code{\link{exp_dat_simulation}()},
\code{\link{exp_dat_task}()},
\code{\link{genexp}()},
\code{\link{setparamsexp}()},
\code{\link{useparamsexp}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{exponential X data functions}
\keyword{exp_dat}
