% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_complete_dat.R
\name{exp_dat_task}
\alias{exp_dat_task}
\title{Generate Data with Exponential X
following a Simple Mediation Model
(Single Task)}
\usage{
exp_dat_task(taskid, dir = getwd(), overwrite = FALSE)
}
\arguments{
\item{taskid}{Numeric.
Task ID.}

\item{dir}{Character string.
Directory where results of \verb{exp_dat_*} are stored.}

\item{overwrite}{Logical.
Overwrite existing results.}
}
\description{
Generate Data with Exponential X
following a Simple Mediation Model
(Single Task)
}
\seealso{
Other exponential X data functions: 
\code{\link{exp_dat_simulation}()},
\code{\link{exp_dat}()},
\code{\link{genexp}()},
\code{\link{setparamsexp}()},
\code{\link{useparamsexp}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{exponential X data functions}
\keyword{exp_dat}
