% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_complete_unstd_jack.R
\name{exp_jack_task}
\alias{exp_jack_task}
\title{Jackknife Estimates of Indirect Effect in a Simple Mediation Model for Data with Exponential X (Single Task)}
\usage{
exp_jack_task(taskid, dir = getwd(), overwrite = FALSE)
}
\arguments{
\item{taskid}{Numeric.
Task ID.}

\item{dir}{Character string.
Directory where results of \verb{exp_dat_*} are stored.}

\item{overwrite}{Logical.
Overwrite existing results.}
}
\description{
Jackknife Estimates of Indirect Effect in a Simple Mediation Model for Data with Exponential X (Single Task)
}
\seealso{
Other jackknife functions: 
\code{\link{beta_jack_simulation}()},
\code{\link{beta_jack_task}()},
\code{\link{beta_jack}()},
\code{\link{exp_jack_simulation}()},
\code{\link{exp_jack}()},
\code{\link{jack}()},
\code{\link{mvn_jack_simulation}()},
\code{\link{mvn_jack_task}()},
\code{\link{mvn_jack}()},
\code{\link{mvn_std_jack_simulation}()},
\code{\link{mvn_std_jack_task}()},
\code{\link{mvn_std_jack}()},
\code{\link{vm_mod_jack_simulation}()},
\code{\link{vm_mod_jack_task}()},
\code{\link{vm_mod_jack}()},
\code{\link{vm_mod_std_jack_simulation}()},
\code{\link{vm_mod_std_jack_task}()},
\code{\link{vm_mod_std_jack}()},
\code{\link{vm_sev_jack_simulation}()},
\code{\link{vm_sev_jack_task}()},
\code{\link{vm_sev_jack}()},
\code{\link{vm_sev_std_jack_simulation}()},
\code{\link{vm_sev_std_jack_task}()},
\code{\link{vm_sev_std_jack}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{jackknife functions}
\keyword{jack}
