% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.sem.mlr.R
\name{fit.sem.mlr}
\alias{fit.sem.mlr}
\title{Fit Simple Mediation Model - Structural Equation Modeling}
\usage{
fit.sem.mlr(data, minimal = FALSE, std = FALSE, fiml = FALSE)
}
\arguments{
\item{data}{\code{n} by 3 matrix or data frame.
\code{data[, 1]} correspond to values for \code{x}.
\code{data[, 2]} correspond to values for \code{m}.
\code{data[, 3]} correspond to values for \code{y}.}

\item{minimal}{Logical.
If \code{TRUE}, only returns the estimate of the indirect effect
\eqn{\left( \hat{\alpha} \hat{\beta} \right)}.
If \code{FALSE}, returns more information.}

\item{std}{Logical.
If \code{TRUE}, estimate standardized simple mediation model
using latent variables and nonlinear constraints.}

\item{fiml}{Logical.
If \code{TRUE}, use \code{missing = "fiml"} to handle missing values.
Note that using \code{missing = "fiml"} sets \code{fixed.x = FALSE}.}
}
\description{
Fits the simple mediation model using structural equation modeling.
}
\examples{
library(lavaan)
summary(fit.sem.mlr(data = jeksterslabRdatarepo::thirst))
summary(fit.sem.mlr(data = jeksterslabRdatarepo::thirst, std = TRUE))
summary(fit.sem.mlr(data = jeksterslabRdatarepo::thirst, minimal = TRUE))
summary(fit.sem.mlr(data = jeksterslabRdatarepo::thirst, minimal = TRUE, std = TRUE))

taskid <- 1
data_mod <- vm_mod_dat(taskid = taskid)
data_sev <- vm_sev_dat(taskid = taskid)

# Moderate ----------------------------------------------------
summary(fit.sem.mlr(data = data_mod))

# Moderate ----------------------------------------------------
summary(fit.sem.mlr(data = data_sev))
}
\seealso{
Other model fit functions: 
\code{\link{fit.cov}()},
\code{\link{fit.ols}()},
\code{\link{fit.sem}()},
\code{\link{mvn_fit.ols_simulation_summary}()},
\code{\link{mvn_fit.ols_simulation}()},
\code{\link{mvn_fit.ols_task_summary}()},
\code{\link{mvn_fit.ols_task}()},
\code{\link{mvn_fit.ols}()},
\code{\link{mvn_fit.sem_simulation_summary}()},
\code{\link{mvn_fit.sem_simulation}()},
\code{\link{mvn_fit.sem_task_summary}()},
\code{\link{mvn_fit.sem_task}()},
\code{\link{mvn_fit.sem}()},
\code{\link{mvn_mar_10_fit.sem_simulation_summary}()},
\code{\link{mvn_mar_10_fit.sem_simulation}()},
\code{\link{mvn_mar_10_fit.sem_task_summary}()},
\code{\link{mvn_mar_10_fit.sem_task}()},
\code{\link{mvn_mar_10_fit.sem}()},
\code{\link{mvn_mar_20_fit.sem_simulation_summary}()},
\code{\link{mvn_mar_20_fit.sem_simulation}()},
\code{\link{mvn_mar_20_fit.sem_task_summary}()},
\code{\link{mvn_mar_20_fit.sem_task}()},
\code{\link{mvn_mar_20_fit.sem}()},
\code{\link{mvn_mar_30_fit.sem_simulation_summary}()},
\code{\link{mvn_mar_30_fit.sem_simulation}()},
\code{\link{mvn_mar_30_fit.sem_task_summary}()},
\code{\link{mvn_mar_30_fit.sem_task}()},
\code{\link{mvn_mar_30_fit.sem}()},
\code{\link{mvn_mcar_10_fit.sem_simulation_summary}()},
\code{\link{mvn_mcar_10_fit.sem_simulation}()},
\code{\link{mvn_mcar_10_fit.sem_task_summary}()},
\code{\link{mvn_mcar_10_fit.sem_task}()},
\code{\link{mvn_mcar_10_fit.sem}()},
\code{\link{mvn_mcar_20_fit.sem_simulation_summary}()},
\code{\link{mvn_mcar_20_fit.sem_simulation}()},
\code{\link{mvn_mcar_20_fit.sem_task_summary}()},
\code{\link{mvn_mcar_20_fit.sem_task}()},
\code{\link{mvn_mcar_20_fit.sem}()},
\code{\link{mvn_mcar_30_fit.sem_simulation_summary}()},
\code{\link{mvn_mcar_30_fit.sem_simulation}()},
\code{\link{mvn_mcar_30_fit.sem_task_summary}()},
\code{\link{mvn_mcar_30_fit.sem_task}()},
\code{\link{mvn_mcar_30_fit.sem}()},
\code{\link{mvn_mnar_10_fit.sem_simulation_summary}()},
\code{\link{mvn_mnar_10_fit.sem_simulation}()},
\code{\link{mvn_mnar_10_fit.sem_task_summary}()},
\code{\link{mvn_mnar_10_fit.sem_task}()},
\code{\link{mvn_mnar_10_fit.sem}()},
\code{\link{mvn_mnar_20_fit.sem_simulation_summary}()},
\code{\link{mvn_mnar_20_fit.sem_simulation}()},
\code{\link{mvn_mnar_20_fit.sem_task_summary}()},
\code{\link{mvn_mnar_20_fit.sem_task}()},
\code{\link{mvn_mnar_20_fit.sem}()},
\code{\link{mvn_mnar_30_fit.sem_simulation_summary}()},
\code{\link{mvn_mnar_30_fit.sem_simulation}()},
\code{\link{mvn_mnar_30_fit.sem_task_summary}()},
\code{\link{mvn_mnar_30_fit.sem_task}()},
\code{\link{mvn_mnar_30_fit.sem}()},
\code{\link{mvn_std_fit.sem_simulation_summary}()},
\code{\link{mvn_std_fit.sem_simulation}()},
\code{\link{mvn_std_fit.sem_task_summary}()},
\code{\link{mvn_std_fit.sem_task}()},
\code{\link{mvn_std_fit.sem}()},
\code{\link{vm_mod_fit.ols_simulation_summary}()},
\code{\link{vm_mod_fit.ols_simulation}()},
\code{\link{vm_mod_fit.ols_task_summary}()},
\code{\link{vm_mod_fit.ols_task}()},
\code{\link{vm_mod_fit.ols}()},
\code{\link{vm_mod_fit.sem.mlr_simulation_summary}()},
\code{\link{vm_mod_fit.sem.mlr_simulation}()},
\code{\link{vm_mod_fit.sem.mlr_task_summary}()},
\code{\link{vm_mod_fit.sem.mlr_task}()},
\code{\link{vm_mod_fit.sem.mlr}()},
\code{\link{vm_mod_std_fit.sem.mlr_simulation_summary}()},
\code{\link{vm_mod_std_fit.sem.mlr_simulation}()},
\code{\link{vm_mod_std_fit.sem.mlr_task_summary}()},
\code{\link{vm_mod_std_fit.sem.mlr_task}()},
\code{\link{vm_mod_std_fit.sem.mlr}()},
\code{\link{vm_sev_fit.ols_simulation_summary}()},
\code{\link{vm_sev_fit.ols_simulation}()},
\code{\link{vm_sev_fit.ols_task_summary}()},
\code{\link{vm_sev_fit.ols_task}()},
\code{\link{vm_sev_fit.ols}()},
\code{\link{vm_sev_fit.sem.mlr_simulation_summary}()},
\code{\link{vm_sev_fit.sem.mlr_simulation}()},
\code{\link{vm_sev_fit.sem.mlr_task_summary}()},
\code{\link{vm_sev_fit.sem.mlr_task}()},
\code{\link{vm_sev_fit.sem.mlr}()},
\code{\link{vm_sev_std_fit.sem.mlr_simulation_summary}()},
\code{\link{vm_sev_std_fit.sem.mlr_simulation}()},
\code{\link{vm_sev_std_fit.sem.mlr_task_summary}()},
\code{\link{vm_sev_std_fit.sem.mlr_task}()},
\code{\link{vm_sev_std_fit.sem.mlr}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{model fit functions}
\keyword{fit}
