% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.sem.R
\name{fit.sem}
\alias{fit.sem}
\title{Fit Simple Mediation Model - Structural Equation Modeling}
\usage{
fit.sem(data, minimal = FALSE, std = FALSE, fiml = FALSE)
}
\arguments{
\item{data}{\code{n} by 3 matrix or data frame.
\code{data[, 1]} correspond to values for \code{x}.
\code{data[, 2]} correspond to values for \code{m}.
\code{data[, 3]} correspond to values for \code{y}.}

\item{minimal}{Logical.
If \code{TRUE}, only returns the estimate of the indirect effect
\eqn{\left( \hat{\alpha} \hat{\beta} \right)}.
If \code{FALSE}, returns more information.}

\item{std}{Logical.
If \code{TRUE}, estimate standardized simple mediation model
using latent variables and nonlinear constraints.}

\item{fiml}{Logical.
If \code{TRUE}, use \code{missing = "fiml"} to handle missing values.
Note that using \code{missing = "fiml"} sets \code{fixed.x = FALSE}.}
}
\description{
Fits the simple mediation model using structural equation modeling.
}
\examples{
library(lavaan)
data <- jeksterslabRdatarepo::thirst

# OLS --------------------------------------------------------
fit.ols(data, minimal = FALSE)

# SEM Unstandardized -----------------------------------------
unstd_sem <- fit.sem(data = data)
coef(unstd_sem)
sqrt(diag(vcov(unstd_sem)))
fit.sem(data = data, minimal = TRUE)

# SEM Standardized -------------------------------------------
std_sem <- fit.sem(data = data, std = TRUE)
coef(std_sem)
sqrt(diag(vcov(std_sem)))
fit.sem(data = data, minimal = TRUE, std = TRUE)
}
\seealso{
Other model fit functions: 
\code{\link{fit.cov}()},
\code{\link{fit.ols}()},
\code{\link{mvn_fit.ols_simulation_summary}()},
\code{\link{mvn_fit.ols_simulation}()},
\code{\link{mvn_fit.ols_task_summary}()},
\code{\link{mvn_fit.ols_task}()},
\code{\link{mvn_fit.ols}()},
\code{\link{mvn_fit.sem_simulation_summary}()},
\code{\link{mvn_fit.sem_simulation}()},
\code{\link{mvn_fit.sem_task_summary}()},
\code{\link{mvn_fit.sem_task}()},
\code{\link{mvn_fit.sem}()},
\code{\link{mvn_mar_10_fit.sem_simulation_summary}()},
\code{\link{mvn_mar_10_fit.sem_simulation}()},
\code{\link{mvn_mar_10_fit.sem_task_summary}()},
\code{\link{mvn_mar_10_fit.sem_task}()},
\code{\link{mvn_mar_10_fit.sem}()},
\code{\link{mvn_mar_20_fit.sem_simulation_summary}()},
\code{\link{mvn_mar_20_fit.sem_simulation}()},
\code{\link{mvn_mar_20_fit.sem_task_summary}()},
\code{\link{mvn_mar_20_fit.sem_task}()},
\code{\link{mvn_mar_20_fit.sem}()},
\code{\link{mvn_mar_30_fit.sem_simulation_summary}()},
\code{\link{mvn_mar_30_fit.sem_simulation}()},
\code{\link{mvn_mar_30_fit.sem_task_summary}()},
\code{\link{mvn_mar_30_fit.sem_task}()},
\code{\link{mvn_mar_30_fit.sem}()},
\code{\link{mvn_mar_50_fit.sem_simulation_summary}()},
\code{\link{mvn_mar_50_fit.sem_simulation}()},
\code{\link{mvn_mar_50_fit.sem_task_summary}()},
\code{\link{mvn_mar_50_fit.sem_task}()},
\code{\link{mvn_mar_50_fit.sem}()},
\code{\link{mvn_mar_fit.sem_simulation_summary}()},
\code{\link{mvn_mar_fit.sem_simulation}()},
\code{\link{mvn_mar_fit.sem_task_summary}()},
\code{\link{mvn_mar_fit.sem_task}()},
\code{\link{mvn_mar_fit.sem}()},
\code{\link{mvn_mar_std_fit.sem_simulation_summary}()},
\code{\link{mvn_mar_std_fit.sem_simulation}()},
\code{\link{mvn_mar_std_fit.sem_task_summary}()},
\code{\link{mvn_mar_std_fit.sem_task}()},
\code{\link{mvn_mar_std_fit.sem}()},
\code{\link{mvn_mcar_10_fit.sem_simulation_summary}()},
\code{\link{mvn_mcar_10_fit.sem_simulation}()},
\code{\link{mvn_mcar_10_fit.sem_task_summary}()},
\code{\link{mvn_mcar_10_fit.sem_task}()},
\code{\link{mvn_mcar_10_fit.sem}()},
\code{\link{mvn_mcar_20_fit.sem_simulation_summary}()},
\code{\link{mvn_mcar_20_fit.sem_simulation}()},
\code{\link{mvn_mcar_20_fit.sem_task_summary}()},
\code{\link{mvn_mcar_20_fit.sem_task}()},
\code{\link{mvn_mcar_20_fit.sem}()},
\code{\link{mvn_mcar_30_fit.sem_simulation_summary}()},
\code{\link{mvn_mcar_30_fit.sem_simulation}()},
\code{\link{mvn_mcar_30_fit.sem_task_summary}()},
\code{\link{mvn_mcar_30_fit.sem_task}()},
\code{\link{mvn_mcar_30_fit.sem}()},
\code{\link{mvn_mcar_50_fit.sem_simulation_summary}()},
\code{\link{mvn_mcar_50_fit.sem_simulation}()},
\code{\link{mvn_mcar_50_fit.sem_task_summary}()},
\code{\link{mvn_mcar_50_fit.sem_task}()},
\code{\link{mvn_mcar_50_fit.sem}()},
\code{\link{mvn_mcar_fit.sem_simulation_summary}()},
\code{\link{mvn_mcar_fit.sem_simulation}()},
\code{\link{mvn_mcar_fit.sem_task_summary}()},
\code{\link{mvn_mcar_fit.sem_task}()},
\code{\link{mvn_mcar_fit.sem}()},
\code{\link{mvn_mcar_std_fit.sem_simulation_summary}()},
\code{\link{mvn_mcar_std_fit.sem_simulation}()},
\code{\link{mvn_mcar_std_fit.sem_task_summary}()},
\code{\link{mvn_mcar_std_fit.sem_task}()},
\code{\link{mvn_mcar_std_fit.sem}()},
\code{\link{mvn_mnar_10_fit.sem_simulation_summary}()},
\code{\link{mvn_mnar_10_fit.sem_simulation}()},
\code{\link{mvn_mnar_10_fit.sem_task_summary}()},
\code{\link{mvn_mnar_10_fit.sem_task}()},
\code{\link{mvn_mnar_10_fit.sem}()},
\code{\link{mvn_mnar_20_fit.sem_simulation_summary}()},
\code{\link{mvn_mnar_20_fit.sem_simulation}()},
\code{\link{mvn_mnar_20_fit.sem_task_summary}()},
\code{\link{mvn_mnar_20_fit.sem_task}()},
\code{\link{mvn_mnar_20_fit.sem}()},
\code{\link{mvn_mnar_30_fit.sem_simulation_summary}()},
\code{\link{mvn_mnar_30_fit.sem_simulation}()},
\code{\link{mvn_mnar_30_fit.sem_task_summary}()},
\code{\link{mvn_mnar_30_fit.sem_task}()},
\code{\link{mvn_mnar_30_fit.sem}()},
\code{\link{mvn_mnar_50_fit.sem_simulation_summary}()},
\code{\link{mvn_mnar_50_fit.sem_simulation}()},
\code{\link{mvn_mnar_50_fit.sem_task_summary}()},
\code{\link{mvn_mnar_50_fit.sem_task}()},
\code{\link{mvn_mnar_50_fit.sem}()},
\code{\link{mvn_mnar_fit.sem_simulation_summary}()},
\code{\link{mvn_mnar_fit.sem_simulation}()},
\code{\link{mvn_mnar_fit.sem_task_summary}()},
\code{\link{mvn_mnar_fit.sem_task}()},
\code{\link{mvn_mnar_fit.sem}()},
\code{\link{mvn_mnar_std_fit.sem_simulation_summary}()},
\code{\link{mvn_mnar_std_fit.sem_simulation}()},
\code{\link{mvn_mnar_std_fit.sem_task_summary}()},
\code{\link{mvn_mnar_std_fit.sem_task}()},
\code{\link{mvn_mnar_std_fit.sem}()},
\code{\link{mvn_std_fit.sem_simulation_summary}()},
\code{\link{mvn_std_fit.sem_simulation}()},
\code{\link{mvn_std_fit.sem_task_summary}()},
\code{\link{mvn_std_fit.sem_task}()},
\code{\link{mvn_std_fit.sem}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{model fit functions}
\keyword{fit}
