% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta.R
\name{genbeta}
\alias{genbeta}
\title{Beta}
\usage{
genbeta(taskid)
}
\arguments{
\item{taskid}{Numeric.
Task ID.}
}
\description{
Beta
}
\examples{
data <- genbeta(taskid = 1)
fit.ols(data)
}
\seealso{
Other beta X data functions: 
\code{\link{beta_dat_simulation}()},
\code{\link{beta_dat_task}()},
\code{\link{beta_dat}()},
\code{\link{setparamsbeta}()},
\code{\link{useparamsbeta}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{beta X data functions}
\keyword{beta}
