% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp.R
\name{genexp}
\alias{genexp}
\title{Exponential}
\usage{
genexp(taskid)
}
\arguments{
\item{taskid}{Numeric.
Task ID.}
}
\description{
Exponential
}
\examples{
data <- genexp(taskid = 1)
fit.ols(data)
}
\seealso{
Other exponential X data functions: 
\code{\link{exp_dat_simulation}()},
\code{\link{exp_dat_task}()},
\code{\link{exp_dat}()},
\code{\link{setparamsexp}()},
\code{\link{useparamsexp}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{exponential X data functions}
\keyword{exp}
