% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval.R
\name{len}
\alias{len}
\title{Confidence Interval - Length}
\usage{
len(lo, up)
}
\arguments{
\item{lo}{Numeric.
Lower limit of the estimated confidence interval
\eqn{\left( \hat{\theta}_{\mathrm{lo}} \right)}.}

\item{up}{Numeric.
Upper limit of the estimated confidence interval
\eqn{\left( \hat{\theta}_{\mathrm{up}} \right)}.}
}
\description{
Calculates confidence interval length.
}
\details{
The confidence interval length is given by
\deqn{
  \mathrm{
    confidence \ interval \ length
  }
  =
  \hat{
    \theta
  }_{
    \mathrm{
      up
    }
  }
  -
  \hat{
    \theta
  }_{
    \mathrm{
      lo
    }
  }
}
}
\seealso{
Other confidence intervals functions: 
\code{\link{bcaci}()},
\code{\link{bcci}()},
\code{\link{evalci}()},
\code{\link{pcci}()},
\code{\link{shape}()},
\code{\link{theta_hit}()},
\code{\link{zero_hit}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{confidence intervals functions}
\keyword{ci}
