% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ram.R
\name{M}
\alias{M}
\title{M Vector}
\usage{
M(mux, deltam, deltay)
}
\arguments{
\item{mux}{Numeric.
Mean of \code{x} \eqn{\left( \mu_x \right)} .}

\item{deltam}{Numeric.
Intercept of \code{m} \eqn{\left( \delta_m \right)} .}

\item{deltay}{Numeric.
Intercept of \code{y} \eqn{\left( \delta_y \right)} .}
}
\description{
Mean of \code{x} and intercepts from the simple mediation model.
}
\examples{
M(
  mux = 70.18000,
  deltam = 26.82246,
  deltay = 29.91071
)
}
\seealso{
Other reticular action model functions: 
\code{\link{A.std}()},
\code{\link{A}()},
\code{\link{Mfrommu}()},
\code{\link{S.std}()},
\code{\link{Sfromsigma2}()},
\code{\link{Sigmatheta.std}()},
\code{\link{Sigmathetafromsigma2}()},
\code{\link{Sigmatheta}()},
\code{\link{S}()},
\code{\link{mutheta}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{reticular action model functions}
\keyword{ram}
