% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc.R
\name{mc.mvn}
\alias{mc.mvn}
\title{Monte Carlo Method for Indirect Effect in a Simple Mediation Model
Assuming \eqn{\alpha} and \eqn{\beta} Follow a Multivariate Normal Distribution (Sampling Distribution)}
\usage{
mc.mvn(R = 20000L, alphahat, sehatalphahat, betahat, sehatbetahat)
}
\arguments{
\item{R}{Integer.
Monte Carlo replications.}

\item{alphahat}{Numeric.
Estimated slope of path from \code{x} to \code{m} \eqn{\left( \hat{\alpha} \right)} .}

\item{sehatalphahat}{Numeric.
Estimated standard error of slope of path from \code{x} to \code{m} \eqn{\left( \widehat{se}_{\hat{\alpha}} \right)} .}

\item{betahat}{Numeric.
Estimated slope of path from \code{m} to \code{y} \eqn{\left( \hat{\beta} \right)} .}

\item{sehatbetahat}{Numeric.
Estimated standard error of slope of path from \code{m} to \code{y} \eqn{\left( \widehat{se}_{\hat{\beta}} \right)} .}
}
\description{
In this method \eqn{\alpha} and \eqn{\beta} are assumed to follow
a multivariate normal distribution.
}
\examples{
thetahatstar <- mc.mvn(
  R = 20000L, alphahat = 0.338593, sehatalphahat = 0.12236736,
  betahat = 0.451039, sehatbetahat = 0.14597405
)
hist(thetahatstar)
}
\seealso{
Other monte carlo method functions: 
\code{\link{mc.t}()},
\code{\link{mvn_mar_10_mc.mvn_pcci_simulation}()},
\code{\link{mvn_mar_10_mc.mvn_pcci_task}()},
\code{\link{mvn_mar_10_mc.mvn_simulation}()},
\code{\link{mvn_mar_10_mc.mvn_task}()},
\code{\link{mvn_mar_10_mc.mvn}()},
\code{\link{mvn_mar_20_mc.mvn_pcci_simulation}()},
\code{\link{mvn_mar_20_mc.mvn_pcci_task}()},
\code{\link{mvn_mar_20_mc.mvn_simulation}()},
\code{\link{mvn_mar_20_mc.mvn_task}()},
\code{\link{mvn_mar_20_mc.mvn}()},
\code{\link{mvn_mar_30_mc.mvn_pcci_simulation}()},
\code{\link{mvn_mar_30_mc.mvn_pcci_task}()},
\code{\link{mvn_mar_30_mc.mvn_simulation}()},
\code{\link{mvn_mar_30_mc.mvn_task}()},
\code{\link{mvn_mar_30_mc.mvn}()},
\code{\link{mvn_mar_50_mc.mvn_pcci_simulation}()},
\code{\link{mvn_mar_50_mc.mvn_pcci_task}()},
\code{\link{mvn_mar_50_mc.mvn_simulation}()},
\code{\link{mvn_mar_50_mc.mvn_task}()},
\code{\link{mvn_mar_50_mc.mvn}()},
\code{\link{mvn_mar_mc.mvn_pcci_simulation}()},
\code{\link{mvn_mar_mc.mvn_pcci_task}()},
\code{\link{mvn_mar_mc.mvn_simulation}()},
\code{\link{mvn_mar_mc.mvn_task}()},
\code{\link{mvn_mar_mc.mvn}()},
\code{\link{mvn_mar_std_mc.mvn.sem_pcci_simulation}()},
\code{\link{mvn_mar_std_mc.mvn.sem_pcci_task}()},
\code{\link{mvn_mar_std_mc.mvn.sem_simulation}()},
\code{\link{mvn_mar_std_mc.mvn.sem_task}()},
\code{\link{mvn_mar_std_mc.mvn.sem}()},
\code{\link{mvn_mc.mvn_pcci_simulation}()},
\code{\link{mvn_mc.mvn_pcci_task}()},
\code{\link{mvn_mc.mvn_simulation}()},
\code{\link{mvn_mc.mvn_task}()},
\code{\link{mvn_mc.mvn}()},
\code{\link{mvn_mc.t_pcci_simulation}()},
\code{\link{mvn_mc.t_pcci_task}()},
\code{\link{mvn_mc.t_simulation}()},
\code{\link{mvn_mc.t_task}()},
\code{\link{mvn_mc.t}()},
\code{\link{mvn_mcar_10_mc.mvn_pcci_simulation}()},
\code{\link{mvn_mcar_10_mc.mvn_pcci_task}()},
\code{\link{mvn_mcar_10_mc.mvn_simulation}()},
\code{\link{mvn_mcar_10_mc.mvn_task}()},
\code{\link{mvn_mcar_10_mc.mvn}()},
\code{\link{mvn_mcar_20_mc.mvn_pcci_simulation}()},
\code{\link{mvn_mcar_20_mc.mvn_pcci_task}()},
\code{\link{mvn_mcar_20_mc.mvn_simulation}()},
\code{\link{mvn_mcar_20_mc.mvn_task}()},
\code{\link{mvn_mcar_20_mc.mvn}()},
\code{\link{mvn_mcar_30_mc.mvn_pcci_simulation}()},
\code{\link{mvn_mcar_30_mc.mvn_pcci_task}()},
\code{\link{mvn_mcar_30_mc.mvn_simulation}()},
\code{\link{mvn_mcar_30_mc.mvn_task}()},
\code{\link{mvn_mcar_30_mc.mvn}()},
\code{\link{mvn_mcar_50_mc.mvn_pcci_simulation}()},
\code{\link{mvn_mcar_50_mc.mvn_pcci_task}()},
\code{\link{mvn_mcar_50_mc.mvn_simulation}()},
\code{\link{mvn_mcar_50_mc.mvn_task}()},
\code{\link{mvn_mcar_50_mc.mvn}()},
\code{\link{mvn_mcar_mc.mvn_pcci_simulation}()},
\code{\link{mvn_mcar_mc.mvn_pcci_task}()},
\code{\link{mvn_mcar_mc.mvn_simulation}()},
\code{\link{mvn_mcar_mc.mvn_task}()},
\code{\link{mvn_mcar_mc.mvn}()},
\code{\link{mvn_mcar_std_mc.mvn.sem_pcci_simulation}()},
\code{\link{mvn_mcar_std_mc.mvn.sem_pcci_task}()},
\code{\link{mvn_mcar_std_mc.mvn.sem_simulation}()},
\code{\link{mvn_mcar_std_mc.mvn.sem_task}()},
\code{\link{mvn_mcar_std_mc.mvn.sem}()},
\code{\link{mvn_mnar_10_mc.mvn_pcci_simulation}()},
\code{\link{mvn_mnar_10_mc.mvn_pcci_task}()},
\code{\link{mvn_mnar_10_mc.mvn_simulation}()},
\code{\link{mvn_mnar_10_mc.mvn_task}()},
\code{\link{mvn_mnar_10_mc.mvn}()},
\code{\link{mvn_mnar_20_mc.mvn_pcci_simulation}()},
\code{\link{mvn_mnar_20_mc.mvn_pcci_task}()},
\code{\link{mvn_mnar_20_mc.mvn_simulation}()},
\code{\link{mvn_mnar_20_mc.mvn_task}()},
\code{\link{mvn_mnar_20_mc.mvn}()},
\code{\link{mvn_mnar_30_mc.mvn_pcci_simulation}()},
\code{\link{mvn_mnar_30_mc.mvn_pcci_task}()},
\code{\link{mvn_mnar_30_mc.mvn_simulation}()},
\code{\link{mvn_mnar_30_mc.mvn_task}()},
\code{\link{mvn_mnar_30_mc.mvn}()},
\code{\link{mvn_mnar_50_mc.mvn_pcci_simulation}()},
\code{\link{mvn_mnar_50_mc.mvn_pcci_task}()},
\code{\link{mvn_mnar_50_mc.mvn_simulation}()},
\code{\link{mvn_mnar_50_mc.mvn_task}()},
\code{\link{mvn_mnar_50_mc.mvn}()},
\code{\link{mvn_mnar_mc.mvn_pcci_simulation}()},
\code{\link{mvn_mnar_mc.mvn_pcci_task}()},
\code{\link{mvn_mnar_mc.mvn_simulation}()},
\code{\link{mvn_mnar_mc.mvn_task}()},
\code{\link{mvn_mnar_mc.mvn}()},
\code{\link{mvn_mnar_std_mc.mvn.sem_pcci_simulation}()},
\code{\link{mvn_mnar_std_mc.mvn.sem_pcci_task}()},
\code{\link{mvn_mnar_std_mc.mvn.sem_simulation}()},
\code{\link{mvn_mnar_std_mc.mvn.sem_task}()},
\code{\link{mvn_mnar_std_mc.mvn.sem}()},
\code{\link{mvn_ols_mc.mvn_pcci_simulation}()},
\code{\link{mvn_ols_mc.mvn_pcci_task}()},
\code{\link{mvn_ols_mc.mvn_simulation}()},
\code{\link{mvn_ols_mc.mvn_task}()},
\code{\link{mvn_ols_mc.mvn}()},
\code{\link{mvn_std_mc.mvn.delta_pcci_simulation}()},
\code{\link{mvn_std_mc.mvn.delta_pcci_task}()},
\code{\link{mvn_std_mc.mvn.delta_simulation}()},
\code{\link{mvn_std_mc.mvn.delta_task}()},
\code{\link{mvn_std_mc.mvn.delta}()},
\code{\link{mvn_std_mc.mvn.sem_pcci_simulation}()},
\code{\link{mvn_std_mc.mvn.sem_pcci_task}()},
\code{\link{mvn_std_mc.mvn.sem_simulation}()},
\code{\link{mvn_std_mc.mvn.sem_task}()},
\code{\link{mvn_std_mc.mvn.sem}()},
\code{\link{mvn_std_mc.mvn.tb_pcci_simulation}()},
\code{\link{mvn_std_mc.mvn.tb_pcci_task}()},
\code{\link{mvn_std_mc.mvn.tb_simulation}()},
\code{\link{mvn_std_mc.mvn.tb_task}()},
\code{\link{mvn_std_mc.mvn.tb}()},
\code{\link{mvn_std_mc.t.delta_pcci_simulation}()},
\code{\link{mvn_std_mc.t.delta_pcci_task}()},
\code{\link{mvn_std_mc.t.delta_simulation}()},
\code{\link{mvn_std_mc.t.delta_task}()},
\code{\link{mvn_std_mc.t.delta}()},
\code{\link{mvn_std_mc.t.sem_pcci_simulation}()},
\code{\link{mvn_std_mc.t.sem_pcci_task}()},
\code{\link{mvn_std_mc.t.sem_simulation}()},
\code{\link{mvn_std_mc.t.sem_task}()},
\code{\link{mvn_std_mc.t.sem}()},
\code{\link{mvn_std_mc.t.tb_pcci_simulation}()},
\code{\link{mvn_std_mc.t.tb_pcci_task}()},
\code{\link{mvn_std_mc.t.tb_simulation}()},
\code{\link{mvn_std_mc.t.tb_task}()},
\code{\link{mvn_std_mc.t.tb}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{monte carlo method functions}
\keyword{mc}
