% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc.R
\name{mc.wishart}
\alias{mc.wishart}
\title{Monte Carlo Method for Indirect Effect in a Standardized Simple Mediation Model
Using the Wishart Distribution (Sampling Distribution)}
\usage{
mc.wishart(R = 20000L, Sigmahat, n, std = TRUE)
}
\arguments{
\item{R}{Integer.
Monte Carlo replications.}

\item{Sigmahat}{Numeric matrix.
Estimated covariance matrix.}

\item{n}{Integer.
Sample size.}

\item{std}{Logical.
Standardize the indirect effect
\eqn{
    \hat{\alpha}^{\prime} \hat{\beta}^{\prime}
    = \hat{\alpha} \hat{\beta} \frac{\hat{\sigma}_x}{\hat{\sigma}_y}}.}
}
\description{
Monte Carlo Method for Indirect Effect in a Standardized Simple Mediation Model
Using the Wishart Distribution (Sampling Distribution)
}
\examples{
Sigmahat <- cov(jeksterslabRdatarepo::thirst)
n <- dim(jeksterslabRdatarepo::thirst)[1]
thetahatstar <- mc.wishart(
  R = 20000L, Sigmahat = Sigmahat, n = n
)
hist(thetahatstar)
}
\seealso{
Other monte carlo method functions: 
\code{\link{beta_ols_mc.mvn_pcci_simulation}()},
\code{\link{beta_ols_mc.mvn_pcci_task}()},
\code{\link{beta_ols_mc.mvn_simulation}()},
\code{\link{beta_ols_mc.mvn_task}()},
\code{\link{beta_ols_mc.mvn}()},
\code{\link{exp_ols_mc.mvn_pcci_simulation}()},
\code{\link{exp_ols_mc.mvn_pcci_task}()},
\code{\link{exp_ols_mc.mvn_simulation}()},
\code{\link{exp_ols_mc.mvn_task}()},
\code{\link{exp_ols_mc.mvn}()},
\code{\link{mc.mvn}()},
\code{\link{mc.t}()},
\code{\link{mvn_mar_10_mc.mvn_pcci_simulation}()},
\code{\link{mvn_mar_10_mc.mvn_pcci_task}()},
\code{\link{mvn_mar_10_mc.mvn_simulation}()},
\code{\link{mvn_mar_10_mc.mvn_task}()},
\code{\link{mvn_mar_10_mc.mvn}()},
\code{\link{mvn_mar_20_mc.mvn_pcci_simulation}()},
\code{\link{mvn_mar_20_mc.mvn_pcci_task}()},
\code{\link{mvn_mar_20_mc.mvn_simulation}()},
\code{\link{mvn_mar_20_mc.mvn_task}()},
\code{\link{mvn_mar_20_mc.mvn}()},
\code{\link{mvn_mar_30_mc.mvn_pcci_simulation}()},
\code{\link{mvn_mar_30_mc.mvn_pcci_task}()},
\code{\link{mvn_mar_30_mc.mvn_simulation}()},
\code{\link{mvn_mar_30_mc.mvn_task}()},
\code{\link{mvn_mar_30_mc.mvn}()},
\code{\link{mvn_mcar_10_mc.mvn_pcci_simulation}()},
\code{\link{mvn_mcar_10_mc.mvn_pcci_task}()},
\code{\link{mvn_mcar_10_mc.mvn_simulation}()},
\code{\link{mvn_mcar_10_mc.mvn_task}()},
\code{\link{mvn_mcar_10_mc.mvn}()},
\code{\link{mvn_mcar_20_mc.mvn_pcci_simulation}()},
\code{\link{mvn_mcar_20_mc.mvn_pcci_task}()},
\code{\link{mvn_mcar_20_mc.mvn_simulation}()},
\code{\link{mvn_mcar_20_mc.mvn_task}()},
\code{\link{mvn_mcar_20_mc.mvn}()},
\code{\link{mvn_mcar_30_mc.mvn_pcci_simulation}()},
\code{\link{mvn_mcar_30_mc.mvn_pcci_task}()},
\code{\link{mvn_mcar_30_mc.mvn_simulation}()},
\code{\link{mvn_mcar_30_mc.mvn_task}()},
\code{\link{mvn_mcar_30_mc.mvn}()},
\code{\link{mvn_mnar_10_mc.mvn_pcci_simulation}()},
\code{\link{mvn_mnar_10_mc.mvn_pcci_task}()},
\code{\link{mvn_mnar_10_mc.mvn_simulation}()},
\code{\link{mvn_mnar_10_mc.mvn_task}()},
\code{\link{mvn_mnar_10_mc.mvn}()},
\code{\link{mvn_mnar_20_mc.mvn_pcci_simulation}()},
\code{\link{mvn_mnar_20_mc.mvn_pcci_task}()},
\code{\link{mvn_mnar_20_mc.mvn_simulation}()},
\code{\link{mvn_mnar_20_mc.mvn_task}()},
\code{\link{mvn_mnar_20_mc.mvn}()},
\code{\link{mvn_mnar_30_mc.mvn_pcci_simulation}()},
\code{\link{mvn_mnar_30_mc.mvn_pcci_task}()},
\code{\link{mvn_mnar_30_mc.mvn_simulation}()},
\code{\link{mvn_mnar_30_mc.mvn_task}()},
\code{\link{mvn_mnar_30_mc.mvn}()},
\code{\link{mvn_ols_mc.mvn_pcci_simulation}()},
\code{\link{mvn_ols_mc.mvn_pcci_task}()},
\code{\link{mvn_ols_mc.mvn_simulation}()},
\code{\link{mvn_ols_mc.mvn_task}()},
\code{\link{mvn_ols_mc.mvn}()},
\code{\link{mvn_sem_mc.mvn_pcci_simulation}()},
\code{\link{mvn_sem_mc.mvn_pcci_task}()},
\code{\link{mvn_sem_mc.mvn_simulation}()},
\code{\link{mvn_sem_mc.mvn_task}()},
\code{\link{mvn_sem_mc.mvn}()},
\code{\link{mvn_std_mc.mvn.delta_pcci_simulation}()},
\code{\link{mvn_std_mc.mvn.delta_pcci_task}()},
\code{\link{mvn_std_mc.mvn.delta_simulation}()},
\code{\link{mvn_std_mc.mvn.delta_task}()},
\code{\link{mvn_std_mc.mvn.delta}()},
\code{\link{mvn_std_mc.mvn.sem_pcci_simulation}()},
\code{\link{mvn_std_mc.mvn.sem_pcci_task}()},
\code{\link{mvn_std_mc.mvn.sem_simulation}()},
\code{\link{mvn_std_mc.mvn.sem_task}()},
\code{\link{mvn_std_mc.mvn.sem}()},
\code{\link{mvn_std_mc.mvn.tb_pcci_simulation}()},
\code{\link{mvn_std_mc.mvn.tb_pcci_task}()},
\code{\link{mvn_std_mc.mvn.tb_simulation}()},
\code{\link{mvn_std_mc.mvn.tb_task}()},
\code{\link{mvn_std_mc.mvn.tb}()},
\code{\link{mvn_std_mc.wishart_pcci_simulation}()},
\code{\link{mvn_std_mc.wishart_pcci_task}()},
\code{\link{mvn_std_mc.wishart_simulation}()},
\code{\link{mvn_std_mc.wishart_task}()},
\code{\link{mvn_std_mc.wishart}()},
\code{\link{vm_mod_ols_mc.mvn_pcci_simulation}()},
\code{\link{vm_mod_ols_mc.mvn_pcci_task}()},
\code{\link{vm_mod_ols_mc.mvn_simulation}()},
\code{\link{vm_mod_ols_mc.mvn_task}()},
\code{\link{vm_mod_ols_mc.mvn}()},
\code{\link{vm_mod_sem_mc.mvn_pcci_simulation}()},
\code{\link{vm_mod_sem_mc.mvn_pcci_task}()},
\code{\link{vm_mod_sem_mc.mvn_simulation}()},
\code{\link{vm_mod_sem_mc.mvn_task}()},
\code{\link{vm_mod_sem_mc.mvn}()},
\code{\link{vm_sev_ols_mc.mvn_pcci_simulation}()},
\code{\link{vm_sev_ols_mc.mvn_pcci_task}()},
\code{\link{vm_sev_ols_mc.mvn_simulation}()},
\code{\link{vm_sev_ols_mc.mvn_task}()},
\code{\link{vm_sev_ols_mc.mvn}()},
\code{\link{vm_sev_sem_mc.mvn_pcci_simulation}()},
\code{\link{vm_sev_sem_mc.mvn_pcci_task}()},
\code{\link{vm_sev_sem_mc.mvn_simulation}()},
\code{\link{vm_sev_sem_mc.mvn_task}()},
\code{\link{vm_sev_sem_mc.mvn}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{monte carlo method functions}
\keyword{mc}
