% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ram.R
\name{mutheta}
\alias{mutheta}
\title{Model-Implied Mean Vector}
\usage{
mutheta(mux, deltam, deltay, taudot, beta, alpha)
}
\arguments{
\item{mux}{Numeric.
Mean of \code{x} \eqn{\left( \mu_x \right)} .}

\item{deltam}{Numeric.
Intercept of \code{m} \eqn{\left( \delta_m \right)} .}

\item{deltay}{Numeric.
Intercept of \code{y} \eqn{\left( \delta_y \right)} .}

\item{taudot}{Numeric.
Slope of path from \code{x} to \code{y} \eqn{\left( \dot{\tau} \right)}.}

\item{beta}{Numeric.
Slope of path from \code{m} to \code{y} \eqn{\left( \beta \right)} .}

\item{alpha}{Numeric.
Slope of path from \code{x} to \code{m} \eqn{\left( \alpha \right)} .}
}
\description{
Model-Implied Mean Vector from the simple mediation model.
}
\details{
The simple mediation model is given by
\deqn{
    y_i = \delta_y + \dot{\tau} x_i + \beta m_i + \varepsilon_{y_{i}}
  }

\deqn{
    m_i = \delta_m + \alpha x_i + \varepsilon_{m_{i}}
  }

The parameters for the mean structure are
\deqn{
    \boldsymbol{\theta}_{\text{mean structure}} = \left\{ \mu_x, \delta_m, \delta_y \right\} .
  }

The parameters for the covariance strcuture are
\deqn{
    \boldsymbol{\theta}_{\text{covariance structure}} = \left\{ \dot{\tau}, \beta, \alpha, \sigma_{x}^{2},
    \sigma_{\varepsilon_{m}}^{2}, \sigma_{\varepsilon_{y}}^{2} \right\} .
  }
}
\examples{
mutheta(
  mux = 70.18000,
  deltam = 26.82246,
  deltay = 29.91071,
  taudot = 0.207648,
  beta = 0.451039,
  alpha = 0.338593
)
colMeans(jeksterslabRdatarepo::thirst)
}
\seealso{
Other reticular action model functions: 
\code{\link{A.std}()},
\code{\link{A}()},
\code{\link{Mfrommu}()},
\code{\link{M}()},
\code{\link{S.std}()},
\code{\link{Sfromsigma2}()},
\code{\link{Sigmatheta.std}()},
\code{\link{Sigmathetafromsigma2}()},
\code{\link{Sigmatheta}()},
\code{\link{S}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{reticular action model functions}
\keyword{ram}
