% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvn_dat_complete.R
\name{mvn_dat}
\alias{mvn_dat}
\title{Generate Data from a Multivariate Normal Distribution
following a Simple Mediation Model}
\usage{
mvn_dat(taskid)
}
\arguments{
\item{taskid}{Numeric.
Task ID.}
}
\description{
Generate Data from a Multivariate Normal Distribution
following a Simple Mediation Model
}
\examples{
data <- mvn_dat(taskid = 1)
fit.ols(data, minimal = TRUE)
fit.ols(data, minimal = TRUE, std = TRUE)
fit.ols(data, minimal = FALSE)
}
\seealso{
Other multivariate normal data functions: 
\code{\link{mvn_dat_mar_10_simulation}()},
\code{\link{mvn_dat_mar_10_task}()},
\code{\link{mvn_dat_mar_10}()},
\code{\link{mvn_dat_mar_20_simulation}()},
\code{\link{mvn_dat_mar_20_task}()},
\code{\link{mvn_dat_mar_20}()},
\code{\link{mvn_dat_mar_30_simulation}()},
\code{\link{mvn_dat_mar_30_task}()},
\code{\link{mvn_dat_mar_30}()},
\code{\link{mvn_dat_mar_50_simulation}()},
\code{\link{mvn_dat_mar_50_task}()},
\code{\link{mvn_dat_mar_50}()},
\code{\link{mvn_dat_mar_simulation}()},
\code{\link{mvn_dat_mar_task}()},
\code{\link{mvn_dat_mar}()},
\code{\link{mvn_dat_mcar_10_simulation}()},
\code{\link{mvn_dat_mcar_10_task}()},
\code{\link{mvn_dat_mcar_10}()},
\code{\link{mvn_dat_mcar_20_simulation}()},
\code{\link{mvn_dat_mcar_20_task}()},
\code{\link{mvn_dat_mcar_20}()},
\code{\link{mvn_dat_mcar_30_simulation}()},
\code{\link{mvn_dat_mcar_30_task}()},
\code{\link{mvn_dat_mcar_30}()},
\code{\link{mvn_dat_mcar_50_simulation}()},
\code{\link{mvn_dat_mcar_50_task}()},
\code{\link{mvn_dat_mcar_50}()},
\code{\link{mvn_dat_mcar_simulation}()},
\code{\link{mvn_dat_mcar_task}()},
\code{\link{mvn_dat_mcar}()},
\code{\link{mvn_dat_mnar_10_simulation}()},
\code{\link{mvn_dat_mnar_10_task}()},
\code{\link{mvn_dat_mnar_10}()},
\code{\link{mvn_dat_mnar_20_simulation}()},
\code{\link{mvn_dat_mnar_20_task}()},
\code{\link{mvn_dat_mnar_20}()},
\code{\link{mvn_dat_mnar_30_simulation}()},
\code{\link{mvn_dat_mnar_30_task}()},
\code{\link{mvn_dat_mnar_30}()},
\code{\link{mvn_dat_mnar_50_simulation}()},
\code{\link{mvn_dat_mnar_50_task}()},
\code{\link{mvn_dat_mnar_50}()},
\code{\link{mvn_dat_mnar_simulation}()},
\code{\link{mvn_dat_mnar_task}()},
\code{\link{mvn_dat_mnar}()},
\code{\link{mvn_dat_simulation}()},
\code{\link{mvn_dat_task}()},
\code{\link{mvn}()},
\code{\link{setparamsmvn}()},
\code{\link{useparamsmvn}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{multivariate normal data functions}
\keyword{mvn}
