% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvn_mar.R
\name{mvn_dat_mar}
\alias{mvn_dat_mar}
\title{Generate Data from a Multivariate Normal Distribution
following a Simple Mediation Model
with Data Missing at Random}
\usage{
mvn_dat_mar(data, taskid)
}
\arguments{
\item{taskid}{Numeric.
Task ID.}
}
\description{
Generate Data from a Multivariate Normal Distribution
following a Simple Mediation Model
with Data Missing at Random
}
\seealso{
Other multivariate normal data functions: 
\code{\link{mvn_dat_mar_simulation}()},
\code{\link{mvn_dat_mar_task}()},
\code{\link{mvn_dat_mcar_simulation}()},
\code{\link{mvn_dat_mcar_task}()},
\code{\link{mvn_dat_mcar}()},
\code{\link{mvn_dat_simulation}()},
\code{\link{mvn_dat_task}()},
\code{\link{mvn_dat}()},
\code{\link{mvn}()},
\code{\link{setparamsmvn}()},
\code{\link{useparamsmvn}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{multivariate normal data functions}
\keyword{mvn}
