% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvn_dat_mar.R
\name{mvn_dat_mar_task}
\alias{mvn_dat_mar_task}
\title{Generate Data from a Multivariate Normal Distribution
following a Simple Mediation Model
with Data Missing at Random
(Single Task)}
\usage{
mvn_dat_mar_task(taskid, dir = getwd(), overwrite = FALSE)
}
\arguments{
\item{taskid}{Numeric.
Task ID.}

\item{dir}{Character string.
Directory where results of \verb{mvn_dat_*} are stored.}

\item{overwrite}{Logical.
Overwrite existing results.}
}
\description{
Generate Data from a Multivariate Normal Distribution
following a Simple Mediation Model
with Data Missing at Random
(Single Task)
}
\seealso{
Other multivariate normal data functions: 
\code{\link{mvn_dat_mar_simulation}()},
\code{\link{mvn_dat_mar}()},
\code{\link{mvn_dat_mcar_simulation}()},
\code{\link{mvn_dat_mcar_task}()},
\code{\link{mvn_dat_mcar}()},
\code{\link{mvn_dat_mnar_simulation}()},
\code{\link{mvn_dat_mnar_task}()},
\code{\link{mvn_dat_mnar}()},
\code{\link{mvn_dat_simulation}()},
\code{\link{mvn_dat_task}()},
\code{\link{mvn_dat}()},
\code{\link{mvn}()},
\code{\link{setparamsmvn}()},
\code{\link{useparamsmvn}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{multivariate normal data functions}
\keyword{mvn}
