% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvn_complete_dat.R
\name{mvn_dat_task}
\alias{mvn_dat_task}
\title{Generate Data from a Multivariate Normal Distribution
following a Simple Mediation Model
(Single Task)}
\usage{
mvn_dat_task(taskid, dir = getwd(), overwrite = FALSE)
}
\arguments{
\item{taskid}{Numeric.
Task ID.}

\item{dir}{Character string.
Directory where results of \verb{mvn_dat_*} are stored.}

\item{overwrite}{Logical.
Overwrite existing results.}
}
\description{
Generate Data from a Multivariate Normal Distribution
following a Simple Mediation Model
(Single Task)
}
\seealso{
Other multivariate normal data functions: 
\code{\link{mvn_dat_simulation}()},
\code{\link{mvn_dat}()},
\code{\link{mvn}()},
\code{\link{setparamsmvn}()},
\code{\link{useparamsmvn}()},
\code{\link{vm_mod_dat_simulation_simulation_summary}()},
\code{\link{vm_mod_dat_simulation_task_summary}()},
\code{\link{vm_mod_dat_simulation}()},
\code{\link{vm_mod_dat_task}()},
\code{\link{vm_mod_dat}()},
\code{\link{vm_sev_dat_simulation_simulation_summary}()},
\code{\link{vm_sev_dat_simulation_task_summary}()},
\code{\link{vm_sev_dat_simulation}()},
\code{\link{vm_sev_dat_task}()},
\code{\link{vm_sev_dat}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{multivariate normal data functions}
\keyword{mvn_dat}
