% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{mvn_fit}
\alias{mvn_fit}
\title{Fit Simple Mediation Model
for Data Generated from a Multivariate Normal Distribution}
\usage{
mvn_fit(data, taskid)
}
\arguments{
\item{data}{\code{n} by 3 matrix or data frame.
\code{data[, 1]} correspond to values for \code{x}.
\code{data[, 2]} correspond to values for \code{m}.
\code{data[, 3]} correspond to values for \code{y}.}

\item{taskid}{Numeric.
Task ID.}
}
\description{
Fit Simple Mediation Model
for Data Generated from a Multivariate Normal Distribution
}
\examples{
taskid <- 1
data <- mvn_dat(taskid = taskid)
mvn_fit(data = data, taskid = taskid)
}
\seealso{
Other model fit functions: 
\code{\link{.fitsem}()},
\code{\link{.fit}()},
\code{\link{fit}()},
\code{\link{mvn_fit_simulation_summary}()},
\code{\link{mvn_fit_simulation}()},
\code{\link{mvn_fit_task_summary}()},
\code{\link{mvn_fit_task}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{model fit functions}
\keyword{fit}
