% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{mvn_fit_task}
\alias{mvn_fit_task}
\title{Fit Simple Mediation Model
for Data Generated from a Multivariate Normal Distribution
(Single Task)}
\usage{
mvn_fit_task(taskid, dir = getwd(), overwrite = FALSE)
}
\arguments{
\item{taskid}{Numeric.
Task ID.}

\item{dir}{Character string.
Directory where results of \verb{mvn_dat_*} are stored.}

\item{overwrite}{Logical.
Overwrite existing results.}
}
\description{
Fit Simple Mediation Model
for Data Generated from a Multivariate Normal Distribution
(Single Task)
}
\seealso{
Other model fit functions: 
\code{\link{.fitfiml}()},
\code{\link{.fitsem}()},
\code{\link{.fit}()},
\code{\link{fit}()},
\code{\link{mvn_fit_simulation_summary}()},
\code{\link{mvn_fit_simulation}()},
\code{\link{mvn_fit_task_summary}()},
\code{\link{mvn_fitfiml_mar_simulation}()},
\code{\link{mvn_fitfiml_mar_task}()},
\code{\link{mvn_fitfiml_mar}()},
\code{\link{mvn_fitfiml_mcar_simulation}()},
\code{\link{mvn_fitfiml_mcar_task}()},
\code{\link{mvn_fitfiml_mcar}()},
\code{\link{mvn_fit}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{model fit functions}
\keyword{fit}
