% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{mvn_fit_task_summary}
\alias{mvn_fit_task_summary}
\title{Fit Simple Mediation Model
for Data Generated from a Multivariate Normal Distribution
(Single Task Summary)}
\usage{
mvn_fit_task_summary(taskid, dir = getwd())
}
\arguments{
\item{taskid}{Numeric.
Task ID.}

\item{dir}{Character string.
Directory where results of \verb{mvn_dat_*} are stored.}
}
\description{
Fit Simple Mediation Model
for Data Generated from a Multivariate Normal Distribution
(Single Task Summary)
}
\seealso{
Other model fit functions: 
\code{\link{.fitsem}()},
\code{\link{.fit}()},
\code{\link{fit}()},
\code{\link{mvn_fit_simulation_summary}()},
\code{\link{mvn_fit_simulation}()},
\code{\link{mvn_fit_task}()},
\code{\link{mvn_fit}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{model fit functions}
\keyword{fit}
