% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvn_fitols.R
\name{mvn_fitols_task_summary}
\alias{mvn_fitols_task_summary}
\title{Fit Simple Mediation Model
for Data Generated from a Multivariate Normal Distribution
(Single Task Summary)}
\usage{
mvn_fitols_task_summary(taskid, dir = getwd())
}
\arguments{
\item{taskid}{Numeric.
Task ID.}

\item{dir}{Character string.
Directory where results of \verb{mvn_dat_*} are stored.}
}
\description{
Fit Simple Mediation Model
for Data Generated from a Multivariate Normal Distribution
(Single Task Summary)
}
\seealso{
Other model fit functions: 
\code{\link{.fitfiml}()},
\code{\link{.fitsem}()},
\code{\link{.fit}()},
\code{\link{fit}()},
\code{\link{mvn_fitfiml_mar_simulation}()},
\code{\link{mvn_fitfiml_mar_task}()},
\code{\link{mvn_fitfiml_mar}()},
\code{\link{mvn_fitfiml_mcar_simulation}()},
\code{\link{mvn_fitfiml_mcar_task}()},
\code{\link{mvn_fitfiml_mcar}()},
\code{\link{mvn_fitols_simulation_summary}()},
\code{\link{mvn_fitols_simulation}()},
\code{\link{mvn_fitols_task}()},
\code{\link{mvn_fitols}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{model fit functions}
\keyword{fit}
