% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jack.R
\name{mvn_jack}
\alias{mvn_jack}
\title{Jackknife Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution}
\usage{
mvn_jack(data, taskid)
}
\arguments{
\item{data}{\code{n} by 3 matrix or data frame.
\code{data[, 1]} correspond to values for \code{x}.
\code{data[, 2]} correspond to values for \code{m}.
\code{data[, 3]} correspond to values for \code{y}.}

\item{taskid}{Numeric.
Task ID.}
}
\description{
Jackknife Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution
}
\examples{
taskid <- 1
data <- mvn_dat(taskid = taskid)
mvn_jack(data = data, taskid = taskid)
}
\seealso{
Other jackknife functions: 
\code{\link{.jack}()},
\code{\link{mvn_jack_mar_simulation}()},
\code{\link{mvn_jack_mar_task}()},
\code{\link{mvn_jack_mar}()},
\code{\link{mvn_jack_mcar_simulation}()},
\code{\link{mvn_jack_mcar_task}()},
\code{\link{mvn_jack_mcar}()},
\code{\link{mvn_jack_simulation}()},
\code{\link{mvn_jack_task}()},
\code{\link{mvn_std_jack_simulation}()},
\code{\link{mvn_std_jack_task}()},
\code{\link{mvn_std_jack}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{jackknife functions}
\keyword{jack}
