% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jack_mar.R
\name{mvn_jack_mar_task}
\alias{mvn_jack_mar_task}
\title{Jackknife Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution (Single Task)}
\usage{
mvn_jack_mar_task(taskid, dir = getwd(), overwrite = FALSE)
}
\arguments{
\item{taskid}{Numeric.
Task ID.}

\item{dir}{Character string.
Directory where results of \verb{mvn_dat_*} are stored.}

\item{overwrite}{Logical.
Overwrite existing results.}
}
\description{
Jackknife Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution (Single Task)
}
\seealso{
Other jackknife functions: 
\code{\link{.jack}()},
\code{\link{mvn_jack_mar_simulation}()},
\code{\link{mvn_jack_mar}()},
\code{\link{mvn_jack_mcar_simulation}()},
\code{\link{mvn_jack_mcar_task}()},
\code{\link{mvn_jack_mcar}()},
\code{\link{mvn_jack_simulation}()},
\code{\link{mvn_jack_task}()},
\code{\link{mvn_jack}()},
\code{\link{mvn_std_jack_simulation}()},
\code{\link{mvn_std_jack_task}()},
\code{\link{mvn_std_jack}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{jackknife functions}
\keyword{jack}
