% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvn_mar_10_nbbc.fiml.R,
%   R/mvn_mar_10_nbpc.fiml.R
\name{mvn_mar_10_nb.fiml_pcci_simulation}
\alias{mvn_mar_10_nb.fiml_pcci_simulation}
\title{Nonparametric Bootstrap with Bias-Corrected Confidence Intervals for Indirect Effect in a Simple Mediation Model
for Data Generated from a Multivariate Normal Distribution with Data Missing at Random
(Simulation)}
\usage{
mvn_mar_10_nb.fiml_pcci_simulation(dir = getwd(), all = TRUE, taskid = NULL)

mvn_mar_10_nb.fiml_pcci_simulation(dir = getwd(), all = TRUE, taskid = NULL)
}
\arguments{
\item{dir}{Character string.
Directory where results of \verb{mvn_dat_*} are stored.}

\item{all}{Logical.
Process all results.}

\item{taskid}{Numeric.
Task ID.}
}
\description{
Nonparametric Bootstrap with Bias-Corrected Confidence Intervals for Indirect Effect in a Simple Mediation Model
for Data Generated from a Multivariate Normal Distribution with Data Missing at Random
(Simulation)

Nonparametric Bootstrap with Percentile Confidence Intervals for Indirect Effect in a Simple Mediation Model
for Data Generated from a Multivariate Normal Distribution with Data Missing at Random
(Simulation)
}
\seealso{
Other nonparametric functions: 
\code{\link{mvn_mar_10_nbbc.fiml_simulation}()},
\code{\link{mvn_mar_10_nbbc.fiml_task}()},
\code{\link{mvn_mar_10_nbpc.fiml_simulation}()},
\code{\link{mvn_mar_10_nbpc.fiml_task}()},
\code{\link{mvn_mar_20_nb.fiml_pcci_simulation}()},
\code{\link{mvn_mar_20_nbbc.fiml_simulation}()},
\code{\link{mvn_mar_20_nbbc.fiml_task}()},
\code{\link{mvn_mar_20_nbpc.fiml_simulation}()},
\code{\link{mvn_mar_20_nbpc.fiml_task}()},
\code{\link{mvn_mar_30_nb.fiml_pcci_simulation}()},
\code{\link{mvn_mar_30_nbbc.fiml_simulation}()},
\code{\link{mvn_mar_30_nbbc.fiml_task}()},
\code{\link{mvn_mar_30_nbpc.fiml_simulation}()},
\code{\link{mvn_mar_30_nbpc.fiml_task}()},
\code{\link{mvn_mar_nb_bcaci_simulation}()},
\code{\link{mvn_mar_nb_bcaci_task}()},
\code{\link{mvn_mar_nb_bcci_simulation}()},
\code{\link{mvn_mar_nb_bcci_task}()},
\code{\link{mvn_mar_nb_pcci_simulation}()},
\code{\link{mvn_mar_nb_pcci_task}()},
\code{\link{mvn_mar_nb_simulation}()},
\code{\link{mvn_mar_nb_task}()},
\code{\link{mvn_mar_nb}()},
\code{\link{mvn_mcar_10_nb.fiml_pcci_simulation}()},
\code{\link{mvn_mcar_10_nbbc.fiml_simulation}()},
\code{\link{mvn_mcar_10_nbbc.fiml_task}()},
\code{\link{mvn_mcar_10_nbpc.fiml_simulation}()},
\code{\link{mvn_mcar_10_nbpc.fiml_task}()},
\code{\link{mvn_mcar_20_nb.fiml_pcci_simulation}()},
\code{\link{mvn_mcar_20_nbbc.fiml_simulation}()},
\code{\link{mvn_mcar_20_nbbc.fiml_task}()},
\code{\link{mvn_mcar_20_nbpc.fiml_simulation}()},
\code{\link{mvn_mcar_20_nbpc.fiml_task}()},
\code{\link{mvn_mcar_30_nb.fiml_pcci_simulation}()},
\code{\link{mvn_mcar_30_nbbc.fiml_simulation}()},
\code{\link{mvn_mcar_30_nbbc.fiml_task}()},
\code{\link{mvn_mcar_30_nbpc.fiml_simulation}()},
\code{\link{mvn_mcar_30_nbpc.fiml_task}()},
\code{\link{mvn_mcar_nb_bcaci_simulation}()},
\code{\link{mvn_mcar_nb_bcaci_task}()},
\code{\link{mvn_mcar_nb_bcci_simulation}()},
\code{\link{mvn_mcar_nb_bcci_task}()},
\code{\link{mvn_mcar_nb_fiml_bcaci_simulation}()},
\code{\link{mvn_mcar_nb_fiml_bcaci_task}()},
\code{\link{mvn_mcar_nb_fiml_bcci_simulation}()},
\code{\link{mvn_mcar_nb_fiml_bcci_task}()},
\code{\link{mvn_mcar_nb_fiml_pcci_simulation}()},
\code{\link{mvn_mcar_nb_fiml_pcci_task}()},
\code{\link{mvn_mcar_nb_fiml_simulation}()},
\code{\link{mvn_mcar_nb_fiml_task}()},
\code{\link{mvn_mcar_nb_fiml}()},
\code{\link{mvn_mcar_nb_pcci_simulation}()},
\code{\link{mvn_mcar_nb_pcci_task}()},
\code{\link{mvn_mcar_nb_simulation}()},
\code{\link{mvn_mcar_nb_task}()},
\code{\link{mvn_mcar_nb}()},
\code{\link{mvn_mnar_10_nb.fiml_pcci_simulation}()},
\code{\link{mvn_mnar_10_nbbc.fiml_simulation}()},
\code{\link{mvn_mnar_10_nbbc.fiml_task}()},
\code{\link{mvn_mnar_10_nbpc.fiml_simulation}()},
\code{\link{mvn_mnar_10_nbpc.fiml_task}()},
\code{\link{mvn_mnar_20_nb.fiml_pcci_simulation}()},
\code{\link{mvn_mnar_20_nbbc.fiml_simulation}()},
\code{\link{mvn_mnar_20_nbbc.fiml_task}()},
\code{\link{mvn_mnar_20_nbpc.fiml_simulation}()},
\code{\link{mvn_mnar_20_nbpc.fiml_task}()},
\code{\link{mvn_mnar_30_nb.fiml_pcci_simulation}()},
\code{\link{mvn_mnar_30_nbbc.fiml_simulation}()},
\code{\link{mvn_mnar_30_nbbc.fiml_task}()},
\code{\link{mvn_mnar_30_nbpc.fiml_simulation}()},
\code{\link{mvn_mnar_30_nbpc.fiml_task}()},
\code{\link{mvn_mnar_nb_bcaci_simulation}()},
\code{\link{mvn_mnar_nb_bcaci_task}()},
\code{\link{mvn_mnar_nb_bcci_simulation}()},
\code{\link{mvn_mnar_nb_bcci_task}()},
\code{\link{mvn_mnar_nb_pcci_simulation}()},
\code{\link{mvn_mnar_nb_pcci_task}()},
\code{\link{mvn_mnar_nb_simulation}()},
\code{\link{mvn_mnar_nb_task}()},
\code{\link{mvn_mnar_nb}()},
\code{\link{mvn_nb_bcaci_simulation}()},
\code{\link{mvn_nb_bcaci_task}()},
\code{\link{mvn_nb_bcci_simulation}()},
\code{\link{mvn_nb_bcci_task}()},
\code{\link{mvn_nb_pcci_simulation}()},
\code{\link{mvn_nb_pcci_task}()},
\code{\link{mvn_nb_simulation}()},
\code{\link{mvn_nb_task}()},
\code{\link{mvn_nb}()},
\code{\link{mvn_std_nb_bcaci_simulation}()},
\code{\link{mvn_std_nb_bcaci_task}()},
\code{\link{mvn_std_nb_bcci_simulation}()},
\code{\link{mvn_std_nb_bcci_task}()},
\code{\link{mvn_std_nb_pcci_simulation}()},
\code{\link{mvn_std_nb_pcci_task}()},
\code{\link{mvn_std_nb_simulation}()},
\code{\link{mvn_std_nb_task}()},
\code{\link{mvn_std_nb}()},
\code{\link{nbbc.fiml}()},
\code{\link{nbpc.fiml}()},
\code{\link{nb}()}

Other nonparametric functions: 
\code{\link{mvn_mar_10_nbbc.fiml_simulation}()},
\code{\link{mvn_mar_10_nbbc.fiml_task}()},
\code{\link{mvn_mar_10_nbpc.fiml_simulation}()},
\code{\link{mvn_mar_10_nbpc.fiml_task}()},
\code{\link{mvn_mar_20_nb.fiml_pcci_simulation}()},
\code{\link{mvn_mar_20_nbbc.fiml_simulation}()},
\code{\link{mvn_mar_20_nbbc.fiml_task}()},
\code{\link{mvn_mar_20_nbpc.fiml_simulation}()},
\code{\link{mvn_mar_20_nbpc.fiml_task}()},
\code{\link{mvn_mar_30_nb.fiml_pcci_simulation}()},
\code{\link{mvn_mar_30_nbbc.fiml_simulation}()},
\code{\link{mvn_mar_30_nbbc.fiml_task}()},
\code{\link{mvn_mar_30_nbpc.fiml_simulation}()},
\code{\link{mvn_mar_30_nbpc.fiml_task}()},
\code{\link{mvn_mar_nb_bcaci_simulation}()},
\code{\link{mvn_mar_nb_bcaci_task}()},
\code{\link{mvn_mar_nb_bcci_simulation}()},
\code{\link{mvn_mar_nb_bcci_task}()},
\code{\link{mvn_mar_nb_pcci_simulation}()},
\code{\link{mvn_mar_nb_pcci_task}()},
\code{\link{mvn_mar_nb_simulation}()},
\code{\link{mvn_mar_nb_task}()},
\code{\link{mvn_mar_nb}()},
\code{\link{mvn_mcar_10_nb.fiml_pcci_simulation}()},
\code{\link{mvn_mcar_10_nbbc.fiml_simulation}()},
\code{\link{mvn_mcar_10_nbbc.fiml_task}()},
\code{\link{mvn_mcar_10_nbpc.fiml_simulation}()},
\code{\link{mvn_mcar_10_nbpc.fiml_task}()},
\code{\link{mvn_mcar_20_nb.fiml_pcci_simulation}()},
\code{\link{mvn_mcar_20_nbbc.fiml_simulation}()},
\code{\link{mvn_mcar_20_nbbc.fiml_task}()},
\code{\link{mvn_mcar_20_nbpc.fiml_simulation}()},
\code{\link{mvn_mcar_20_nbpc.fiml_task}()},
\code{\link{mvn_mcar_30_nb.fiml_pcci_simulation}()},
\code{\link{mvn_mcar_30_nbbc.fiml_simulation}()},
\code{\link{mvn_mcar_30_nbbc.fiml_task}()},
\code{\link{mvn_mcar_30_nbpc.fiml_simulation}()},
\code{\link{mvn_mcar_30_nbpc.fiml_task}()},
\code{\link{mvn_mcar_nb_bcaci_simulation}()},
\code{\link{mvn_mcar_nb_bcaci_task}()},
\code{\link{mvn_mcar_nb_bcci_simulation}()},
\code{\link{mvn_mcar_nb_bcci_task}()},
\code{\link{mvn_mcar_nb_fiml_bcaci_simulation}()},
\code{\link{mvn_mcar_nb_fiml_bcaci_task}()},
\code{\link{mvn_mcar_nb_fiml_bcci_simulation}()},
\code{\link{mvn_mcar_nb_fiml_bcci_task}()},
\code{\link{mvn_mcar_nb_fiml_pcci_simulation}()},
\code{\link{mvn_mcar_nb_fiml_pcci_task}()},
\code{\link{mvn_mcar_nb_fiml_simulation}()},
\code{\link{mvn_mcar_nb_fiml_task}()},
\code{\link{mvn_mcar_nb_fiml}()},
\code{\link{mvn_mcar_nb_pcci_simulation}()},
\code{\link{mvn_mcar_nb_pcci_task}()},
\code{\link{mvn_mcar_nb_simulation}()},
\code{\link{mvn_mcar_nb_task}()},
\code{\link{mvn_mcar_nb}()},
\code{\link{mvn_mnar_10_nb.fiml_pcci_simulation}()},
\code{\link{mvn_mnar_10_nbbc.fiml_simulation}()},
\code{\link{mvn_mnar_10_nbbc.fiml_task}()},
\code{\link{mvn_mnar_10_nbpc.fiml_simulation}()},
\code{\link{mvn_mnar_10_nbpc.fiml_task}()},
\code{\link{mvn_mnar_20_nb.fiml_pcci_simulation}()},
\code{\link{mvn_mnar_20_nbbc.fiml_simulation}()},
\code{\link{mvn_mnar_20_nbbc.fiml_task}()},
\code{\link{mvn_mnar_20_nbpc.fiml_simulation}()},
\code{\link{mvn_mnar_20_nbpc.fiml_task}()},
\code{\link{mvn_mnar_30_nb.fiml_pcci_simulation}()},
\code{\link{mvn_mnar_30_nbbc.fiml_simulation}()},
\code{\link{mvn_mnar_30_nbbc.fiml_task}()},
\code{\link{mvn_mnar_30_nbpc.fiml_simulation}()},
\code{\link{mvn_mnar_30_nbpc.fiml_task}()},
\code{\link{mvn_mnar_nb_bcaci_simulation}()},
\code{\link{mvn_mnar_nb_bcaci_task}()},
\code{\link{mvn_mnar_nb_bcci_simulation}()},
\code{\link{mvn_mnar_nb_bcci_task}()},
\code{\link{mvn_mnar_nb_pcci_simulation}()},
\code{\link{mvn_mnar_nb_pcci_task}()},
\code{\link{mvn_mnar_nb_simulation}()},
\code{\link{mvn_mnar_nb_task}()},
\code{\link{mvn_mnar_nb}()},
\code{\link{mvn_nb_bcaci_simulation}()},
\code{\link{mvn_nb_bcaci_task}()},
\code{\link{mvn_nb_bcci_simulation}()},
\code{\link{mvn_nb_bcci_task}()},
\code{\link{mvn_nb_pcci_simulation}()},
\code{\link{mvn_nb_pcci_task}()},
\code{\link{mvn_nb_simulation}()},
\code{\link{mvn_nb_task}()},
\code{\link{mvn_nb}()},
\code{\link{mvn_std_nb_bcaci_simulation}()},
\code{\link{mvn_std_nb_bcaci_task}()},
\code{\link{mvn_std_nb_bcci_simulation}()},
\code{\link{mvn_std_nb_bcci_task}()},
\code{\link{mvn_std_nb_pcci_simulation}()},
\code{\link{mvn_std_nb_pcci_task}()},
\code{\link{mvn_std_nb_simulation}()},
\code{\link{mvn_std_nb_task}()},
\code{\link{mvn_std_nb}()},
\code{\link{nbbc.fiml}()},
\code{\link{nbpc.fiml}()},
\code{\link{nb}()}
}
\author{
Ivan Jacob Agaloos Pesigan

Ivan Jacob Agaloos Pesigan
}
\concept{nonparametric functions}
\keyword{mvn_mar_unstd}
