% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvn_mar_30_unstd_pb.mvn.R
\name{mvn_mar_30_pb.mvn_pcci_task}
\alias{mvn_mar_30_pb.mvn_pcci_task}
\title{Parametric Bootstrap Percentile Confidence Intervals for Indirect Effect in a Simple Mediation Model
for Data Generated from a Multivariate Normal Distribution
with Data Missing at Random - 30\%
(Single Task)}
\usage{
mvn_mar_30_pb.mvn_pcci_task(taskid, dir = getwd())
}
\arguments{
\item{taskid}{Numeric.
Task ID.}

\item{dir}{Character string.
Directory where results of \verb{mvn_dat_*} are stored.}
}
\description{
Parametric Bootstrap Percentile Confidence Intervals for Indirect Effect in a Simple Mediation Model
for Data Generated from a Multivariate Normal Distribution
with Data Missing at Random - 30\%
(Single Task)
}
\seealso{
Other parametric functions: 
\code{\link{mvn_mar_10_pb.mvn_bcci_simulation}()},
\code{\link{mvn_mar_10_pb.mvn_bcci_task}()},
\code{\link{mvn_mar_10_pb.mvn_pcci_simulation}()},
\code{\link{mvn_mar_10_pb.mvn_pcci_task}()},
\code{\link{mvn_mar_10_pb.mvn_simulation}()},
\code{\link{mvn_mar_10_pb.mvn_task}()},
\code{\link{mvn_mar_10_pb.mvn}()},
\code{\link{mvn_mar_20_pb.mvn_bcci_simulation}()},
\code{\link{mvn_mar_20_pb.mvn_bcci_task}()},
\code{\link{mvn_mar_20_pb.mvn_pcci_simulation}()},
\code{\link{mvn_mar_20_pb.mvn_pcci_task}()},
\code{\link{mvn_mar_20_pb.mvn_simulation}()},
\code{\link{mvn_mar_20_pb.mvn_task}()},
\code{\link{mvn_mar_20_pb.mvn}()},
\code{\link{mvn_mar_30_pb.mvn_bcci_simulation}()},
\code{\link{mvn_mar_30_pb.mvn_bcci_task}()},
\code{\link{mvn_mar_30_pb.mvn_pcci_simulation}()},
\code{\link{mvn_mar_30_pb.mvn_simulation}()},
\code{\link{mvn_mar_30_pb.mvn_task}()},
\code{\link{mvn_mar_30_pb.mvn}()},
\code{\link{mvn_mcar_10_pb.mvn_bcci_simulation}()},
\code{\link{mvn_mcar_10_pb.mvn_bcci_task}()},
\code{\link{mvn_mcar_10_pb.mvn_pcci_simulation}()},
\code{\link{mvn_mcar_10_pb.mvn_pcci_task}()},
\code{\link{mvn_mcar_10_pb.mvn_simulation}()},
\code{\link{mvn_mcar_10_pb.mvn_task}()},
\code{\link{mvn_mcar_10_pb.mvn}()},
\code{\link{mvn_mcar_20_pb.mvn_bcci_simulation}()},
\code{\link{mvn_mcar_20_pb.mvn_bcci_task}()},
\code{\link{mvn_mcar_20_pb.mvn_pcci_simulation}()},
\code{\link{mvn_mcar_20_pb.mvn_pcci_task}()},
\code{\link{mvn_mcar_20_pb.mvn_simulation}()},
\code{\link{mvn_mcar_20_pb.mvn_task}()},
\code{\link{mvn_mcar_20_pb.mvn}()},
\code{\link{mvn_mcar_30_pb.mvn_bcci_simulation}()},
\code{\link{mvn_mcar_30_pb.mvn_bcci_task}()},
\code{\link{mvn_mcar_30_pb.mvn_pcci_simulation}()},
\code{\link{mvn_mcar_30_pb.mvn_pcci_task}()},
\code{\link{mvn_mcar_30_pb.mvn_simulation}()},
\code{\link{mvn_mcar_30_pb.mvn_task}()},
\code{\link{mvn_mcar_30_pb.mvn}()},
\code{\link{mvn_mnar_10_pb.mvn_bcci_simulation}()},
\code{\link{mvn_mnar_10_pb.mvn_bcci_task}()},
\code{\link{mvn_mnar_10_pb.mvn_pcci_simulation}()},
\code{\link{mvn_mnar_10_pb.mvn_pcci_task}()},
\code{\link{mvn_mnar_10_pb.mvn_simulation}()},
\code{\link{mvn_mnar_10_pb.mvn_task}()},
\code{\link{mvn_mnar_10_pb.mvn}()},
\code{\link{mvn_mnar_20_pb.mvn_bcci_simulation}()},
\code{\link{mvn_mnar_20_pb.mvn_bcci_task}()},
\code{\link{mvn_mnar_20_pb.mvn_pcci_simulation}()},
\code{\link{mvn_mnar_20_pb.mvn_pcci_task}()},
\code{\link{mvn_mnar_20_pb.mvn_simulation}()},
\code{\link{mvn_mnar_20_pb.mvn_task}()},
\code{\link{mvn_mnar_20_pb.mvn}()},
\code{\link{mvn_mnar_30_pb.mvn_bcci_simulation}()},
\code{\link{mvn_mnar_30_pb.mvn_bcci_task}()},
\code{\link{mvn_mnar_30_pb.mvn_pcci_simulation}()},
\code{\link{mvn_mnar_30_pb.mvn_pcci_task}()},
\code{\link{mvn_mnar_30_pb.mvn_simulation}()},
\code{\link{mvn_mnar_30_pb.mvn_task}()},
\code{\link{mvn_mnar_30_pb.mvn}()},
\code{\link{mvn_pb.mvn_bcaci_simulation}()},
\code{\link{mvn_pb.mvn_bcaci_task}()},
\code{\link{mvn_pb.mvn_bcci_simulation}()},
\code{\link{mvn_pb.mvn_bcci_task}()},
\code{\link{mvn_pb.mvn_pcci_simulation}()},
\code{\link{mvn_pb.mvn_pcci_task}()},
\code{\link{mvn_pb.mvn_simulation}()},
\code{\link{mvn_pb.mvn_task}()},
\code{\link{mvn_pb.mvn}()},
\code{\link{mvn_std_pb.mvn_bcaci_simulation}()},
\code{\link{mvn_std_pb.mvn_bcaci_task}()},
\code{\link{mvn_std_pb.mvn_bcci_simulation}()},
\code{\link{mvn_std_pb.mvn_bcci_task}()},
\code{\link{mvn_std_pb.mvn_pcci_simulation}()},
\code{\link{mvn_std_pb.mvn_pcci_task}()},
\code{\link{mvn_std_pb.mvn_simulation}()},
\code{\link{mvn_std_pb.mvn_task}()},
\code{\link{mvn_std_pb.mvn}()},
\code{\link{pb.mvn}()},
\code{\link{pb.vm}()},
\code{\link{vm_mod_pb.mvn_bcaci_simulation}()},
\code{\link{vm_mod_pb.mvn_bcaci_task}()},
\code{\link{vm_mod_pb.mvn_bcci_simulation}()},
\code{\link{vm_mod_pb.mvn_bcci_task}()},
\code{\link{vm_mod_pb.mvn_pcci_simulation}()},
\code{\link{vm_mod_pb.mvn_pcci_task}()},
\code{\link{vm_mod_pb.mvn_simulation}()},
\code{\link{vm_mod_pb.mvn_task}()},
\code{\link{vm_mod_pb.mvn}()},
\code{\link{vm_mod_pb.vm_bcaci_simulation}()},
\code{\link{vm_mod_pb.vm_bcaci_task}()},
\code{\link{vm_mod_pb.vm_bcci_simulation}()},
\code{\link{vm_mod_pb.vm_bcci_task}()},
\code{\link{vm_mod_pb.vm_pcci_simulation}()},
\code{\link{vm_mod_pb.vm_pcci_task}()},
\code{\link{vm_mod_pb.vm_simulation}()},
\code{\link{vm_mod_pb.vm_task}()},
\code{\link{vm_mod_pb.vm}()},
\code{\link{vm_mod_std_pb.mvn_bcaci_simulation}()},
\code{\link{vm_mod_std_pb.mvn_bcaci_task}()},
\code{\link{vm_mod_std_pb.mvn_bcci_simulation}()},
\code{\link{vm_mod_std_pb.mvn_bcci_task}()},
\code{\link{vm_mod_std_pb.mvn_pcci_simulation}()},
\code{\link{vm_mod_std_pb.mvn_pcci_task}()},
\code{\link{vm_mod_std_pb.mvn_simulation}()},
\code{\link{vm_mod_std_pb.mvn_task}()},
\code{\link{vm_mod_std_pb.mvn}()},
\code{\link{vm_sev_pb.mvn_bcaci_simulation}()},
\code{\link{vm_sev_pb.mvn_bcaci_task}()},
\code{\link{vm_sev_pb.mvn_bcci_simulation}()},
\code{\link{vm_sev_pb.mvn_bcci_task}()},
\code{\link{vm_sev_pb.mvn_pcci_simulation}()},
\code{\link{vm_sev_pb.mvn_pcci_task}()},
\code{\link{vm_sev_pb.mvn_simulation}()},
\code{\link{vm_sev_pb.mvn_task}()},
\code{\link{vm_sev_pb.mvn}()},
\code{\link{vm_sev_pb.vm_bcaci_simulation}()},
\code{\link{vm_sev_pb.vm_bcaci_task}()},
\code{\link{vm_sev_pb.vm_bcci_simulation}()},
\code{\link{vm_sev_pb.vm_bcci_task}()},
\code{\link{vm_sev_pb.vm_pcci_simulation}()},
\code{\link{vm_sev_pb.vm_pcci_task}()},
\code{\link{vm_sev_pb.vm_simulation}()},
\code{\link{vm_sev_pb.vm_task}()},
\code{\link{vm_sev_pb.vm}()},
\code{\link{vm_sev_std_pb.mvn_bcaci_simulation}()},
\code{\link{vm_sev_std_pb.mvn_bcaci_task}()},
\code{\link{vm_sev_std_pb.mvn_bcci_simulation}()},
\code{\link{vm_sev_std_pb.mvn_bcci_task}()},
\code{\link{vm_sev_std_pb.mvn_pcci_simulation}()},
\code{\link{vm_sev_std_pb.mvn_pcci_task}()},
\code{\link{vm_sev_std_pb.mvn_simulation}()},
\code{\link{vm_sev_std_pb.mvn_task}()},
\code{\link{vm_sev_std_pb.mvn}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{parametric functions}
\keyword{pb}
