% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvn_mar_unstd_fit.sem.R
\name{mvn_mar_fit.sem_task}
\alias{mvn_mar_fit.sem_task}
\title{Fit Simple Mediation Model
for Data Generated from a Multivariate Normal Distribution
with Data Missing at Random - Structural Equation Modeling
(Single Task)}
\usage{
mvn_mar_fit.sem_task(taskid, dir = getwd(), overwrite = FALSE)
}
\arguments{
\item{taskid}{Numeric.
Task ID.}

\item{dir}{Character string.
Directory where results of \verb{mvn_dat_*} are stored.}

\item{overwrite}{Logical.
Overwrite existing results.}
}
\description{
Fit Simple Mediation Model
for Data Generated from a Multivariate Normal Distribution
with Data Missing at Random - Structural Equation Modeling
(Single Task)
}
\seealso{
Other model fit functions: 
\code{\link{fit.cov}()},
\code{\link{fit.ols}()},
\code{\link{fit.sem}()},
\code{\link{mvn_fit.ols_simulation_summary}()},
\code{\link{mvn_fit.ols_simulation}()},
\code{\link{mvn_fit.ols_task_summary}()},
\code{\link{mvn_fit.ols_task}()},
\code{\link{mvn_fit.ols}()},
\code{\link{mvn_fit.sem_simulation_summary}()},
\code{\link{mvn_fit.sem_simulation}()},
\code{\link{mvn_fit.sem_task_summary}()},
\code{\link{mvn_fit.sem_task}()},
\code{\link{mvn_fit.sem}()},
\code{\link{mvn_mar_10_fit.sem_simulation_summary}()},
\code{\link{mvn_mar_10_fit.sem_simulation}()},
\code{\link{mvn_mar_10_fit.sem_task_summary}()},
\code{\link{mvn_mar_10_fit.sem_task}()},
\code{\link{mvn_mar_10_fit.sem}()},
\code{\link{mvn_mar_20_fit.sem_simulation_summary}()},
\code{\link{mvn_mar_20_fit.sem_simulation}()},
\code{\link{mvn_mar_20_fit.sem_task_summary}()},
\code{\link{mvn_mar_20_fit.sem_task}()},
\code{\link{mvn_mar_20_fit.sem}()},
\code{\link{mvn_mar_30_fit.sem_simulation_summary}()},
\code{\link{mvn_mar_30_fit.sem_simulation}()},
\code{\link{mvn_mar_30_fit.sem_task_summary}()},
\code{\link{mvn_mar_30_fit.sem_task}()},
\code{\link{mvn_mar_30_fit.sem}()},
\code{\link{mvn_mar_50_fit.sem_simulation_summary}()},
\code{\link{mvn_mar_50_fit.sem_simulation}()},
\code{\link{mvn_mar_50_fit.sem_task_summary}()},
\code{\link{mvn_mar_50_fit.sem_task}()},
\code{\link{mvn_mar_50_fit.sem}()},
\code{\link{mvn_mar_fit.sem_simulation_summary}()},
\code{\link{mvn_mar_fit.sem_simulation}()},
\code{\link{mvn_mar_fit.sem_task_summary}()},
\code{\link{mvn_mar_fit.sem}()},
\code{\link{mvn_mar_std_fit.sem_simulation_summary}()},
\code{\link{mvn_mar_std_fit.sem_simulation}()},
\code{\link{mvn_mar_std_fit.sem_task_summary}()},
\code{\link{mvn_mar_std_fit.sem_task}()},
\code{\link{mvn_mar_std_fit.sem}()},
\code{\link{mvn_mcar_10_fit.sem_simulation_summary}()},
\code{\link{mvn_mcar_10_fit.sem_simulation}()},
\code{\link{mvn_mcar_10_fit.sem_task_summary}()},
\code{\link{mvn_mcar_10_fit.sem_task}()},
\code{\link{mvn_mcar_10_fit.sem}()},
\code{\link{mvn_mcar_20_fit.sem_simulation_summary}()},
\code{\link{mvn_mcar_20_fit.sem_simulation}()},
\code{\link{mvn_mcar_20_fit.sem_task_summary}()},
\code{\link{mvn_mcar_20_fit.sem_task}()},
\code{\link{mvn_mcar_20_fit.sem}()},
\code{\link{mvn_mcar_30_fit.sem_simulation_summary}()},
\code{\link{mvn_mcar_30_fit.sem_simulation}()},
\code{\link{mvn_mcar_30_fit.sem_task_summary}()},
\code{\link{mvn_mcar_30_fit.sem_task}()},
\code{\link{mvn_mcar_30_fit.sem}()},
\code{\link{mvn_mcar_50_fit.sem_simulation_summary}()},
\code{\link{mvn_mcar_50_fit.sem_simulation}()},
\code{\link{mvn_mcar_50_fit.sem_task_summary}()},
\code{\link{mvn_mcar_50_fit.sem_task}()},
\code{\link{mvn_mcar_50_fit.sem}()},
\code{\link{mvn_mcar_fit.sem_simulation_summary}()},
\code{\link{mvn_mcar_fit.sem_simulation}()},
\code{\link{mvn_mcar_fit.sem_task_summary}()},
\code{\link{mvn_mcar_fit.sem_task}()},
\code{\link{mvn_mcar_fit.sem}()},
\code{\link{mvn_mcar_std_fit.sem_simulation_summary}()},
\code{\link{mvn_mcar_std_fit.sem_simulation}()},
\code{\link{mvn_mcar_std_fit.sem_task_summary}()},
\code{\link{mvn_mcar_std_fit.sem_task}()},
\code{\link{mvn_mcar_std_fit.sem}()},
\code{\link{mvn_mnar_10_fit.sem_simulation_summary}()},
\code{\link{mvn_mnar_10_fit.sem_simulation}()},
\code{\link{mvn_mnar_10_fit.sem_task_summary}()},
\code{\link{mvn_mnar_10_fit.sem_task}()},
\code{\link{mvn_mnar_10_fit.sem}()},
\code{\link{mvn_mnar_20_fit.sem_simulation_summary}()},
\code{\link{mvn_mnar_20_fit.sem_simulation}()},
\code{\link{mvn_mnar_20_fit.sem_task_summary}()},
\code{\link{mvn_mnar_20_fit.sem_task}()},
\code{\link{mvn_mnar_20_fit.sem}()},
\code{\link{mvn_mnar_30_fit.sem_simulation_summary}()},
\code{\link{mvn_mnar_30_fit.sem_simulation}()},
\code{\link{mvn_mnar_30_fit.sem_task_summary}()},
\code{\link{mvn_mnar_30_fit.sem_task}()},
\code{\link{mvn_mnar_30_fit.sem}()},
\code{\link{mvn_mnar_50_fit.sem_simulation_summary}()},
\code{\link{mvn_mnar_50_fit.sem_simulation}()},
\code{\link{mvn_mnar_50_fit.sem_task_summary}()},
\code{\link{mvn_mnar_50_fit.sem_task}()},
\code{\link{mvn_mnar_50_fit.sem}()},
\code{\link{mvn_mnar_fit.sem_simulation_summary}()},
\code{\link{mvn_mnar_fit.sem_simulation}()},
\code{\link{mvn_mnar_fit.sem_task_summary}()},
\code{\link{mvn_mnar_fit.sem_task}()},
\code{\link{mvn_mnar_fit.sem}()},
\code{\link{mvn_mnar_std_fit.sem_simulation_summary}()},
\code{\link{mvn_mnar_std_fit.sem_simulation}()},
\code{\link{mvn_mnar_std_fit.sem_task_summary}()},
\code{\link{mvn_mnar_std_fit.sem_task}()},
\code{\link{mvn_mnar_std_fit.sem}()},
\code{\link{mvn_std_fit.sem_simulation_summary}()},
\code{\link{mvn_std_fit.sem_simulation}()},
\code{\link{mvn_std_fit.sem_task_summary}()},
\code{\link{mvn_std_fit.sem_task}()},
\code{\link{mvn_std_fit.sem}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{model fit functions}
\keyword{mvn_mar_unstd}
