% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvn_mar_unstd_pb.mvn.R
\name{mvn_mar_pb.mvn}
\alias{mvn_mar_pb.mvn}
\title{Parametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution
with Data Missing at Random}
\usage{
mvn_mar_pb.mvn(data, taskid)
}
\arguments{
\item{data}{\code{n} by 3 matrix or data frame.
\code{data[, 1]} correspond to values for \code{x}.
\code{data[, 2]} correspond to values for \code{m}.
\code{data[, 3]} correspond to values for \code{y}.}

\item{taskid}{Numeric.
Task ID.}
}
\description{
Parametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution
with Data Missing at Random
}
\examples{
taskid <- 1
data <- mvn_dat(taskid = taskid)
data <- mvn_dat_mar(data = data, taskid = taskid)
mvn_mar_pb.mvn(data = data, taskid = taskid)
}
\seealso{
Other parametric functions: 
\code{\link{mvn_mar_10_pb.mvn_bcci_simulation}()},
\code{\link{mvn_mar_10_pb.mvn_bcci_task}()},
\code{\link{mvn_mar_10_pb.mvn_pcci_simulation}()},
\code{\link{mvn_mar_10_pb.mvn_pcci_task}()},
\code{\link{mvn_mar_10_pb.mvn_simulation}()},
\code{\link{mvn_mar_10_pb.mvn_task}()},
\code{\link{mvn_mar_10_pb.mvn}()},
\code{\link{mvn_mar_20_pb.mvn_bcci_simulation}()},
\code{\link{mvn_mar_20_pb.mvn_bcci_task}()},
\code{\link{mvn_mar_20_pb.mvn_pcci_simulation}()},
\code{\link{mvn_mar_20_pb.mvn_pcci_task}()},
\code{\link{mvn_mar_20_pb.mvn_simulation}()},
\code{\link{mvn_mar_20_pb.mvn_task}()},
\code{\link{mvn_mar_20_pb.mvn}()},
\code{\link{mvn_mar_30_pb.mvn_bcci_simulation}()},
\code{\link{mvn_mar_30_pb.mvn_bcci_task}()},
\code{\link{mvn_mar_30_pb.mvn_pcci_simulation}()},
\code{\link{mvn_mar_30_pb.mvn_pcci_task}()},
\code{\link{mvn_mar_30_pb.mvn_simulation}()},
\code{\link{mvn_mar_30_pb.mvn_task}()},
\code{\link{mvn_mar_30_pb.mvn}()},
\code{\link{mvn_mar_pb.mvn_bcci_simulation}()},
\code{\link{mvn_mar_pb.mvn_bcci_task}()},
\code{\link{mvn_mar_pb.mvn_pcci_simulation}()},
\code{\link{mvn_mar_pb.mvn_pcci_task}()},
\code{\link{mvn_mar_pb.mvn_simulation}()},
\code{\link{mvn_mar_pb.mvn_task}()},
\code{\link{mvn_mar_std_pb.mvn_bcci_simulation}()},
\code{\link{mvn_mar_std_pb.mvn_bcci_task}()},
\code{\link{mvn_mar_std_pb.mvn_pcci_simulation}()},
\code{\link{mvn_mar_std_pb.mvn_pcci_task}()},
\code{\link{mvn_mar_std_pb.mvn_simulation}()},
\code{\link{mvn_mar_std_pb.mvn_task}()},
\code{\link{mvn_mar_std_pb.mvn}()},
\code{\link{mvn_mcar_10_pb.mvn_bcci_simulation}()},
\code{\link{mvn_mcar_10_pb.mvn_bcci_task}()},
\code{\link{mvn_mcar_10_pb.mvn_pcci_simulation}()},
\code{\link{mvn_mcar_10_pb.mvn_pcci_task}()},
\code{\link{mvn_mcar_10_pb.mvn_simulation}()},
\code{\link{mvn_mcar_10_pb.mvn_task}()},
\code{\link{mvn_mcar_10_pb.mvn}()},
\code{\link{mvn_mcar_20_pb.mvn_bcci_simulation}()},
\code{\link{mvn_mcar_20_pb.mvn_bcci_task}()},
\code{\link{mvn_mcar_20_pb.mvn_pcci_simulation}()},
\code{\link{mvn_mcar_20_pb.mvn_pcci_task}()},
\code{\link{mvn_mcar_20_pb.mvn_simulation}()},
\code{\link{mvn_mcar_20_pb.mvn_task}()},
\code{\link{mvn_mcar_20_pb.mvn}()},
\code{\link{mvn_mcar_30_pb.mvn_bcci_simulation}()},
\code{\link{mvn_mcar_30_pb.mvn_bcci_task}()},
\code{\link{mvn_mcar_30_pb.mvn_pcci_simulation}()},
\code{\link{mvn_mcar_30_pb.mvn_pcci_task}()},
\code{\link{mvn_mcar_30_pb.mvn_simulation}()},
\code{\link{mvn_mcar_30_pb.mvn_task}()},
\code{\link{mvn_mcar_30_pb.mvn}()},
\code{\link{mvn_mcar_pb.mvn_bcci_simulation}()},
\code{\link{mvn_mcar_pb.mvn_bcci_task}()},
\code{\link{mvn_mcar_pb.mvn_pcci_simulation}()},
\code{\link{mvn_mcar_pb.mvn_pcci_task}()},
\code{\link{mvn_mcar_pb.mvn_simulation}()},
\code{\link{mvn_mcar_pb.mvn_task}()},
\code{\link{mvn_mcar_pb.mvn}()},
\code{\link{mvn_mcar_std_pb.mvn_bcci_simulation}()},
\code{\link{mvn_mcar_std_pb.mvn_bcci_task}()},
\code{\link{mvn_mcar_std_pb.mvn_pcci_simulation}()},
\code{\link{mvn_mcar_std_pb.mvn_pcci_task}()},
\code{\link{mvn_mcar_std_pb.mvn_simulation}()},
\code{\link{mvn_mcar_std_pb.mvn_task}()},
\code{\link{mvn_mcar_std_pb.mvn}()},
\code{\link{mvn_mnar_10_pb.mvn_bcci_simulation}()},
\code{\link{mvn_mnar_10_pb.mvn_bcci_task}()},
\code{\link{mvn_mnar_10_pb.mvn_pcci_simulation}()},
\code{\link{mvn_mnar_10_pb.mvn_pcci_task}()},
\code{\link{mvn_mnar_10_pb.mvn_simulation}()},
\code{\link{mvn_mnar_10_pb.mvn_task}()},
\code{\link{mvn_mnar_10_pb.mvn}()},
\code{\link{mvn_mnar_20_pb.mvn_bcci_simulation}()},
\code{\link{mvn_mnar_20_pb.mvn_bcci_task}()},
\code{\link{mvn_mnar_20_pb.mvn_pcci_simulation}()},
\code{\link{mvn_mnar_20_pb.mvn_pcci_task}()},
\code{\link{mvn_mnar_20_pb.mvn_simulation}()},
\code{\link{mvn_mnar_20_pb.mvn_task}()},
\code{\link{mvn_mnar_20_pb.mvn}()},
\code{\link{mvn_mnar_30_pb.mvn_bcci_simulation}()},
\code{\link{mvn_mnar_30_pb.mvn_bcci_task}()},
\code{\link{mvn_mnar_30_pb.mvn_pcci_simulation}()},
\code{\link{mvn_mnar_30_pb.mvn_pcci_task}()},
\code{\link{mvn_mnar_30_pb.mvn_simulation}()},
\code{\link{mvn_mnar_30_pb.mvn_task}()},
\code{\link{mvn_mnar_30_pb.mvn}()},
\code{\link{mvn_mnar_pb.mvn_bcci_simulation}()},
\code{\link{mvn_mnar_pb.mvn_bcci_task}()},
\code{\link{mvn_mnar_pb.mvn_pcci_simulation}()},
\code{\link{mvn_mnar_pb.mvn_pcci_task}()},
\code{\link{mvn_mnar_pb.mvn_simulation}()},
\code{\link{mvn_mnar_pb.mvn_task}()},
\code{\link{mvn_mnar_pb.mvn}()},
\code{\link{mvn_pb.mvn_bcaci_simulation}()},
\code{\link{mvn_pb.mvn_bcaci_task}()},
\code{\link{mvn_pb.mvn_bcci_simulation}()},
\code{\link{mvn_pb.mvn_bcci_task}()},
\code{\link{mvn_pb.mvn_pcci_simulation}()},
\code{\link{mvn_pb.mvn_pcci_task}()},
\code{\link{mvn_pb.mvn_simulation}()},
\code{\link{mvn_pb.mvn_task}()},
\code{\link{mvn_pb.mvn}()},
\code{\link{mvn_std_pb.mvn_bcaci_simulation}()},
\code{\link{mvn_std_pb.mvn_bcaci_task}()},
\code{\link{mvn_std_pb.mvn_bcci_simulation}()},
\code{\link{mvn_std_pb.mvn_bcci_task}()},
\code{\link{mvn_std_pb.mvn_pcci_simulation}()},
\code{\link{mvn_std_pb.mvn_pcci_task}()},
\code{\link{mvn_std_pb.mvn_simulation}()},
\code{\link{mvn_std_pb.mvn_task}()},
\code{\link{mvn_std_pb.mvn}()},
\code{\link{pb.mvn}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{parametric functions}
\keyword{mvn_mar_unstd}
