% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvn_mcar_20_dat.R
\name{mvn_mcar_20_dat_task}
\alias{mvn_mcar_20_dat_task}
\title{Generate Data from a Multivariate Normal Distribution
following a Simple Mediation Model
with Data Missing Completely at Random - 20\%
(Single Task)}
\usage{
mvn_mcar_20_dat_task(taskid, dir = getwd(), overwrite = FALSE)
}
\arguments{
\item{taskid}{Numeric.
Task ID.}

\item{dir}{Character string.
Directory where results of \verb{mvn_dat_*} are stored.}

\item{overwrite}{Logical.
Overwrite existing results.}
}
\description{
Generate Data from a Multivariate Normal Distribution
following a Simple Mediation Model
with Data Missing Completely at Random - 20\%
(Single Task)
}
\seealso{
Other missing completely at random: 
\code{\link{mvn_mcar_10_dat_simulation}()},
\code{\link{mvn_mcar_10_dat_task}()},
\code{\link{mvn_mcar_10_dat}()},
\code{\link{mvn_mcar_20_dat_simulation}()},
\code{\link{mvn_mcar_20_dat}()},
\code{\link{mvn_mcar_30_dat_simulation}()},
\code{\link{mvn_mcar_30_dat_task}()},
\code{\link{mvn_mcar_30_dat}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{missing completely at random}
\keyword{mvn_dat}
