% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvn_mcar_unstd_nb.R
\name{mvn_mcar_nb}
\alias{mvn_mcar_nb}
\title{Nonparametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution
with Data Missing Completely at Random}
\usage{
mvn_mcar_nb(data, taskid)
}
\arguments{
\item{data}{\code{n} by 3 matrix or data frame.
\code{data[, 1]} correspond to values for \code{x}.
\code{data[, 2]} correspond to values for \code{m}.
\code{data[, 3]} correspond to values for \code{y}.}

\item{taskid}{Numeric.
Task ID.}
}
\description{
Nonparametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution
with Data Missing Completely at Random
}
\examples{
taskid <- 1
data <- mvn_dat(taskid = taskid)
mvn_mcar_nb(data = data, taskid = taskid)
}
\seealso{
Other nonparametric functions: 
\code{\link{mvn_mar_nb_bcaci_simulation}()},
\code{\link{mvn_mar_nb_bcaci_task}()},
\code{\link{mvn_mar_nb_bcci_simulation}()},
\code{\link{mvn_mar_nb_bcci_task}()},
\code{\link{mvn_mar_nb_pcci_simulation}()},
\code{\link{mvn_mar_nb_pcci_task}()},
\code{\link{mvn_mar_nb_simulation}()},
\code{\link{mvn_mar_nb_task}()},
\code{\link{mvn_mar_nb}()},
\code{\link{mvn_mcar_nb_bcaci_simulation}()},
\code{\link{mvn_mcar_nb_bcaci_task}()},
\code{\link{mvn_mcar_nb_bcci_simulation}()},
\code{\link{mvn_mcar_nb_bcci_task}()},
\code{\link{mvn_mcar_nb_fiml_bcaci_simulation}()},
\code{\link{mvn_mcar_nb_fiml_bcaci_task}()},
\code{\link{mvn_mcar_nb_fiml_bcci_simulation}()},
\code{\link{mvn_mcar_nb_fiml_bcci_task}()},
\code{\link{mvn_mcar_nb_fiml_pcci_simulation}()},
\code{\link{mvn_mcar_nb_fiml_pcci_task}()},
\code{\link{mvn_mcar_nb_fiml_simulation}()},
\code{\link{mvn_mcar_nb_fiml_task}()},
\code{\link{mvn_mcar_nb_fiml}()},
\code{\link{mvn_mcar_nb_pcci_simulation}()},
\code{\link{mvn_mcar_nb_pcci_task}()},
\code{\link{mvn_mcar_nb_simulation}()},
\code{\link{mvn_mcar_nb_task}()},
\code{\link{mvn_mnar_nb_bcaci_simulation}()},
\code{\link{mvn_mnar_nb_bcaci_task}()},
\code{\link{mvn_mnar_nb_bcci_simulation}()},
\code{\link{mvn_mnar_nb_bcci_task}()},
\code{\link{mvn_mnar_nb_pcci_simulation}()},
\code{\link{mvn_mnar_nb_pcci_task}()},
\code{\link{mvn_mnar_nb_simulation}()},
\code{\link{mvn_mnar_nb_task}()},
\code{\link{mvn_mnar_nb}()},
\code{\link{mvn_nb_bcaci_simulation}()},
\code{\link{mvn_nb_bcaci_task}()},
\code{\link{mvn_nb_bcci_simulation}()},
\code{\link{mvn_nb_bcci_task}()},
\code{\link{mvn_nb_pcci_simulation}()},
\code{\link{mvn_nb_pcci_task}()},
\code{\link{mvn_nb_simulation}()},
\code{\link{mvn_nb_task}()},
\code{\link{mvn_nb}()},
\code{\link{mvn_std_nb_bcaci_simulation}()},
\code{\link{mvn_std_nb_bcaci_task}()},
\code{\link{mvn_std_nb_bcci_simulation}()},
\code{\link{mvn_std_nb_bcci_task}()},
\code{\link{mvn_std_nb_pcci_simulation}()},
\code{\link{mvn_std_nb_pcci_task}()},
\code{\link{mvn_std_nb_simulation}()},
\code{\link{mvn_std_nb_task}()},
\code{\link{mvn_std_nb}()},
\code{\link{nb}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{nonparametric functions}
\keyword{mvn_complete_unstd}
