% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvn_mnar_10_unstd_fit.sem.R
\name{mvn_mnar_10_fit.sem_task_summary}
\alias{mvn_mnar_10_fit.sem_task_summary}
\title{Fit Simple Mediation Model
for Data Generated from a Multivariate Normal Distribution
with Data Missing Not at Random - 10\% - Structural Equation Modeling
(Single Task Summary)}
\usage{
mvn_mnar_10_fit.sem_task_summary(taskid, dir = getwd())
}
\arguments{
\item{taskid}{Numeric.
Task ID.}

\item{dir}{Character string.
Directory where results of \verb{mvn_dat_*} are stored.}
}
\description{
Fit Simple Mediation Model
for Data Generated from a Multivariate Normal Distribution
with Data Missing Not at Random - 10\% - Structural Equation Modeling
(Single Task Summary)
}
\seealso{
Other model fit functions: 
\code{\link{beta_fit.ols_simulation_summary}()},
\code{\link{beta_fit.ols_simulation}()},
\code{\link{beta_fit.ols_task_summary}()},
\code{\link{beta_fit.ols_task}()},
\code{\link{beta_fit.ols}()},
\code{\link{beta_fit.sem.mlr_simulation_summary}()},
\code{\link{beta_fit.sem.mlr_simulation}()},
\code{\link{beta_fit.sem.mlr_task_summary}()},
\code{\link{beta_fit.sem.mlr_task}()},
\code{\link{beta_fit.sem.mlr}()},
\code{\link{beta_std_fit.sem.mlr_simulation_summary}()},
\code{\link{beta_std_fit.sem.mlr_simulation}()},
\code{\link{beta_std_fit.sem.mlr_task_summary}()},
\code{\link{beta_std_fit.sem.mlr_task}()},
\code{\link{beta_std_fit.sem.mlr}()},
\code{\link{exp_fit.ols_simulation_summary}()},
\code{\link{exp_fit.ols_simulation}()},
\code{\link{exp_fit.ols_task_summary}()},
\code{\link{exp_fit.ols_task}()},
\code{\link{exp_fit.ols}()},
\code{\link{exp_fit.sem.mlr_simulation_summary}()},
\code{\link{exp_fit.sem.mlr_simulation}()},
\code{\link{exp_fit.sem.mlr_task_summary}()},
\code{\link{exp_fit.sem.mlr_task}()},
\code{\link{exp_fit.sem.mlr}()},
\code{\link{exp_std_fit.sem.mlr_simulation_summary}()},
\code{\link{exp_std_fit.sem.mlr_simulation}()},
\code{\link{exp_std_fit.sem.mlr_task_summary}()},
\code{\link{exp_std_fit.sem.mlr_task}()},
\code{\link{exp_std_fit.sem.mlr}()},
\code{\link{fit.cov}()},
\code{\link{fit.ols}()},
\code{\link{fit.sem.mlr}()},
\code{\link{fit.sem}()},
\code{\link{mvn_fit.ols_simulation_summary}()},
\code{\link{mvn_fit.ols_simulation}()},
\code{\link{mvn_fit.ols_task_summary}()},
\code{\link{mvn_fit.ols_task}()},
\code{\link{mvn_fit.ols}()},
\code{\link{mvn_fit.sem_simulation_summary}()},
\code{\link{mvn_fit.sem_simulation}()},
\code{\link{mvn_fit.sem_task_summary}()},
\code{\link{mvn_fit.sem_task}()},
\code{\link{mvn_fit.sem}()},
\code{\link{mvn_mar_10_fit.sem_simulation_summary}()},
\code{\link{mvn_mar_10_fit.sem_simulation}()},
\code{\link{mvn_mar_10_fit.sem_task_summary}()},
\code{\link{mvn_mar_10_fit.sem_task}()},
\code{\link{mvn_mar_10_fit.sem}()},
\code{\link{mvn_mar_20_fit.sem_simulation_summary}()},
\code{\link{mvn_mar_20_fit.sem_simulation}()},
\code{\link{mvn_mar_20_fit.sem_task_summary}()},
\code{\link{mvn_mar_20_fit.sem_task}()},
\code{\link{mvn_mar_20_fit.sem}()},
\code{\link{mvn_mar_30_fit.sem_simulation_summary}()},
\code{\link{mvn_mar_30_fit.sem_simulation}()},
\code{\link{mvn_mar_30_fit.sem_task_summary}()},
\code{\link{mvn_mar_30_fit.sem_task}()},
\code{\link{mvn_mar_30_fit.sem}()},
\code{\link{mvn_mcar_10_fit.sem_simulation_summary}()},
\code{\link{mvn_mcar_10_fit.sem_simulation}()},
\code{\link{mvn_mcar_10_fit.sem_task_summary}()},
\code{\link{mvn_mcar_10_fit.sem_task}()},
\code{\link{mvn_mcar_10_fit.sem}()},
\code{\link{mvn_mcar_20_fit.sem_simulation_summary}()},
\code{\link{mvn_mcar_20_fit.sem_simulation}()},
\code{\link{mvn_mcar_20_fit.sem_task_summary}()},
\code{\link{mvn_mcar_20_fit.sem_task}()},
\code{\link{mvn_mcar_20_fit.sem}()},
\code{\link{mvn_mcar_30_fit.sem_simulation_summary}()},
\code{\link{mvn_mcar_30_fit.sem_simulation}()},
\code{\link{mvn_mcar_30_fit.sem_task_summary}()},
\code{\link{mvn_mcar_30_fit.sem_task}()},
\code{\link{mvn_mcar_30_fit.sem}()},
\code{\link{mvn_mnar_10_fit.sem_simulation_summary}()},
\code{\link{mvn_mnar_10_fit.sem_simulation}()},
\code{\link{mvn_mnar_10_fit.sem_task}()},
\code{\link{mvn_mnar_10_fit.sem}()},
\code{\link{mvn_mnar_20_fit.sem_simulation_summary}()},
\code{\link{mvn_mnar_20_fit.sem_simulation}()},
\code{\link{mvn_mnar_20_fit.sem_task_summary}()},
\code{\link{mvn_mnar_20_fit.sem_task}()},
\code{\link{mvn_mnar_20_fit.sem}()},
\code{\link{mvn_mnar_30_fit.sem_simulation_summary}()},
\code{\link{mvn_mnar_30_fit.sem_simulation}()},
\code{\link{mvn_mnar_30_fit.sem_task_summary}()},
\code{\link{mvn_mnar_30_fit.sem_task}()},
\code{\link{mvn_mnar_30_fit.sem}()},
\code{\link{mvn_std_fit.sem_simulation_summary}()},
\code{\link{mvn_std_fit.sem_simulation}()},
\code{\link{mvn_std_fit.sem_task_summary}()},
\code{\link{mvn_std_fit.sem_task}()},
\code{\link{mvn_std_fit.sem}()},
\code{\link{vm_mod_fit.ols_simulation_summary}()},
\code{\link{vm_mod_fit.ols_simulation}()},
\code{\link{vm_mod_fit.ols_task_summary}()},
\code{\link{vm_mod_fit.ols_task}()},
\code{\link{vm_mod_fit.ols}()},
\code{\link{vm_mod_fit.sem.mlr_simulation_summary}()},
\code{\link{vm_mod_fit.sem.mlr_simulation}()},
\code{\link{vm_mod_fit.sem.mlr_task_summary}()},
\code{\link{vm_mod_fit.sem.mlr_task}()},
\code{\link{vm_mod_fit.sem.mlr}()},
\code{\link{vm_mod_std_fit.sem.mlr_simulation_summary}()},
\code{\link{vm_mod_std_fit.sem.mlr_simulation}()},
\code{\link{vm_mod_std_fit.sem.mlr_task_summary}()},
\code{\link{vm_mod_std_fit.sem.mlr_task}()},
\code{\link{vm_mod_std_fit.sem.mlr}()},
\code{\link{vm_sev_fit.ols_simulation_summary}()},
\code{\link{vm_sev_fit.ols_simulation}()},
\code{\link{vm_sev_fit.ols_task_summary}()},
\code{\link{vm_sev_fit.ols_task}()},
\code{\link{vm_sev_fit.ols}()},
\code{\link{vm_sev_fit.sem.mlr_simulation_summary}()},
\code{\link{vm_sev_fit.sem.mlr_simulation}()},
\code{\link{vm_sev_fit.sem.mlr_task_summary}()},
\code{\link{vm_sev_fit.sem.mlr_task}()},
\code{\link{vm_sev_fit.sem.mlr}()},
\code{\link{vm_sev_std_fit.sem.mlr_simulation_summary}()},
\code{\link{vm_sev_std_fit.sem.mlr_simulation}()},
\code{\link{vm_sev_std_fit.sem.mlr_task_summary}()},
\code{\link{vm_sev_std_fit.sem.mlr_task}()},
\code{\link{vm_sev_std_fit.sem.mlr}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{model fit functions}
\keyword{fit}
