% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvn_mnar_20_unstd_pb.mvn.R
\name{mvn_mnar_20_pb.mvn}
\alias{mvn_mnar_20_pb.mvn}
\title{Parametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution
with Data Missing Completely at Random - 20\%}
\usage{
mvn_mnar_20_pb.mvn(data, taskid)
}
\arguments{
\item{data}{\code{n} by 3 matrix or data frame.
\code{data[, 1]} correspond to values for \code{x}.
\code{data[, 2]} correspond to values for \code{m}.
\code{data[, 3]} correspond to values for \code{y}.}

\item{taskid}{Numeric.
Task ID.}
}
\description{
Parametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution
with Data Missing Completely at Random - 20\%
}
\examples{
taskid <- 1
data <- mvn_dat(taskid = taskid)
data <- mvn_mnar_20_dat(data = data, taskid = taskid)
mvn_mnar_20_pb.mvn(data = data, taskid = taskid)
}
\seealso{
Other parametric functions: 
\code{\link{beta_pb.beta_bcaci_simulation}()},
\code{\link{beta_pb.beta_bcaci_task}()},
\code{\link{beta_pb.beta_bcci_simulation}()},
\code{\link{beta_pb.beta_bcci_task}()},
\code{\link{beta_pb.beta_pcci_simulation}()},
\code{\link{beta_pb.beta_pcci_task}()},
\code{\link{beta_pb.beta_simulation}()},
\code{\link{beta_pb.beta_task}()},
\code{\link{beta_pb.beta}()},
\code{\link{beta_pb.mvn_bcaci_simulation}()},
\code{\link{beta_pb.mvn_bcaci_task}()},
\code{\link{beta_pb.mvn_bcci_simulation}()},
\code{\link{beta_pb.mvn_bcci_task}()},
\code{\link{beta_pb.mvn_pcci_simulation}()},
\code{\link{beta_pb.mvn_pcci_task}()},
\code{\link{beta_pb.mvn_simulation}()},
\code{\link{beta_pb.mvn_task}()},
\code{\link{beta_pb.mvn}()},
\code{\link{exp_pb.exp_bcaci_simulation}()},
\code{\link{exp_pb.exp_bcaci_task}()},
\code{\link{exp_pb.exp_bcci_simulation}()},
\code{\link{exp_pb.exp_bcci_task}()},
\code{\link{exp_pb.exp_pcci_simulation}()},
\code{\link{exp_pb.exp_pcci_task}()},
\code{\link{exp_pb.exp_simulation}()},
\code{\link{exp_pb.exp_task}()},
\code{\link{exp_pb.exp}()},
\code{\link{exp_pb.mvn_bcaci_simulation}()},
\code{\link{exp_pb.mvn_bcaci_task}()},
\code{\link{exp_pb.mvn_bcci_simulation}()},
\code{\link{exp_pb.mvn_bcci_task}()},
\code{\link{exp_pb.mvn_pcci_simulation}()},
\code{\link{exp_pb.mvn_pcci_task}()},
\code{\link{exp_pb.mvn_simulation}()},
\code{\link{exp_pb.mvn_task}()},
\code{\link{exp_pb.mvn}()},
\code{\link{mvn_mar_10_pb.mvn_bcci_simulation}()},
\code{\link{mvn_mar_10_pb.mvn_bcci_task}()},
\code{\link{mvn_mar_10_pb.mvn_pcci_simulation}()},
\code{\link{mvn_mar_10_pb.mvn_pcci_task}()},
\code{\link{mvn_mar_10_pb.mvn_simulation}()},
\code{\link{mvn_mar_10_pb.mvn_task}()},
\code{\link{mvn_mar_10_pb.mvn}()},
\code{\link{mvn_mar_20_pb.mvn_bcci_simulation}()},
\code{\link{mvn_mar_20_pb.mvn_bcci_task}()},
\code{\link{mvn_mar_20_pb.mvn_pcci_simulation}()},
\code{\link{mvn_mar_20_pb.mvn_pcci_task}()},
\code{\link{mvn_mar_20_pb.mvn_simulation}()},
\code{\link{mvn_mar_20_pb.mvn_task}()},
\code{\link{mvn_mar_20_pb.mvn}()},
\code{\link{mvn_mar_30_pb.mvn_bcci_simulation}()},
\code{\link{mvn_mar_30_pb.mvn_bcci_task}()},
\code{\link{mvn_mar_30_pb.mvn_pcci_simulation}()},
\code{\link{mvn_mar_30_pb.mvn_pcci_task}()},
\code{\link{mvn_mar_30_pb.mvn_simulation}()},
\code{\link{mvn_mar_30_pb.mvn_task}()},
\code{\link{mvn_mar_30_pb.mvn}()},
\code{\link{mvn_mcar_10_pb.mvn_bcci_simulation}()},
\code{\link{mvn_mcar_10_pb.mvn_bcci_task}()},
\code{\link{mvn_mcar_10_pb.mvn_pcci_simulation}()},
\code{\link{mvn_mcar_10_pb.mvn_pcci_task}()},
\code{\link{mvn_mcar_10_pb.mvn_simulation}()},
\code{\link{mvn_mcar_10_pb.mvn_task}()},
\code{\link{mvn_mcar_10_pb.mvn}()},
\code{\link{mvn_mcar_20_pb.mvn_bcci_simulation}()},
\code{\link{mvn_mcar_20_pb.mvn_bcci_task}()},
\code{\link{mvn_mcar_20_pb.mvn_pcci_simulation}()},
\code{\link{mvn_mcar_20_pb.mvn_pcci_task}()},
\code{\link{mvn_mcar_20_pb.mvn_simulation}()},
\code{\link{mvn_mcar_20_pb.mvn_task}()},
\code{\link{mvn_mcar_20_pb.mvn}()},
\code{\link{mvn_mcar_30_pb.mvn_bcci_simulation}()},
\code{\link{mvn_mcar_30_pb.mvn_bcci_task}()},
\code{\link{mvn_mcar_30_pb.mvn_pcci_simulation}()},
\code{\link{mvn_mcar_30_pb.mvn_pcci_task}()},
\code{\link{mvn_mcar_30_pb.mvn_simulation}()},
\code{\link{mvn_mcar_30_pb.mvn_task}()},
\code{\link{mvn_mcar_30_pb.mvn}()},
\code{\link{mvn_mnar_10_pb.mvn_bcci_simulation}()},
\code{\link{mvn_mnar_10_pb.mvn_bcci_task}()},
\code{\link{mvn_mnar_10_pb.mvn_pcci_simulation}()},
\code{\link{mvn_mnar_10_pb.mvn_pcci_task}()},
\code{\link{mvn_mnar_10_pb.mvn_simulation}()},
\code{\link{mvn_mnar_10_pb.mvn_task}()},
\code{\link{mvn_mnar_10_pb.mvn}()},
\code{\link{mvn_mnar_20_pb.mvn_bcci_simulation}()},
\code{\link{mvn_mnar_20_pb.mvn_bcci_task}()},
\code{\link{mvn_mnar_20_pb.mvn_pcci_simulation}()},
\code{\link{mvn_mnar_20_pb.mvn_pcci_task}()},
\code{\link{mvn_mnar_20_pb.mvn_simulation}()},
\code{\link{mvn_mnar_20_pb.mvn_task}()},
\code{\link{mvn_mnar_30_pb.mvn_bcci_simulation}()},
\code{\link{mvn_mnar_30_pb.mvn_bcci_task}()},
\code{\link{mvn_mnar_30_pb.mvn_pcci_simulation}()},
\code{\link{mvn_mnar_30_pb.mvn_pcci_task}()},
\code{\link{mvn_mnar_30_pb.mvn_simulation}()},
\code{\link{mvn_mnar_30_pb.mvn_task}()},
\code{\link{mvn_mnar_30_pb.mvn}()},
\code{\link{mvn_pb.mvn_bcaci_simulation}()},
\code{\link{mvn_pb.mvn_bcaci_task}()},
\code{\link{mvn_pb.mvn_bcci_simulation}()},
\code{\link{mvn_pb.mvn_bcci_task}()},
\code{\link{mvn_pb.mvn_pcci_simulation}()},
\code{\link{mvn_pb.mvn_pcci_task}()},
\code{\link{mvn_pb.mvn_simulation}()},
\code{\link{mvn_pb.mvn_task}()},
\code{\link{mvn_pb.mvn}()},
\code{\link{mvn_std_pb.mvn_bcaci_simulation}()},
\code{\link{mvn_std_pb.mvn_bcaci_task}()},
\code{\link{mvn_std_pb.mvn_bcci_simulation}()},
\code{\link{mvn_std_pb.mvn_bcci_task}()},
\code{\link{mvn_std_pb.mvn_pcci_simulation}()},
\code{\link{mvn_std_pb.mvn_pcci_task}()},
\code{\link{mvn_std_pb.mvn_simulation}()},
\code{\link{mvn_std_pb.mvn_task}()},
\code{\link{mvn_std_pb.mvn}()},
\code{\link{pb.beta}()},
\code{\link{pb.exp}()},
\code{\link{pb.mvn}()},
\code{\link{pb.vm}()},
\code{\link{vm_mod_pb.mvn_bcaci_simulation}()},
\code{\link{vm_mod_pb.mvn_bcaci_task}()},
\code{\link{vm_mod_pb.mvn_bcci_simulation}()},
\code{\link{vm_mod_pb.mvn_bcci_task}()},
\code{\link{vm_mod_pb.mvn_pcci_simulation}()},
\code{\link{vm_mod_pb.mvn_pcci_task}()},
\code{\link{vm_mod_pb.mvn_simulation}()},
\code{\link{vm_mod_pb.mvn_task}()},
\code{\link{vm_mod_pb.mvn}()},
\code{\link{vm_mod_pb.vm_bcaci_simulation}()},
\code{\link{vm_mod_pb.vm_bcaci_task}()},
\code{\link{vm_mod_pb.vm_bcci_simulation}()},
\code{\link{vm_mod_pb.vm_bcci_task}()},
\code{\link{vm_mod_pb.vm_pcci_simulation}()},
\code{\link{vm_mod_pb.vm_pcci_task}()},
\code{\link{vm_mod_pb.vm_simulation}()},
\code{\link{vm_mod_pb.vm_task}()},
\code{\link{vm_mod_pb.vm}()},
\code{\link{vm_mod_std_pb.mvn_bcaci_simulation}()},
\code{\link{vm_mod_std_pb.mvn_bcaci_task}()},
\code{\link{vm_mod_std_pb.mvn_bcci_simulation}()},
\code{\link{vm_mod_std_pb.mvn_bcci_task}()},
\code{\link{vm_mod_std_pb.mvn_pcci_simulation}()},
\code{\link{vm_mod_std_pb.mvn_pcci_task}()},
\code{\link{vm_mod_std_pb.mvn_simulation}()},
\code{\link{vm_mod_std_pb.mvn_task}()},
\code{\link{vm_mod_std_pb.mvn}()},
\code{\link{vm_sev_pb.mvn_bcaci_simulation}()},
\code{\link{vm_sev_pb.mvn_bcaci_task}()},
\code{\link{vm_sev_pb.mvn_bcci_simulation}()},
\code{\link{vm_sev_pb.mvn_bcci_task}()},
\code{\link{vm_sev_pb.mvn_pcci_simulation}()},
\code{\link{vm_sev_pb.mvn_pcci_task}()},
\code{\link{vm_sev_pb.mvn_simulation}()},
\code{\link{vm_sev_pb.mvn_task}()},
\code{\link{vm_sev_pb.mvn}()},
\code{\link{vm_sev_pb.vm_bcaci_simulation}()},
\code{\link{vm_sev_pb.vm_bcaci_task}()},
\code{\link{vm_sev_pb.vm_bcci_simulation}()},
\code{\link{vm_sev_pb.vm_bcci_task}()},
\code{\link{vm_sev_pb.vm_pcci_simulation}()},
\code{\link{vm_sev_pb.vm_pcci_task}()},
\code{\link{vm_sev_pb.vm_simulation}()},
\code{\link{vm_sev_pb.vm_task}()},
\code{\link{vm_sev_pb.vm}()},
\code{\link{vm_sev_std_pb.mvn_bcaci_simulation}()},
\code{\link{vm_sev_std_pb.mvn_bcaci_task}()},
\code{\link{vm_sev_std_pb.mvn_bcci_simulation}()},
\code{\link{vm_sev_std_pb.mvn_bcci_task}()},
\code{\link{vm_sev_std_pb.mvn_pcci_simulation}()},
\code{\link{vm_sev_std_pb.mvn_pcci_task}()},
\code{\link{vm_sev_std_pb.mvn_simulation}()},
\code{\link{vm_sev_std_pb.mvn_task}()},
\code{\link{vm_sev_std_pb.mvn}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{parametric functions}
\keyword{pb}
