% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvn_mnar_30_dat.R
\name{mvn_mnar_30_dat}
\alias{mvn_mnar_30_dat}
\title{Generate Data from a Multivariate Normal Distribution
following a Simple Mediation Model
with Data Missing Not at Random - 30\%}
\usage{
mvn_mnar_30_dat(data, taskid)
}
\arguments{
\item{data}{\code{n} by 3 matrix or data frame.
\code{data[, 1]} correspond to values for \code{x}.
\code{data[, 2]} correspond to values for \code{m}.
\code{data[, 3]} correspond to values for \code{y}.}

\item{taskid}{Numeric.
Task ID.}
}
\description{
Generate Data from a Multivariate Normal Distribution
following a Simple Mediation Model
with Data Missing Not at Random - 30\%
}
\examples{
taskid <- 1
data <- mvn_dat(taskid = taskid)
mvn_mnar_30_dat(data = data, taskid = taskid)
}
\seealso{
Other missing not at random: 
\code{\link{mvn_mnar_10_dat_simulation}()},
\code{\link{mvn_mnar_10_dat_task}()},
\code{\link{mvn_mnar_10_dat}()},
\code{\link{mvn_mnar_20_dat_simulation}()},
\code{\link{mvn_mnar_20_dat_task}()},
\code{\link{mvn_mnar_20_dat}()},
\code{\link{mvn_mnar_30_dat_simulation}()},
\code{\link{mvn_mnar_30_dat_task}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{missing not at random}
\keyword{mvn_dat}
