% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvn_complete_unstd_nb.R
\name{mvn_nb_bcaci_task}
\alias{mvn_nb_bcaci_task}
\title{Nonparametric Bootstrap Bias-Corrected and Accelerated Confidence Intervals for Indirect Effect in a Simple Mediation Model
for Data Generated from a Multivariate Normal Distribution
(Single Task)}
\usage{
mvn_nb_bcaci_task(taskid, dir = getwd())
}
\arguments{
\item{taskid}{Numeric.
Task ID.}

\item{dir}{Character string.
Directory where results of \verb{mvn_dat_*} are stored.}
}
\description{
Nonparametric Bootstrap Bias-Corrected and Accelerated Confidence Intervals for Indirect Effect in a Simple Mediation Model
for Data Generated from a Multivariate Normal Distribution
(Single Task)
}
\seealso{
Other nonparametric functions: 
\code{\link{mvn_mar_nb_bcaci_simulation}()},
\code{\link{mvn_mar_nb_bcaci_task}()},
\code{\link{mvn_mar_nb_bcci_simulation}()},
\code{\link{mvn_mar_nb_bcci_task}()},
\code{\link{mvn_mar_nb_pcci_simulation}()},
\code{\link{mvn_mar_nb_pcci_task}()},
\code{\link{mvn_mar_nb_simulation}()},
\code{\link{mvn_mar_nb_task}()},
\code{\link{mvn_mar_nb}()},
\code{\link{mvn_mcar_nb_bcaci_simulation}()},
\code{\link{mvn_mcar_nb_bcaci_task}()},
\code{\link{mvn_mcar_nb_bcci_simulation}()},
\code{\link{mvn_mcar_nb_bcci_task}()},
\code{\link{mvn_mcar_nb_fiml_bcaci_simulation}()},
\code{\link{mvn_mcar_nb_fiml_bcaci_task}()},
\code{\link{mvn_mcar_nb_fiml_bcci_simulation}()},
\code{\link{mvn_mcar_nb_fiml_bcci_task}()},
\code{\link{mvn_mcar_nb_fiml_pcci_simulation}()},
\code{\link{mvn_mcar_nb_fiml_pcci_task}()},
\code{\link{mvn_mcar_nb_fiml_simulation}()},
\code{\link{mvn_mcar_nb_fiml_task}()},
\code{\link{mvn_mcar_nb_fiml}()},
\code{\link{mvn_mcar_nb_pcci_simulation}()},
\code{\link{mvn_mcar_nb_pcci_task}()},
\code{\link{mvn_mcar_nb_simulation}()},
\code{\link{mvn_mcar_nb_task}()},
\code{\link{mvn_mcar_nb}()},
\code{\link{mvn_mnar_nb_bcaci_simulation}()},
\code{\link{mvn_mnar_nb_bcaci_task}()},
\code{\link{mvn_mnar_nb_bcci_simulation}()},
\code{\link{mvn_mnar_nb_bcci_task}()},
\code{\link{mvn_mnar_nb_pcci_simulation}()},
\code{\link{mvn_mnar_nb_pcci_task}()},
\code{\link{mvn_mnar_nb_simulation}()},
\code{\link{mvn_mnar_nb_task}()},
\code{\link{mvn_mnar_nb}()},
\code{\link{mvn_nb_bcaci_simulation}()},
\code{\link{mvn_nb_bcci_simulation}()},
\code{\link{mvn_nb_bcci_task}()},
\code{\link{mvn_nb_pcci_simulation}()},
\code{\link{mvn_nb_pcci_task}()},
\code{\link{mvn_nb_simulation}()},
\code{\link{mvn_nb_task}()},
\code{\link{mvn_nb}()},
\code{\link{mvn_std_nb_bcaci_simulation}()},
\code{\link{mvn_std_nb_bcaci_task}()},
\code{\link{mvn_std_nb_bcci_simulation}()},
\code{\link{mvn_std_nb_bcci_task}()},
\code{\link{mvn_std_nb_pcci_simulation}()},
\code{\link{mvn_std_nb_pcci_task}()},
\code{\link{mvn_std_nb_simulation}()},
\code{\link{mvn_std_nb_task}()},
\code{\link{mvn_std_nb}()},
\code{\link{nb}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{nonparametric functions}
\keyword{mvn_complete_unstd}
