% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nb_mar.R
\name{mvn_nb_mar}
\alias{mvn_nb_mar}
\title{Nonparametric Bootstrapping Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution}
\usage{
mvn_nb_mar(data, taskid, B = 5000)
}
\arguments{
\item{data}{\code{n} by 3 matrix or data frame.
\code{data[, 1]} correspond to values for \code{x}.
\code{data[, 2]} correspond to values for \code{m}.
\code{data[, 3]} correspond to values for \code{y}.}

\item{taskid}{Numeric.
Task ID.}

\item{B}{Integer.
Number of bootstrap samples.}
}
\description{
Nonparametric Bootstrapping Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution
}
\examples{
taskid <- 1
data <- mvn_dat(taskid = taskid)
mvn_nb_mar(data = data, taskid = taskid, B = 200)
}
\seealso{
Other nonparametric functions: 
\code{\link{.nb}()},
\code{\link{mvn_nb_bcaci_simulation}()},
\code{\link{mvn_nb_bcaci_task}()},
\code{\link{mvn_nb_bcci_simulation}()},
\code{\link{mvn_nb_bcci_task}()},
\code{\link{mvn_nb_mar_bcaci_simulation}()},
\code{\link{mvn_nb_mar_bcaci_task}()},
\code{\link{mvn_nb_mar_bcci_simulation}()},
\code{\link{mvn_nb_mar_bcci_task}()},
\code{\link{mvn_nb_mar_pcci_simulation}()},
\code{\link{mvn_nb_mar_pcci_task}()},
\code{\link{mvn_nb_mar_simulation}()},
\code{\link{mvn_nb_mar_task}()},
\code{\link{mvn_nb_mcar_bcaci_simulation}()},
\code{\link{mvn_nb_mcar_bcaci_task}()},
\code{\link{mvn_nb_mcar_bcci_simulation}()},
\code{\link{mvn_nb_mcar_bcci_task}()},
\code{\link{mvn_nb_mcar_pcci_simulation}()},
\code{\link{mvn_nb_mcar_pcci_task}()},
\code{\link{mvn_nb_mcar_simulation}()},
\code{\link{mvn_nb_mcar_task}()},
\code{\link{mvn_nb_mcar}()},
\code{\link{mvn_nb_pcci_simulation}()},
\code{\link{mvn_nb_pcci_task}()},
\code{\link{mvn_nb_simulation}()},
\code{\link{mvn_nb_task}()},
\code{\link{mvn_nb}()},
\code{\link{mvn_pbmvn_bcaci_simulation}()},
\code{\link{mvn_pbmvn_bcaci_task}()},
\code{\link{mvn_pbmvn_bcci_simulation}()},
\code{\link{mvn_pbmvn_bcci_task}()},
\code{\link{mvn_pbmvn_mar_bcaci_simulation}()},
\code{\link{mvn_pbmvn_mar_bcaci_task}()},
\code{\link{mvn_pbmvn_mar_bcci_simulation}()},
\code{\link{mvn_pbmvn_mar_bcci_task}()},
\code{\link{mvn_pbmvn_mcar_bcaci_simulation}()},
\code{\link{mvn_pbmvn_mcar_bcaci_task}()},
\code{\link{mvn_pbmvn_mcar_bcci_simulation}()},
\code{\link{mvn_pbmvn_mcar_bcci_task}()},
\code{\link{mvn_std_nb_bcaci_simulation}()},
\code{\link{mvn_std_nb_bcaci_task}()},
\code{\link{mvn_std_nb_bcci_simulation}()},
\code{\link{mvn_std_nb_bcci_task}()},
\code{\link{mvn_std_nb_pcci_simulation}()},
\code{\link{mvn_std_nb_pcci_task}()},
\code{\link{mvn_std_nb_simulation}()},
\code{\link{mvn_std_nb_task}()},
\code{\link{mvn_std_nb}()},
\code{\link{mvn_std_pbmvn_bcaci_simulation}()},
\code{\link{mvn_std_pbmvn_bcaci_task}()},
\code{\link{mvn_std_pbmvn_bcci_simulation}()},
\code{\link{mvn_std_pbmvn_bcci_task}()},
\code{\link{nb}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{nonparametric functions}
\keyword{nb}
