% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nb_mcar.R
\name{mvn_nb_mcar_bcci_simulation}
\alias{mvn_nb_mcar_bcci_simulation}
\title{Nonparametric Bootstrap Bias-Corrected Confidence Intervals for Indirect Effect in a Simple Mediation Model
for Data Generated from a Multivariate Normal Distribution (Simulation)}
\usage{
mvn_nb_mcar_bcci_simulation(
  dir = getwd(),
  all = TRUE,
  taskid = NULL,
  alpha = c(0.001, 0.01, 0.05),
  par = TRUE,
  ncores = NULL,
  blas_threads = TRUE,
  mc = TRUE,
  lb = FALSE,
  cl_eval = FALSE,
  cl_export = FALSE,
  cl_expr,
  cl_vars
)
}
\arguments{
\item{dir}{Character string.
Directory where results of \verb{mvn_dat_*} are stored.}

\item{all}{Logical.
Process all results.}

\item{taskid}{Numeric.
Task ID.}

\item{alpha}{Numeric vector.
Alpha level.
By default \code{alpha = c(0.001, 0.01, 0.05)}.}

\item{par}{Logical.
If \code{TRUE}, use multiple cores.
If \code{FALSE}, use \code{\link[=lapply]{lapply()}}.}

\item{ncores}{Integer.
Number of cores to use if \code{par = TRUE}.
If unspecified, defaults to \code{detectCores() - 1}.}

\item{blas_threads}{Logical.
If \code{TRUE}, set BLAS threads using \code{blas_set_num_threads(threads = blas_get_num_procs())}.
If \code{FALSE}, set BLAS threads using \code{blas_set_num_threads(threads = 1)}.
If \code{par = TRUE}, \code{blas_threads} is automatically set to \code{FALSE}
to prevent conflicts in parallel processing.
This argument is useful when \code{FUN} can handle implicit parallelism
when \code{par = FALSE}, for example linear algebra operations.}

\item{mc}{Logical.
If \code{TRUE}, use \code{\link[parallel:mclapply]{parallel::mclapply()}}.
If \code{FALSE}, use \code{\link[parallel:clusterApply]{parallel::parLapply()}} or \code{\link[parallel:clusterApply]{parallel::parLapplyLB()}}.
Ignored if \code{par = FALSE}.}

\item{lb}{Logical.
If \code{TRUE} use \code{\link[parallel:clusterApply]{parallel::parLapplyLB()}}.
If \code{FALSE}, use \code{\link[parallel:clusterApply]{parallel::parLapply()}}.
Ignored if \code{par = FALSE} and \code{mc = TRUE}.}

\item{cl_eval}{Logical.
Execute \code{\link[parallel:clusterApply]{parallel::clusterEvalQ()}} using \code{cl_expr}.
Ignored if \code{mc = TRUE}.}

\item{cl_export}{Logical.
Execute \code{\link[parallel:clusterApply]{parallel::clusterExport()}} using \code{cl_vars}.
Ignored if \code{mc = TRUE}.}

\item{cl_expr}{Expression.
Expression passed to \code{\link[parallel:clusterApply]{parallel::clusterEvalQ()}}
Ignored if \code{mc = TRUE}.}

\item{cl_vars}{Character vector.
Names of objects to pass to \code{\link[parallel:clusterApply]{parallel::clusterExport()}}
Ignored if \code{mc = TRUE}.}
}
\description{
Nonparametric Bootstrap Bias-Corrected Confidence Intervals for Indirect Effect in a Simple Mediation Model
for Data Generated from a Multivariate Normal Distribution (Simulation)
}
\seealso{
Other nonparametric functions: 
\code{\link{.nb}()},
\code{\link{mvn_nb_bcaci_simulation}()},
\code{\link{mvn_nb_bcaci_task}()},
\code{\link{mvn_nb_bcci_simulation}()},
\code{\link{mvn_nb_bcci_task}()},
\code{\link{mvn_nb_mar_bcaci_simulation}()},
\code{\link{mvn_nb_mar_bcaci_task}()},
\code{\link{mvn_nb_mar_bcci_simulation}()},
\code{\link{mvn_nb_mar_bcci_task}()},
\code{\link{mvn_nb_mar_pcci_simulation}()},
\code{\link{mvn_nb_mar_pcci_task}()},
\code{\link{mvn_nb_mar_simulation}()},
\code{\link{mvn_nb_mar_task}()},
\code{\link{mvn_nb_mar}()},
\code{\link{mvn_nb_mcar_bcaci_simulation}()},
\code{\link{mvn_nb_mcar_bcaci_task}()},
\code{\link{mvn_nb_mcar_bcci_task}()},
\code{\link{mvn_nb_mcar_pcci_simulation}()},
\code{\link{mvn_nb_mcar_pcci_task}()},
\code{\link{mvn_nb_mcar_simulation}()},
\code{\link{mvn_nb_mcar_task}()},
\code{\link{mvn_nb_mcar}()},
\code{\link{mvn_nb_pcci_simulation}()},
\code{\link{mvn_nb_pcci_task}()},
\code{\link{mvn_nb_simulation}()},
\code{\link{mvn_nb_task}()},
\code{\link{mvn_nb}()},
\code{\link{mvn_pbmvn_bcaci_simulation}()},
\code{\link{mvn_pbmvn_bcaci_task}()},
\code{\link{mvn_pbmvn_bcci_simulation}()},
\code{\link{mvn_pbmvn_bcci_task}()},
\code{\link{mvn_pbmvn_mar_bcaci_simulation}()},
\code{\link{mvn_pbmvn_mar_bcaci_task}()},
\code{\link{mvn_pbmvn_mar_bcci_simulation}()},
\code{\link{mvn_pbmvn_mar_bcci_task}()},
\code{\link{mvn_pbmvn_mcar_bcaci_simulation}()},
\code{\link{mvn_pbmvn_mcar_bcaci_task}()},
\code{\link{mvn_pbmvn_mcar_bcci_simulation}()},
\code{\link{mvn_pbmvn_mcar_bcci_task}()},
\code{\link{mvn_std_nb_bcaci_simulation}()},
\code{\link{mvn_std_nb_bcaci_task}()},
\code{\link{mvn_std_nb_bcci_simulation}()},
\code{\link{mvn_std_nb_bcci_task}()},
\code{\link{mvn_std_nb_pcci_simulation}()},
\code{\link{mvn_std_nb_pcci_task}()},
\code{\link{mvn_std_nb_simulation}()},
\code{\link{mvn_std_nb_task}()},
\code{\link{mvn_std_nb}()},
\code{\link{mvn_std_pbmvn_bcaci_simulation}()},
\code{\link{mvn_std_pbmvn_bcaci_task}()},
\code{\link{mvn_std_pbmvn_bcci_simulation}()},
\code{\link{mvn_std_pbmvn_bcci_task}()},
\code{\link{nb}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{nonparametric functions}
\keyword{nb}
