% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nb.R
\name{mvn_nb_pcci_task}
\alias{mvn_nb_pcci_task}
\title{Nonparametric Bootstrap Percentile Confidence Intervals for Indirect Effect in a Simple Mediation Model
for Data Generated from a Multivariate Normal Distribution (Single Task)}
\usage{
mvn_nb_pcci_task(taskid, dir = getwd(), alpha = c(0.001, 0.01, 0.05))
}
\arguments{
\item{taskid}{Numeric.
Task ID.}

\item{dir}{Character string.
Directory where results of \verb{mvn_dat_*} are stored.}

\item{alpha}{Numeric vector.
Alpha level.
By default \code{alpha = c(0.001, 0.01, 0.05)}.}
}
\description{
Nonparametric Bootstrap Percentile Confidence Intervals for Indirect Effect in a Simple Mediation Model
for Data Generated from a Multivariate Normal Distribution (Single Task)
}
\seealso{
Other nonparametric functions: 
\code{\link{.nb}()},
\code{\link{mvn_nb_bcaci_simulation}()},
\code{\link{mvn_nb_bcaci_task}()},
\code{\link{mvn_nb_bcci_simulation}()},
\code{\link{mvn_nb_bcci_task}()},
\code{\link{mvn_nb_mar_bcaci_simulation}()},
\code{\link{mvn_nb_mar_bcaci_task}()},
\code{\link{mvn_nb_mar_bcci_simulation}()},
\code{\link{mvn_nb_mar_bcci_task}()},
\code{\link{mvn_nb_mar_pcci_simulation}()},
\code{\link{mvn_nb_mar_pcci_task}()},
\code{\link{mvn_nb_mar_simulation}()},
\code{\link{mvn_nb_mar_task}()},
\code{\link{mvn_nb_mar}()},
\code{\link{mvn_nb_mcar_bcaci_simulation}()},
\code{\link{mvn_nb_mcar_bcaci_task}()},
\code{\link{mvn_nb_mcar_bcci_simulation}()},
\code{\link{mvn_nb_mcar_bcci_task}()},
\code{\link{mvn_nb_mcar_pcci_simulation}()},
\code{\link{mvn_nb_mcar_pcci_task}()},
\code{\link{mvn_nb_mcar_simulation}()},
\code{\link{mvn_nb_mcar_task}()},
\code{\link{mvn_nb_mcar}()},
\code{\link{mvn_nb_pcci_simulation}()},
\code{\link{mvn_nb_simulation}()},
\code{\link{mvn_nb_task}()},
\code{\link{mvn_nb}()},
\code{\link{mvn_pbmvn_bcaci_simulation}()},
\code{\link{mvn_pbmvn_bcaci_task}()},
\code{\link{mvn_pbmvn_bcci_simulation}()},
\code{\link{mvn_pbmvn_bcci_task}()},
\code{\link{mvn_pbmvn_mar_bcaci_simulation}()},
\code{\link{mvn_pbmvn_mar_bcaci_task}()},
\code{\link{mvn_pbmvn_mar_bcci_simulation}()},
\code{\link{mvn_pbmvn_mar_bcci_task}()},
\code{\link{mvn_pbmvn_mcar_bcaci_simulation}()},
\code{\link{mvn_pbmvn_mcar_bcaci_task}()},
\code{\link{mvn_pbmvn_mcar_bcci_simulation}()},
\code{\link{mvn_pbmvn_mcar_bcci_task}()},
\code{\link{mvn_std_nb_bcaci_simulation}()},
\code{\link{mvn_std_nb_bcaci_task}()},
\code{\link{mvn_std_nb_bcci_simulation}()},
\code{\link{mvn_std_nb_bcci_task}()},
\code{\link{mvn_std_nb_pcci_simulation}()},
\code{\link{mvn_std_nb_pcci_task}()},
\code{\link{mvn_std_nb_simulation}()},
\code{\link{mvn_std_nb_task}()},
\code{\link{mvn_std_nb}()},
\code{\link{mvn_std_pbmvn_bcaci_simulation}()},
\code{\link{mvn_std_pbmvn_bcaci_task}()},
\code{\link{mvn_std_pbmvn_bcci_simulation}()},
\code{\link{mvn_std_pbmvn_bcci_task}()},
\code{\link{nb}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{nonparametric functions}
\keyword{nb}
