% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbmvn_mar.R
\name{mvn_pbmvn_mar_pcci_task}
\alias{mvn_pbmvn_mar_pcci_task}
\title{Parametric Bootstrap Percentile Confidence Intervals for Indirect Effect in a Simple Mediation Model Assuming Multivariate Normal Distribution
for Data Generated from a Multivariate Normal Distribution (Single Task)}
\usage{
mvn_pbmvn_mar_pcci_task(taskid, dir = getwd(), alpha = c(0.001, 0.01, 0.05))
}
\arguments{
\item{taskid}{Numeric.
Task ID.}

\item{alpha}{Numeric vector.
Alpha level.
By default \code{alpha = c(0.001, 0.01, 0.05)}.}
}
\description{
Parametric Bootstrap Percentile Confidence Intervals for Indirect Effect in a Simple Mediation Model Assuming Multivariate Normal Distribution
for Data Generated from a Multivariate Normal Distribution (Single Task)
}
\seealso{
Other parametric functions: 
\code{\link{.pbmvn}()},
\code{\link{mvn_pbmvn_mar_pcci_simulation}()},
\code{\link{mvn_pbmvn_mar_simulation}()},
\code{\link{mvn_pbmvn_mar_task}()},
\code{\link{mvn_pbmvn_mar}()},
\code{\link{mvn_pbmvn_mcar_pcci_simulation}()},
\code{\link{mvn_pbmvn_mcar_pcci_task}()},
\code{\link{mvn_pbmvn_mcar_simulation}()},
\code{\link{mvn_pbmvn_mcar_task}()},
\code{\link{mvn_pbmvn_mcar}()},
\code{\link{mvn_pbmvn_pcci_simulation}()},
\code{\link{mvn_pbmvn_pcci_task}()},
\code{\link{mvn_pbmvn_simulation}()},
\code{\link{mvn_pbmvn_task}()},
\code{\link{mvn_pbmvn}()},
\code{\link{mvn_std_pbmvn_pcci_simulation}()},
\code{\link{mvn_std_pbmvn_pcci_task}()},
\code{\link{mvn_std_pbmvn_simulation}()},
\code{\link{mvn_std_pbmvn_task}()},
\code{\link{mvn_std_pbmvn}()},
\code{\link{pbmvn}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{parametric functions}
\keyword{pb}
