% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbmvn_mcar.R
\name{mvn_pbmvn_mcar}
\alias{mvn_pbmvn_mcar}
\title{Parametric Bootstrapping Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution}
\usage{
mvn_pbmvn_mcar(data, taskid, B = 5000)
}
\arguments{
\item{data}{\code{n} by 3 matrix or data frame.
\code{data[, 1]} correspond to values for \code{x}.
\code{data[, 2]} correspond to values for \code{m}.
\code{data[, 3]} correspond to values for \code{y}.}

\item{taskid}{Numeric.
Task ID.}

\item{B}{Integer.
Number of bootstrap samples.}
}
\description{
Parametric Bootstrapping Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution
}
\examples{
taskid <- 1
data <- mvn_dat(taskid = taskid)
mvn_pbmvn_mcar(data = data, taskid = taskid, B = 200)
}
\seealso{
Other parametric functions: 
\code{\link{.pbmvn}()},
\code{\link{mvn_pbmvn_mar_pcci_simulation}()},
\code{\link{mvn_pbmvn_mar_pcci_task}()},
\code{\link{mvn_pbmvn_mar_simulation}()},
\code{\link{mvn_pbmvn_mar_task}()},
\code{\link{mvn_pbmvn_mar}()},
\code{\link{mvn_pbmvn_mcar_pcci_simulation}()},
\code{\link{mvn_pbmvn_mcar_pcci_task}()},
\code{\link{mvn_pbmvn_mcar_simulation}()},
\code{\link{mvn_pbmvn_mcar_task}()},
\code{\link{mvn_pbmvn_pcci_simulation}()},
\code{\link{mvn_pbmvn_pcci_task}()},
\code{\link{mvn_pbmvn_simulation}()},
\code{\link{mvn_pbmvn_task}()},
\code{\link{mvn_pbmvn}()},
\code{\link{mvn_std_pbmvn_pcci_simulation}()},
\code{\link{mvn_std_pbmvn_pcci_task}()},
\code{\link{mvn_std_pbmvn_simulation}()},
\code{\link{mvn_std_pbmvn_task}()},
\code{\link{mvn_std_pbmvn}()},
\code{\link{pbmvn}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{parametric functions}
\keyword{pb}
