% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvn_complete_std_mc.mvn.delta.R
\name{mvn_std_mc.mvn.delta}
\alias{mvn_std_mc.mvn.delta}
\title{Monte Carlo Method Assuming Multivariate Normal Distribution using Delta Method Standard Errors for Indirect Effect in a Standardized Simple Mediation Model
for Data Generated from a Multivariate Normal Distribution}
\usage{
mvn_std_mc.mvn.delta(
  taskid,
  R = 20000L,
  alphahatprime,
  sehatalphahatprimedelta,
  betahatprime,
  sehatbetahatprimedelta
)
}
\arguments{
\item{taskid}{Numeric.
Task ID.}

\item{R}{Integer.
Monte Carlo replications.}

\item{alphahatprime}{Numeric.
Estimated standardized slope of path from \code{x} to \code{m} \eqn{\left( \hat{\alpha}^{\prime} \right)} .}

\item{sehatalphahatprimedelta}{Numeric.
Estimated delta method standard error of standardized slope of path from \code{x} to \code{m} \eqn{\left( \widehat{se}_{\hat{\alpha}}^{\prime} \right)} .}

\item{betahatprime}{Numeric.
Estimated standardized slope of path from \code{m} to \code{y} \eqn{\left( \hat{\beta}^{\prime} \right)} .}

\item{sehatbetahatprimedelta}{Numeric.
Estimated delta method standard error of standardized slope of path from \code{m} to \code{y} \eqn{\left( \widehat{se}_{\hat{\beta}}^{\prime} \right)} .}
}
\description{
Monte Carlo Method Assuming Multivariate Normal Distribution using Delta Method Standard Errors for Indirect Effect in a Standardized Simple Mediation Model
for Data Generated from a Multivariate Normal Distribution
}
\examples{
taskid <- 1
data <- mvn_dat(taskid = taskid)
fit.ols(data = data, minimal = TRUE)

fit <- mvn_fit.ols(data = data, taskid = taskid)
thetahatstar <- mvn_std_mc.mvn.delta(
  taskid = taskid, R = 20000L,
  alphahatprime = fit["alphahatprime"], sehatalphahatprimedelta = fit["sehatalphahatprimedelta"],
  betahatprime = fit["betahatprime"], sehatbetahatprimedelta = fit["sehatbetahatprimedelta"]
)
hist(thetahatstar)
}
\seealso{
Other monte carlo method functions: 
\code{\link{mc.mvn}()},
\code{\link{mc.t}()},
\code{\link{mvn_mar_10_mc.mvn_pcci_simulation}()},
\code{\link{mvn_mar_10_mc.mvn_pcci_task}()},
\code{\link{mvn_mar_10_mc.mvn_simulation}()},
\code{\link{mvn_mar_10_mc.mvn_task}()},
\code{\link{mvn_mar_10_mc.mvn}()},
\code{\link{mvn_mar_20_mc.mvn_pcci_simulation}()},
\code{\link{mvn_mar_20_mc.mvn_pcci_task}()},
\code{\link{mvn_mar_20_mc.mvn_simulation}()},
\code{\link{mvn_mar_20_mc.mvn_task}()},
\code{\link{mvn_mar_20_mc.mvn}()},
\code{\link{mvn_mar_30_mc.mvn_pcci_simulation}()},
\code{\link{mvn_mar_30_mc.mvn_pcci_task}()},
\code{\link{mvn_mar_30_mc.mvn_simulation}()},
\code{\link{mvn_mar_30_mc.mvn_task}()},
\code{\link{mvn_mar_30_mc.mvn}()},
\code{\link{mvn_mar_50_mc.mvn_pcci_simulation}()},
\code{\link{mvn_mar_50_mc.mvn_pcci_task}()},
\code{\link{mvn_mar_50_mc.mvn_simulation}()},
\code{\link{mvn_mar_50_mc.mvn_task}()},
\code{\link{mvn_mar_50_mc.mvn}()},
\code{\link{mvn_mar_mc.mvn_pcci_simulation}()},
\code{\link{mvn_mar_mc.mvn_pcci_task}()},
\code{\link{mvn_mar_mc.mvn_simulation}()},
\code{\link{mvn_mar_mc.mvn_task}()},
\code{\link{mvn_mar_mc.mvn}()},
\code{\link{mvn_mar_std_mc.mvn.sem_pcci_simulation}()},
\code{\link{mvn_mar_std_mc.mvn.sem_pcci_task}()},
\code{\link{mvn_mar_std_mc.mvn.sem_simulation}()},
\code{\link{mvn_mar_std_mc.mvn.sem_task}()},
\code{\link{mvn_mar_std_mc.mvn.sem}()},
\code{\link{mvn_mc.mvn_pcci_simulation}()},
\code{\link{mvn_mc.mvn_pcci_task}()},
\code{\link{mvn_mc.mvn_simulation}()},
\code{\link{mvn_mc.mvn_task}()},
\code{\link{mvn_mc.mvn}()},
\code{\link{mvn_mc.t_pcci_simulation}()},
\code{\link{mvn_mc.t_pcci_task}()},
\code{\link{mvn_mc.t_simulation}()},
\code{\link{mvn_mc.t_task}()},
\code{\link{mvn_mc.t}()},
\code{\link{mvn_mcar_10_mc.mvn_pcci_simulation}()},
\code{\link{mvn_mcar_10_mc.mvn_pcci_task}()},
\code{\link{mvn_mcar_10_mc.mvn_simulation}()},
\code{\link{mvn_mcar_10_mc.mvn_task}()},
\code{\link{mvn_mcar_10_mc.mvn}()},
\code{\link{mvn_mcar_20_mc.mvn_pcci_simulation}()},
\code{\link{mvn_mcar_20_mc.mvn_pcci_task}()},
\code{\link{mvn_mcar_20_mc.mvn_simulation}()},
\code{\link{mvn_mcar_20_mc.mvn_task}()},
\code{\link{mvn_mcar_20_mc.mvn}()},
\code{\link{mvn_mcar_30_mc.mvn_pcci_simulation}()},
\code{\link{mvn_mcar_30_mc.mvn_pcci_task}()},
\code{\link{mvn_mcar_30_mc.mvn_simulation}()},
\code{\link{mvn_mcar_30_mc.mvn_task}()},
\code{\link{mvn_mcar_30_mc.mvn}()},
\code{\link{mvn_mcar_50_mc.mvn_pcci_simulation}()},
\code{\link{mvn_mcar_50_mc.mvn_pcci_task}()},
\code{\link{mvn_mcar_50_mc.mvn_simulation}()},
\code{\link{mvn_mcar_50_mc.mvn_task}()},
\code{\link{mvn_mcar_50_mc.mvn}()},
\code{\link{mvn_mcar_mc.mvn_pcci_simulation}()},
\code{\link{mvn_mcar_mc.mvn_pcci_task}()},
\code{\link{mvn_mcar_mc.mvn_simulation}()},
\code{\link{mvn_mcar_mc.mvn_task}()},
\code{\link{mvn_mcar_mc.mvn}()},
\code{\link{mvn_mcar_std_mc.mvn.sem_pcci_simulation}()},
\code{\link{mvn_mcar_std_mc.mvn.sem_pcci_task}()},
\code{\link{mvn_mcar_std_mc.mvn.sem_simulation}()},
\code{\link{mvn_mcar_std_mc.mvn.sem_task}()},
\code{\link{mvn_mcar_std_mc.mvn.sem}()},
\code{\link{mvn_mnar_10_mc.mvn_pcci_simulation}()},
\code{\link{mvn_mnar_10_mc.mvn_pcci_task}()},
\code{\link{mvn_mnar_10_mc.mvn_simulation}()},
\code{\link{mvn_mnar_10_mc.mvn_task}()},
\code{\link{mvn_mnar_10_mc.mvn}()},
\code{\link{mvn_mnar_20_mc.mvn_pcci_simulation}()},
\code{\link{mvn_mnar_20_mc.mvn_pcci_task}()},
\code{\link{mvn_mnar_20_mc.mvn_simulation}()},
\code{\link{mvn_mnar_20_mc.mvn_task}()},
\code{\link{mvn_mnar_20_mc.mvn}()},
\code{\link{mvn_mnar_30_mc.mvn_pcci_simulation}()},
\code{\link{mvn_mnar_30_mc.mvn_pcci_task}()},
\code{\link{mvn_mnar_30_mc.mvn_simulation}()},
\code{\link{mvn_mnar_30_mc.mvn_task}()},
\code{\link{mvn_mnar_30_mc.mvn}()},
\code{\link{mvn_mnar_50_mc.mvn_pcci_simulation}()},
\code{\link{mvn_mnar_50_mc.mvn_pcci_task}()},
\code{\link{mvn_mnar_50_mc.mvn_simulation}()},
\code{\link{mvn_mnar_50_mc.mvn_task}()},
\code{\link{mvn_mnar_50_mc.mvn}()},
\code{\link{mvn_mnar_mc.mvn_pcci_simulation}()},
\code{\link{mvn_mnar_mc.mvn_pcci_task}()},
\code{\link{mvn_mnar_mc.mvn_simulation}()},
\code{\link{mvn_mnar_mc.mvn_task}()},
\code{\link{mvn_mnar_mc.mvn}()},
\code{\link{mvn_mnar_std_mc.mvn.sem_pcci_simulation}()},
\code{\link{mvn_mnar_std_mc.mvn.sem_pcci_task}()},
\code{\link{mvn_mnar_std_mc.mvn.sem_simulation}()},
\code{\link{mvn_mnar_std_mc.mvn.sem_task}()},
\code{\link{mvn_mnar_std_mc.mvn.sem}()},
\code{\link{mvn_ols_mc.mvn_pcci_simulation}()},
\code{\link{mvn_ols_mc.mvn_pcci_task}()},
\code{\link{mvn_ols_mc.mvn_simulation}()},
\code{\link{mvn_ols_mc.mvn_task}()},
\code{\link{mvn_ols_mc.mvn}()},
\code{\link{mvn_std_mc.mvn.delta_pcci_simulation}()},
\code{\link{mvn_std_mc.mvn.delta_pcci_task}()},
\code{\link{mvn_std_mc.mvn.delta_simulation}()},
\code{\link{mvn_std_mc.mvn.delta_task}()},
\code{\link{mvn_std_mc.mvn.sem_pcci_simulation}()},
\code{\link{mvn_std_mc.mvn.sem_pcci_task}()},
\code{\link{mvn_std_mc.mvn.sem_simulation}()},
\code{\link{mvn_std_mc.mvn.sem_task}()},
\code{\link{mvn_std_mc.mvn.sem}()},
\code{\link{mvn_std_mc.mvn.tb_pcci_simulation}()},
\code{\link{mvn_std_mc.mvn.tb_pcci_task}()},
\code{\link{mvn_std_mc.mvn.tb_simulation}()},
\code{\link{mvn_std_mc.mvn.tb_task}()},
\code{\link{mvn_std_mc.mvn.tb}()},
\code{\link{mvn_std_mc.t.delta_pcci_simulation}()},
\code{\link{mvn_std_mc.t.delta_pcci_task}()},
\code{\link{mvn_std_mc.t.delta_simulation}()},
\code{\link{mvn_std_mc.t.delta_task}()},
\code{\link{mvn_std_mc.t.delta}()},
\code{\link{mvn_std_mc.t.sem_pcci_simulation}()},
\code{\link{mvn_std_mc.t.sem_pcci_task}()},
\code{\link{mvn_std_mc.t.sem_simulation}()},
\code{\link{mvn_std_mc.t.sem_task}()},
\code{\link{mvn_std_mc.t.sem}()},
\code{\link{mvn_std_mc.t.tb_pcci_simulation}()},
\code{\link{mvn_std_mc.t.tb_pcci_task}()},
\code{\link{mvn_std_mc.t.tb_simulation}()},
\code{\link{mvn_std_mc.t.tb_task}()},
\code{\link{mvn_std_mc.t.tb}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{monte carlo method functions}
\keyword{mvn_complete_std}
