% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvn_complete_std_mc.wishart.R
\name{mvn_std_mc.wishart}
\alias{mvn_std_mc.wishart}
\title{Monte Carlo Method Assuming Wishart Distribution using Estimated Covariance Matrix
for Data Generated from a Multivariate Normal Distribution}
\usage{
mvn_std_mc.wishart(data, R = 20000L, taskid)
}
\arguments{
\item{data}{\code{n} by 3 matrix or data frame.
\code{data[, 1]} correspond to values for \code{x}.
\code{data[, 2]} correspond to values for \code{m}.
\code{data[, 3]} correspond to values for \code{y}.}

\item{R}{Integer.
Monte Carlo replications.}

\item{taskid}{Numeric.
Task ID.}
}
\description{
Monte Carlo Method Assuming Wishart Distribution using Estimated Covariance Matrix
for Data Generated from a Multivariate Normal Distribution
}
\examples{
taskid <- 1
data <- mvn_dat(taskid = taskid)
fit.ols(data = data, minimal = TRUE, std = TRUE)

thetahatstar <- mvn_std_mc.wishart(
  data = data, R = 20000L, taskid = taskid
)
hist(thetahatstar)
}
\seealso{
Other monte carlo method functions: 
\code{\link{beta_ols_mc.mvn_pcci_simulation}()},
\code{\link{beta_ols_mc.mvn_pcci_task}()},
\code{\link{beta_ols_mc.mvn_simulation}()},
\code{\link{beta_ols_mc.mvn_task}()},
\code{\link{beta_ols_mc.mvn}()},
\code{\link{exp_ols_mc.mvn_pcci_simulation}()},
\code{\link{exp_ols_mc.mvn_pcci_task}()},
\code{\link{exp_ols_mc.mvn_simulation}()},
\code{\link{exp_ols_mc.mvn_task}()},
\code{\link{exp_ols_mc.mvn}()},
\code{\link{mc.mvn}()},
\code{\link{mc.t}()},
\code{\link{mc.wishart}()},
\code{\link{mvn_mar_10_mc.mvn_pcci_simulation}()},
\code{\link{mvn_mar_10_mc.mvn_pcci_task}()},
\code{\link{mvn_mar_10_mc.mvn_simulation}()},
\code{\link{mvn_mar_10_mc.mvn_task}()},
\code{\link{mvn_mar_10_mc.mvn}()},
\code{\link{mvn_mar_20_mc.mvn_pcci_simulation}()},
\code{\link{mvn_mar_20_mc.mvn_pcci_task}()},
\code{\link{mvn_mar_20_mc.mvn_simulation}()},
\code{\link{mvn_mar_20_mc.mvn_task}()},
\code{\link{mvn_mar_20_mc.mvn}()},
\code{\link{mvn_mar_30_mc.mvn_pcci_simulation}()},
\code{\link{mvn_mar_30_mc.mvn_pcci_task}()},
\code{\link{mvn_mar_30_mc.mvn_simulation}()},
\code{\link{mvn_mar_30_mc.mvn_task}()},
\code{\link{mvn_mar_30_mc.mvn}()},
\code{\link{mvn_mcar_10_mc.mvn_pcci_simulation}()},
\code{\link{mvn_mcar_10_mc.mvn_pcci_task}()},
\code{\link{mvn_mcar_10_mc.mvn_simulation}()},
\code{\link{mvn_mcar_10_mc.mvn_task}()},
\code{\link{mvn_mcar_10_mc.mvn}()},
\code{\link{mvn_mcar_20_mc.mvn_pcci_simulation}()},
\code{\link{mvn_mcar_20_mc.mvn_pcci_task}()},
\code{\link{mvn_mcar_20_mc.mvn_simulation}()},
\code{\link{mvn_mcar_20_mc.mvn_task}()},
\code{\link{mvn_mcar_20_mc.mvn}()},
\code{\link{mvn_mcar_30_mc.mvn_pcci_simulation}()},
\code{\link{mvn_mcar_30_mc.mvn_pcci_task}()},
\code{\link{mvn_mcar_30_mc.mvn_simulation}()},
\code{\link{mvn_mcar_30_mc.mvn_task}()},
\code{\link{mvn_mcar_30_mc.mvn}()},
\code{\link{mvn_mnar_10_mc.mvn_pcci_simulation}()},
\code{\link{mvn_mnar_10_mc.mvn_pcci_task}()},
\code{\link{mvn_mnar_10_mc.mvn_simulation}()},
\code{\link{mvn_mnar_10_mc.mvn_task}()},
\code{\link{mvn_mnar_10_mc.mvn}()},
\code{\link{mvn_mnar_20_mc.mvn_pcci_simulation}()},
\code{\link{mvn_mnar_20_mc.mvn_pcci_task}()},
\code{\link{mvn_mnar_20_mc.mvn_simulation}()},
\code{\link{mvn_mnar_20_mc.mvn_task}()},
\code{\link{mvn_mnar_20_mc.mvn}()},
\code{\link{mvn_mnar_30_mc.mvn_pcci_simulation}()},
\code{\link{mvn_mnar_30_mc.mvn_pcci_task}()},
\code{\link{mvn_mnar_30_mc.mvn_simulation}()},
\code{\link{mvn_mnar_30_mc.mvn_task}()},
\code{\link{mvn_mnar_30_mc.mvn}()},
\code{\link{mvn_ols_mc.mvn_pcci_simulation}()},
\code{\link{mvn_ols_mc.mvn_pcci_task}()},
\code{\link{mvn_ols_mc.mvn_simulation}()},
\code{\link{mvn_ols_mc.mvn_task}()},
\code{\link{mvn_ols_mc.mvn}()},
\code{\link{mvn_sem_mc.mvn_pcci_simulation}()},
\code{\link{mvn_sem_mc.mvn_pcci_task}()},
\code{\link{mvn_sem_mc.mvn_simulation}()},
\code{\link{mvn_sem_mc.mvn_task}()},
\code{\link{mvn_sem_mc.mvn}()},
\code{\link{mvn_std_mc.mvn.delta_pcci_simulation}()},
\code{\link{mvn_std_mc.mvn.delta_pcci_task}()},
\code{\link{mvn_std_mc.mvn.delta_simulation}()},
\code{\link{mvn_std_mc.mvn.delta_task}()},
\code{\link{mvn_std_mc.mvn.delta}()},
\code{\link{mvn_std_mc.mvn.sem_pcci_simulation}()},
\code{\link{mvn_std_mc.mvn.sem_pcci_task}()},
\code{\link{mvn_std_mc.mvn.sem_simulation}()},
\code{\link{mvn_std_mc.mvn.sem_task}()},
\code{\link{mvn_std_mc.mvn.sem}()},
\code{\link{mvn_std_mc.mvn.tb_pcci_simulation}()},
\code{\link{mvn_std_mc.mvn.tb_pcci_task}()},
\code{\link{mvn_std_mc.mvn.tb_simulation}()},
\code{\link{mvn_std_mc.mvn.tb_task}()},
\code{\link{mvn_std_mc.mvn.tb}()},
\code{\link{mvn_std_mc.wishart_pcci_simulation}()},
\code{\link{mvn_std_mc.wishart_pcci_task}()},
\code{\link{mvn_std_mc.wishart_simulation}()},
\code{\link{mvn_std_mc.wishart_task}()},
\code{\link{vm_mod_ols_mc.mvn_pcci_simulation}()},
\code{\link{vm_mod_ols_mc.mvn_pcci_task}()},
\code{\link{vm_mod_ols_mc.mvn_simulation}()},
\code{\link{vm_mod_ols_mc.mvn_task}()},
\code{\link{vm_mod_ols_mc.mvn}()},
\code{\link{vm_mod_sem_mc.mvn_pcci_simulation}()},
\code{\link{vm_mod_sem_mc.mvn_pcci_task}()},
\code{\link{vm_mod_sem_mc.mvn_simulation}()},
\code{\link{vm_mod_sem_mc.mvn_task}()},
\code{\link{vm_mod_sem_mc.mvn}()},
\code{\link{vm_sev_ols_mc.mvn_pcci_simulation}()},
\code{\link{vm_sev_ols_mc.mvn_pcci_task}()},
\code{\link{vm_sev_ols_mc.mvn_simulation}()},
\code{\link{vm_sev_ols_mc.mvn_task}()},
\code{\link{vm_sev_ols_mc.mvn}()},
\code{\link{vm_sev_sem_mc.mvn_pcci_simulation}()},
\code{\link{vm_sev_sem_mc.mvn_pcci_task}()},
\code{\link{vm_sev_sem_mc.mvn_simulation}()},
\code{\link{vm_sev_sem_mc.mvn_task}()},
\code{\link{vm_sev_sem_mc.mvn}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{monte carlo method functions}
\keyword{mc}
