% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcstdsem.R
\name{mvn_std_mcmvnsem_pcci_task}
\alias{mvn_std_mcmvnsem_pcci_task}
\title{Monte Carlo Method Confidence Intervals for Indirect Effect in a Simple Mediation Model Assuming \eqn{t} Distribution
for Data Generated from a Multivariate Normal Distribution (Single Task)}
\usage{
mvn_std_mcmvnsem_pcci_task(taskid, dir = getwd(), alpha = c(0.001, 0.01, 0.05))
}
\arguments{
\item{taskid}{Numeric.
Task ID.}

\item{alpha}{Numeric vector.
Alpha level.
By default \code{alpha = c(0.001, 0.01, 0.05)}.}
}
\description{
Monte Carlo Method Confidence Intervals for Indirect Effect in a Simple Mediation Model Assuming \eqn{t} Distribution
for Data Generated from a Multivariate Normal Distribution (Single Task)
}
\seealso{
Other monte carlo method functions: 
\code{\link{.mcmvn}()},
\code{\link{.mct}()},
\code{\link{mc}()},
\code{\link{mvn_mcmvn_mar_pcci_simulation}()},
\code{\link{mvn_mcmvn_mar_pcci_task}()},
\code{\link{mvn_mcmvn_mar_simulation}()},
\code{\link{mvn_mcmvn_mar_task}()},
\code{\link{mvn_mcmvn_mar}()},
\code{\link{mvn_mcmvn_mcar_pcci_simulation}()},
\code{\link{mvn_mcmvn_mcar_pcci_task}()},
\code{\link{mvn_mcmvn_mcar_simulation}()},
\code{\link{mvn_mcmvn_mcar_task}()},
\code{\link{mvn_mcmvn_mcar}()},
\code{\link{mvn_mcmvn_pcci_simulation}()},
\code{\link{mvn_mcmvn_pcci_task}()},
\code{\link{mvn_mcmvn_simulation}()},
\code{\link{mvn_mcmvn_task}()},
\code{\link{mvn_mcmvn}()},
\code{\link{mvn_mct_mar_pcci_simulation}()},
\code{\link{mvn_mct_mar_pcci_task}()},
\code{\link{mvn_mct_mar_simulation}()},
\code{\link{mvn_mct_mar_task}()},
\code{\link{mvn_mct_mar}()},
\code{\link{mvn_mct_mcar_pcci_simulation}()},
\code{\link{mvn_mct_mcar_pcci_task}()},
\code{\link{mvn_mct_mcar_simulation}()},
\code{\link{mvn_mct_mcar_task}()},
\code{\link{mvn_mct_mcar}()},
\code{\link{mvn_mct_pcci_simulation}()},
\code{\link{mvn_mct_pcci_task}()},
\code{\link{mvn_mct_simulation}()},
\code{\link{mvn_mct_task}()},
\code{\link{mvn_mct}()},
\code{\link{mvn_std_mcmvndelta_pcci_simulation}()},
\code{\link{mvn_std_mcmvndelta_pcci_task}()},
\code{\link{mvn_std_mcmvndelta_simulation}()},
\code{\link{mvn_std_mcmvndelta_task}()},
\code{\link{mvn_std_mcmvndelta}()},
\code{\link{mvn_std_mcmvnsem_pcci_simulation}()},
\code{\link{mvn_std_mcmvnsem_simulation}()},
\code{\link{mvn_std_mcmvnsem_task}()},
\code{\link{mvn_std_mcmvnsem}()},
\code{\link{mvn_std_mcmvntb_pcci_simulation}()},
\code{\link{mvn_std_mcmvntb_pcci_task}()},
\code{\link{mvn_std_mcmvntb_simulation}()},
\code{\link{mvn_std_mcmvntb_task}()},
\code{\link{mvn_std_mcmvntb}()},
\code{\link{mvn_std_mctdelta_pcci_simulation}()},
\code{\link{mvn_std_mctdelta_pcci_task}()},
\code{\link{mvn_std_mctdelta_simulation}()},
\code{\link{mvn_std_mctdelta_task}()},
\code{\link{mvn_std_mctdelta}()},
\code{\link{mvn_std_mctsem_pcci_simulation}()},
\code{\link{mvn_std_mctsem_pcci_task}()},
\code{\link{mvn_std_mctsem_simulation}()},
\code{\link{mvn_std_mctsem_task}()},
\code{\link{mvn_std_mctsem}()},
\code{\link{mvn_std_mcttb_pcci_simulation}()},
\code{\link{mvn_std_mcttb_pcci_task}()},
\code{\link{mvn_std_mcttb_simulation}()},
\code{\link{mvn_std_mcttb_task}()},
\code{\link{mvn_std_mcttb}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{monte carlo method functions}
\keyword{mc}
