% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvn_complete_std_nb.R
\name{mvn_std_nb_pcci_simulation}
\alias{mvn_std_nb_pcci_simulation}
\title{Nonparametric Bootstrap Percentile Confidence Intervals for Indirect Effect in a Standardized Simple Mediation Model
for Data Generated from a Multivariate Normal Distribution
(Simulation)}
\usage{
mvn_std_nb_pcci_simulation(
  dir = getwd(),
  all = TRUE,
  taskid = NULL,
  par = TRUE,
  ncores = NULL,
  blas_threads = TRUE,
  mc = TRUE,
  lb = FALSE,
  cl_eval = FALSE,
  cl_export = FALSE,
  cl_expr,
  cl_vars
)
}
\arguments{
\item{dir}{Character string.
Directory where results of \verb{mvn_dat_*} are stored.}

\item{all}{Logical.
Process all results.}

\item{taskid}{Numeric.
Task ID.}

\item{par}{Logical.
If \code{TRUE}, use multiple cores.
If \code{FALSE}, use \code{\link[=lapply]{lapply()}}.}

\item{ncores}{Integer.
Number of cores to use if \code{par = TRUE}.
If unspecified, defaults to \code{detectCores() - 1}.}

\item{blas_threads}{Logical.
If \code{TRUE}, set BLAS threads using \code{blas_set_num_threads(threads = blas_get_num_procs())}.
If \code{FALSE}, set BLAS threads using \code{blas_set_num_threads(threads = 1)}.
If \code{par = TRUE}, \code{blas_threads} is automatically set to \code{FALSE}
to prevent conflicts in parallel processing.
This argument is useful when \code{FUN} can handle implicit parallelism
when \code{par = FALSE}, for example linear algebra operations.}

\item{mc}{Logical.
If \code{TRUE}, use \code{\link[parallel:mclapply]{parallel::mclapply()}}.
If \code{FALSE}, use \code{\link[parallel:clusterApply]{parallel::parLapply()}} or \code{\link[parallel:clusterApply]{parallel::parLapplyLB()}}.
Ignored if \code{par = FALSE}.}

\item{lb}{Logical.
If \code{TRUE} use \code{\link[parallel:clusterApply]{parallel::parLapplyLB()}}.
If \code{FALSE}, use \code{\link[parallel:clusterApply]{parallel::parLapply()}}.
Ignored if \code{par = FALSE} and \code{mc = TRUE}.}

\item{cl_eval}{Logical.
Execute \code{\link[parallel:clusterApply]{parallel::clusterEvalQ()}} using \code{cl_expr}.
Ignored if \code{mc = TRUE}.}

\item{cl_export}{Logical.
Execute \code{\link[parallel:clusterApply]{parallel::clusterExport()}} using \code{cl_vars}.
Ignored if \code{mc = TRUE}.}

\item{cl_expr}{Expression.
Expression passed to \code{\link[parallel:clusterApply]{parallel::clusterEvalQ()}}
Ignored if \code{mc = TRUE}.}

\item{cl_vars}{Character vector.
Names of objects to pass to \code{\link[parallel:clusterApply]{parallel::clusterExport()}}
Ignored if \code{mc = TRUE}.}
}
\description{
Nonparametric Bootstrap Percentile Confidence Intervals for Indirect Effect in a Standardized Simple Mediation Model
for Data Generated from a Multivariate Normal Distribution
(Simulation)
}
\seealso{
Other nonparametric functions: 
\code{\link{mvn_mar_nb_bcaci_simulation}()},
\code{\link{mvn_mar_nb_bcaci_task}()},
\code{\link{mvn_mar_nb_bcci_simulation}()},
\code{\link{mvn_mar_nb_bcci_task}()},
\code{\link{mvn_mar_nb_pcci_simulation}()},
\code{\link{mvn_mar_nb_pcci_task}()},
\code{\link{mvn_mar_nb_simulation}()},
\code{\link{mvn_mar_nb_task}()},
\code{\link{mvn_mar_nb}()},
\code{\link{mvn_mcar_nb_bcaci_simulation}()},
\code{\link{mvn_mcar_nb_bcaci_task}()},
\code{\link{mvn_mcar_nb_bcci_simulation}()},
\code{\link{mvn_mcar_nb_bcci_task}()},
\code{\link{mvn_mcar_nb_fiml_bcaci_simulation}()},
\code{\link{mvn_mcar_nb_fiml_bcaci_task}()},
\code{\link{mvn_mcar_nb_fiml_bcci_simulation}()},
\code{\link{mvn_mcar_nb_fiml_bcci_task}()},
\code{\link{mvn_mcar_nb_fiml_pcci_simulation}()},
\code{\link{mvn_mcar_nb_fiml_pcci_task}()},
\code{\link{mvn_mcar_nb_fiml_simulation}()},
\code{\link{mvn_mcar_nb_fiml_task}()},
\code{\link{mvn_mcar_nb_fiml}()},
\code{\link{mvn_mcar_nb_pcci_simulation}()},
\code{\link{mvn_mcar_nb_pcci_task}()},
\code{\link{mvn_mcar_nb_simulation}()},
\code{\link{mvn_mcar_nb_task}()},
\code{\link{mvn_mcar_nb}()},
\code{\link{mvn_mnar_nb_bcaci_simulation}()},
\code{\link{mvn_mnar_nb_bcaci_task}()},
\code{\link{mvn_mnar_nb_bcci_simulation}()},
\code{\link{mvn_mnar_nb_bcci_task}()},
\code{\link{mvn_mnar_nb_pcci_simulation}()},
\code{\link{mvn_mnar_nb_pcci_task}()},
\code{\link{mvn_mnar_nb_simulation}()},
\code{\link{mvn_mnar_nb_task}()},
\code{\link{mvn_mnar_nb}()},
\code{\link{mvn_nb_bcaci_simulation}()},
\code{\link{mvn_nb_bcaci_task}()},
\code{\link{mvn_nb_bcci_simulation}()},
\code{\link{mvn_nb_bcci_task}()},
\code{\link{mvn_nb_pcci_simulation}()},
\code{\link{mvn_nb_pcci_task}()},
\code{\link{mvn_nb_simulation}()},
\code{\link{mvn_nb_task}()},
\code{\link{mvn_nb}()},
\code{\link{mvn_std_nb_bcaci_simulation}()},
\code{\link{mvn_std_nb_bcaci_task}()},
\code{\link{mvn_std_nb_bcci_simulation}()},
\code{\link{mvn_std_nb_bcci_task}()},
\code{\link{mvn_std_nb_pcci_task}()},
\code{\link{mvn_std_nb_simulation}()},
\code{\link{mvn_std_nb_task}()},
\code{\link{mvn_std_nb}()},
\code{\link{nb}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{nonparametric functions}
\keyword{mvn_complete_std}
