% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbstd.R
\name{mvn_std_nb_task}
\alias{mvn_std_nb_task}
\title{Nonparametric Bootstrapping for Data Generated from a Multivariate Normal Distribution (Single Task)}
\usage{
mvn_std_nb_task(taskid, dir = getwd(), B = 5000, overwrite = FALSE)
}
\arguments{
\item{taskid}{Numeric.
Task ID.}

\item{dir}{Character string.
Directory where results of \verb{mvn_dat_*} are stored.}

\item{B}{Integer.
Number of bootstrap samples.}

\item{overwrite}{Logical.
Overwrite existing results.}
}
\description{
Nonparametric Bootstrapping for Data Generated from a Multivariate Normal Distribution (Single Task)
}
\seealso{
Other nonparametric functions: 
\code{\link{.nb}()},
\code{\link{mvn_nb_bcaci_simulation}()},
\code{\link{mvn_nb_bcaci_task}()},
\code{\link{mvn_nb_bcci_simulation}()},
\code{\link{mvn_nb_bcci_task}()},
\code{\link{mvn_nb_pcci_simulation}()},
\code{\link{mvn_nb_pcci_task}()},
\code{\link{mvn_nb_simulation}()},
\code{\link{mvn_nb_task}()},
\code{\link{mvn_nb}()},
\code{\link{mvn_pbmvn_bcaci_simulation}()},
\code{\link{mvn_pbmvn_bcaci_task}()},
\code{\link{mvn_pbmvn_bcci_simulation}()},
\code{\link{mvn_pbmvn_bcci_task}()},
\code{\link{mvn_std_nb_bcaci_simulation}()},
\code{\link{mvn_std_nb_bcaci_task}()},
\code{\link{mvn_std_nb_bcci_simulation}()},
\code{\link{mvn_std_nb_bcci_task}()},
\code{\link{mvn_std_nb_pcci_simulation}()},
\code{\link{mvn_std_nb_pcci_task}()},
\code{\link{mvn_std_nb_simulation}()},
\code{\link{mvn_std_nb}()},
\code{\link{mvn_std_pbmvn_bcaci_simulation}()},
\code{\link{mvn_std_pbmvn_bcaci_task}()},
\code{\link{mvn_std_pbmvn_bcci_simulation}()},
\code{\link{mvn_std_pbmvn_bcci_task}()},
\code{\link{nb}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{nonparametric functions}
\keyword{nb}
