% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pb.beta.R
\name{pb.beta}
\alias{pb.beta}
\title{Parametric Bootstrapping Assuming Beta X}
\usage{
pb.beta(
  data,
  std = FALSE,
  B = 5000,
  par = TRUE,
  ncores = NULL,
  blas_threads = TRUE,
  mc = TRUE,
  lb = FALSE
)
}
\arguments{
\item{data}{\code{n} by 3 matrix or data frame.
\code{data[, 1]} correspond to values for \code{x}.
\code{data[, 2]} correspond to values for \code{m}.
\code{data[, 3]} correspond to values for \code{y}.}

\item{std}{Logical.
Standardize the indirect effect
\eqn{
    \hat{\alpha}^{\prime} \hat{\beta}^{\prime}
    = \hat{\alpha} \hat{\beta} \frac{\hat{\sigma}_x}{\hat{\sigma}_y}}.}

\item{B}{Integer.
Number of bootstrap samples.}

\item{par}{Logical.
If \code{TRUE}, use multiple cores.
If \code{FALSE}, use \code{\link[=lapply]{lapply()}}.}

\item{ncores}{Integer.
Number of cores to use if \code{par = TRUE}.
If unspecified, defaults to \code{detectCores() - 1}.}

\item{blas_threads}{Logical.
If \code{TRUE}, set BLAS threads using \code{blas_set_num_threads(threads = blas_get_num_procs())}.
If \code{FALSE}, set BLAS threads using \code{blas_set_num_threads(threads = 1)}.
If \code{par = TRUE}, \code{blas_threads} is automatically set to \code{FALSE}
to prevent conflicts in parallel processing.
This argument is useful when \code{FUN} can handle implicit parallelism
when \code{par = FALSE}, for example linear algebra operations.}

\item{mc}{Logical.
If \code{TRUE}, use \code{\link[parallel:mclapply]{parallel::mclapply()}}.
If \code{FALSE}, use \code{\link[parallel:clusterApply]{parallel::parLapply()}} or \code{\link[parallel:clusterApply]{parallel::parLapplyLB()}}.
Ignored if \code{par = FALSE}.}

\item{lb}{Logical.
If \code{TRUE} use \code{\link[parallel:clusterApply]{parallel::parLapplyLB()}}.
If \code{FALSE}, use \code{\link[parallel:clusterApply]{parallel::parLapply()}}.
Ignored if \code{par = FALSE} and \code{mc = TRUE}.}
}
\description{
Parametric Bootstrapping Assuming Beta X
}
\examples{
data <- genbeta(taskid = 1)

# Unstandardized -------------------------------------------------------------
thetahatstar <- pb.beta(
  data = data, B = 5000, par = FALSE
)
hist(thetahatstar)

# Standardized ---------------------------------------------------------------
thetahatstar <- pb.beta(
  data = data, std = TRUE, B = 5000, par = FALSE
)
hist(thetahatstar)
}
\seealso{
Other parametric functions: 
\code{\link{beta_pb.beta_bcaci_simulation}()},
\code{\link{beta_pb.beta_bcaci_task}()},
\code{\link{beta_pb.beta_bcci_simulation}()},
\code{\link{beta_pb.beta_bcci_task}()},
\code{\link{beta_pb.beta_pcci_simulation}()},
\code{\link{beta_pb.beta_pcci_task}()},
\code{\link{beta_pb.beta_simulation}()},
\code{\link{beta_pb.beta_task}()},
\code{\link{beta_pb.beta}()},
\code{\link{beta_pb.mvn_bcaci_simulation}()},
\code{\link{beta_pb.mvn_bcaci_task}()},
\code{\link{beta_pb.mvn_bcci_simulation}()},
\code{\link{beta_pb.mvn_bcci_task}()},
\code{\link{beta_pb.mvn_pcci_simulation}()},
\code{\link{beta_pb.mvn_pcci_task}()},
\code{\link{beta_pb.mvn_simulation}()},
\code{\link{beta_pb.mvn_task}()},
\code{\link{beta_pb.mvn}()},
\code{\link{exp_pb.exp_bcaci_simulation}()},
\code{\link{exp_pb.exp_bcaci_task}()},
\code{\link{exp_pb.exp_bcci_simulation}()},
\code{\link{exp_pb.exp_bcci_task}()},
\code{\link{exp_pb.exp_pcci_simulation}()},
\code{\link{exp_pb.exp_pcci_task}()},
\code{\link{exp_pb.exp_simulation}()},
\code{\link{exp_pb.exp_task}()},
\code{\link{exp_pb.exp}()},
\code{\link{exp_pb.mvn_bcaci_simulation}()},
\code{\link{exp_pb.mvn_bcaci_task}()},
\code{\link{exp_pb.mvn_bcci_simulation}()},
\code{\link{exp_pb.mvn_bcci_task}()},
\code{\link{exp_pb.mvn_pcci_simulation}()},
\code{\link{exp_pb.mvn_pcci_task}()},
\code{\link{exp_pb.mvn_simulation}()},
\code{\link{exp_pb.mvn_task}()},
\code{\link{exp_pb.mvn}()},
\code{\link{mvn_mar_10_pb.mvn_bcci_simulation}()},
\code{\link{mvn_mar_10_pb.mvn_bcci_task}()},
\code{\link{mvn_mar_10_pb.mvn_pcci_simulation}()},
\code{\link{mvn_mar_10_pb.mvn_pcci_task}()},
\code{\link{mvn_mar_10_pb.mvn_simulation}()},
\code{\link{mvn_mar_10_pb.mvn_task}()},
\code{\link{mvn_mar_10_pb.mvn}()},
\code{\link{mvn_mar_20_pb.mvn_bcci_simulation}()},
\code{\link{mvn_mar_20_pb.mvn_bcci_task}()},
\code{\link{mvn_mar_20_pb.mvn_pcci_simulation}()},
\code{\link{mvn_mar_20_pb.mvn_pcci_task}()},
\code{\link{mvn_mar_20_pb.mvn_simulation}()},
\code{\link{mvn_mar_20_pb.mvn_task}()},
\code{\link{mvn_mar_20_pb.mvn}()},
\code{\link{mvn_mar_30_pb.mvn_bcci_simulation}()},
\code{\link{mvn_mar_30_pb.mvn_bcci_task}()},
\code{\link{mvn_mar_30_pb.mvn_pcci_simulation}()},
\code{\link{mvn_mar_30_pb.mvn_pcci_task}()},
\code{\link{mvn_mar_30_pb.mvn_simulation}()},
\code{\link{mvn_mar_30_pb.mvn_task}()},
\code{\link{mvn_mar_30_pb.mvn}()},
\code{\link{mvn_mcar_10_pb.mvn_bcci_simulation}()},
\code{\link{mvn_mcar_10_pb.mvn_bcci_task}()},
\code{\link{mvn_mcar_10_pb.mvn_pcci_simulation}()},
\code{\link{mvn_mcar_10_pb.mvn_pcci_task}()},
\code{\link{mvn_mcar_10_pb.mvn_simulation}()},
\code{\link{mvn_mcar_10_pb.mvn_task}()},
\code{\link{mvn_mcar_10_pb.mvn}()},
\code{\link{mvn_mcar_20_pb.mvn_bcci_simulation}()},
\code{\link{mvn_mcar_20_pb.mvn_bcci_task}()},
\code{\link{mvn_mcar_20_pb.mvn_pcci_simulation}()},
\code{\link{mvn_mcar_20_pb.mvn_pcci_task}()},
\code{\link{mvn_mcar_20_pb.mvn_simulation}()},
\code{\link{mvn_mcar_20_pb.mvn_task}()},
\code{\link{mvn_mcar_20_pb.mvn}()},
\code{\link{mvn_mcar_30_pb.mvn_bcci_simulation}()},
\code{\link{mvn_mcar_30_pb.mvn_bcci_task}()},
\code{\link{mvn_mcar_30_pb.mvn_pcci_simulation}()},
\code{\link{mvn_mcar_30_pb.mvn_pcci_task}()},
\code{\link{mvn_mcar_30_pb.mvn_simulation}()},
\code{\link{mvn_mcar_30_pb.mvn_task}()},
\code{\link{mvn_mcar_30_pb.mvn}()},
\code{\link{mvn_mnar_10_pb.mvn_bcci_simulation}()},
\code{\link{mvn_mnar_10_pb.mvn_bcci_task}()},
\code{\link{mvn_mnar_10_pb.mvn_pcci_simulation}()},
\code{\link{mvn_mnar_10_pb.mvn_pcci_task}()},
\code{\link{mvn_mnar_10_pb.mvn_simulation}()},
\code{\link{mvn_mnar_10_pb.mvn_task}()},
\code{\link{mvn_mnar_10_pb.mvn}()},
\code{\link{mvn_mnar_20_pb.mvn_bcci_simulation}()},
\code{\link{mvn_mnar_20_pb.mvn_bcci_task}()},
\code{\link{mvn_mnar_20_pb.mvn_pcci_simulation}()},
\code{\link{mvn_mnar_20_pb.mvn_pcci_task}()},
\code{\link{mvn_mnar_20_pb.mvn_simulation}()},
\code{\link{mvn_mnar_20_pb.mvn_task}()},
\code{\link{mvn_mnar_20_pb.mvn}()},
\code{\link{mvn_mnar_30_pb.mvn_bcci_simulation}()},
\code{\link{mvn_mnar_30_pb.mvn_bcci_task}()},
\code{\link{mvn_mnar_30_pb.mvn_pcci_simulation}()},
\code{\link{mvn_mnar_30_pb.mvn_pcci_task}()},
\code{\link{mvn_mnar_30_pb.mvn_simulation}()},
\code{\link{mvn_mnar_30_pb.mvn_task}()},
\code{\link{mvn_mnar_30_pb.mvn}()},
\code{\link{mvn_pb.mvn_bcaci_simulation}()},
\code{\link{mvn_pb.mvn_bcaci_task}()},
\code{\link{mvn_pb.mvn_bcci_simulation}()},
\code{\link{mvn_pb.mvn_bcci_task}()},
\code{\link{mvn_pb.mvn_pcci_simulation}()},
\code{\link{mvn_pb.mvn_pcci_task}()},
\code{\link{mvn_pb.mvn_simulation}()},
\code{\link{mvn_pb.mvn_task}()},
\code{\link{mvn_pb.mvn}()},
\code{\link{mvn_std_pb.mvn_bcaci_simulation}()},
\code{\link{mvn_std_pb.mvn_bcaci_task}()},
\code{\link{mvn_std_pb.mvn_bcci_simulation}()},
\code{\link{mvn_std_pb.mvn_bcci_task}()},
\code{\link{mvn_std_pb.mvn_pcci_simulation}()},
\code{\link{mvn_std_pb.mvn_pcci_task}()},
\code{\link{mvn_std_pb.mvn_simulation}()},
\code{\link{mvn_std_pb.mvn_task}()},
\code{\link{mvn_std_pb.mvn}()},
\code{\link{pb.exp}()},
\code{\link{pb.mvn}()},
\code{\link{pb.vm}()},
\code{\link{vm_mod_pb.mvn_bcaci_simulation}()},
\code{\link{vm_mod_pb.mvn_bcaci_task}()},
\code{\link{vm_mod_pb.mvn_bcci_simulation}()},
\code{\link{vm_mod_pb.mvn_bcci_task}()},
\code{\link{vm_mod_pb.mvn_pcci_simulation}()},
\code{\link{vm_mod_pb.mvn_pcci_task}()},
\code{\link{vm_mod_pb.mvn_simulation}()},
\code{\link{vm_mod_pb.mvn_task}()},
\code{\link{vm_mod_pb.mvn}()},
\code{\link{vm_mod_pb.vm_bcaci_simulation}()},
\code{\link{vm_mod_pb.vm_bcaci_task}()},
\code{\link{vm_mod_pb.vm_bcci_simulation}()},
\code{\link{vm_mod_pb.vm_bcci_task}()},
\code{\link{vm_mod_pb.vm_pcci_simulation}()},
\code{\link{vm_mod_pb.vm_pcci_task}()},
\code{\link{vm_mod_pb.vm_simulation}()},
\code{\link{vm_mod_pb.vm_task}()},
\code{\link{vm_mod_pb.vm}()},
\code{\link{vm_mod_std_pb.mvn_bcaci_simulation}()},
\code{\link{vm_mod_std_pb.mvn_bcaci_task}()},
\code{\link{vm_mod_std_pb.mvn_bcci_simulation}()},
\code{\link{vm_mod_std_pb.mvn_bcci_task}()},
\code{\link{vm_mod_std_pb.mvn_pcci_simulation}()},
\code{\link{vm_mod_std_pb.mvn_pcci_task}()},
\code{\link{vm_mod_std_pb.mvn_simulation}()},
\code{\link{vm_mod_std_pb.mvn_task}()},
\code{\link{vm_mod_std_pb.mvn}()},
\code{\link{vm_sev_pb.mvn_bcaci_simulation}()},
\code{\link{vm_sev_pb.mvn_bcaci_task}()},
\code{\link{vm_sev_pb.mvn_bcci_simulation}()},
\code{\link{vm_sev_pb.mvn_bcci_task}()},
\code{\link{vm_sev_pb.mvn_pcci_simulation}()},
\code{\link{vm_sev_pb.mvn_pcci_task}()},
\code{\link{vm_sev_pb.mvn_simulation}()},
\code{\link{vm_sev_pb.mvn_task}()},
\code{\link{vm_sev_pb.mvn}()},
\code{\link{vm_sev_pb.vm_bcaci_simulation}()},
\code{\link{vm_sev_pb.vm_bcaci_task}()},
\code{\link{vm_sev_pb.vm_bcci_simulation}()},
\code{\link{vm_sev_pb.vm_bcci_task}()},
\code{\link{vm_sev_pb.vm_pcci_simulation}()},
\code{\link{vm_sev_pb.vm_pcci_task}()},
\code{\link{vm_sev_pb.vm_simulation}()},
\code{\link{vm_sev_pb.vm_task}()},
\code{\link{vm_sev_pb.vm}()},
\code{\link{vm_sev_std_pb.mvn_bcaci_simulation}()},
\code{\link{vm_sev_std_pb.mvn_bcaci_task}()},
\code{\link{vm_sev_std_pb.mvn_bcci_simulation}()},
\code{\link{vm_sev_std_pb.mvn_bcci_task}()},
\code{\link{vm_sev_std_pb.mvn_pcci_simulation}()},
\code{\link{vm_sev_std_pb.mvn_pcci_task}()},
\code{\link{vm_sev_std_pb.mvn_simulation}()},
\code{\link{vm_sev_std_pb.mvn_task}()},
\code{\link{vm_sev_std_pb.mvn}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{parametric functions}
\keyword{pb}
