% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci.R
\name{pcci}
\alias{pcci}
\title{Percentile Confidence Intervals}
\usage{
pcci(thetahatstar, thetahat, theta = NULL, alpha = c(0.001, 0.01, 0.05))
}
\arguments{
\item{thetahatstar}{Numeric vector.
Sampling distribution of thetahat.}

\item{thetahat}{Numeric.
Parameter estimate.}

\item{theta}{Numeric.
Parameter.
Optional argument.}

\item{alpha}{Numeric vector.
Alpha level.
By default \code{alpha = c(0.001, 0.01, 0.05)}.}
}
\description{
Percentile Confidence Intervals
}
\examples{
B <- 5000
data <- jeksterslabRdatarepo::thirst
thetahat <- .fit(data, minimal = TRUE)
thetahatstar <- .pbmvn(data = data, B = 5000, par = FALSE)
pcci(
  thetahatstar = thetahatstar,
  thetahat = thetahat,
  theta = 0.15 # assuming that the true indirect effect is 0.15
)
}
\seealso{
Other confidence intervals functions: 
\code{\link{bcaci}()},
\code{\link{bcci}()},
\code{\link{evalci}()},
\code{\link{jackknife}()},
\code{\link{len}()},
\code{\link{shape}()},
\code{\link{theta_hit}()},
\code{\link{zero_hit}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{confidence intervals functions}
\keyword{ci}
