% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_kurt}
\alias{plot_kurt}
\title{Plot Kurtosis}
\usage{
plot_kurt(data, std = FALSE)
}
\arguments{
\item{data}{Dataframe of results.}

\item{std}{Logical.
If \code{TRUE}, standardized indirect effect.
If \code{FALSE}, unstandardized indirect effect.}
}
\description{
Plot Kurtosis
}
\seealso{
Other plotting functions: 
\code{\link{plot_bias}()},
\code{\link{plot_miss}()},
\code{\link{plot_power}()},
\code{\link{plot_rmse}()},
\code{\link{plot_type1}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{plotting functions}
\keyword{plot}
