% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_power}
\alias{plot_power}
\title{Plot Power}
\usage{
plot_power(data, alpha_level = 0.05)
}
\arguments{
\item{data}{Dataframe of results.}

\item{alpha_level}{Numeric.
Available values are \code{0.05}, \code{0.01}, \code{0.001}.}
}
\description{
Plot Power
}
\seealso{
Other plotting functions: 
\code{\link{plot_bias}()},
\code{\link{plot_kurt}()},
\code{\link{plot_miss}()},
\code{\link{plot_rmse}()},
\code{\link{plot_type1}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{plotting functions}
\keyword{plot}
