% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_rmse}
\alias{plot_rmse}
\title{Plot Root Mean Square Error}
\usage{
plot_rmse(data, std = FALSE)
}
\arguments{
\item{data}{Dataframe of results.}

\item{std}{Logical.
If \code{TRUE}, standardized indirect effect.
If \code{FALSE}, unstandardized indirect effect.}
}
\description{
Plot Root Mean Square Error
}
\seealso{
Other plotting functions: 
\code{\link{plot_bias}()},
\code{\link{plot_kurt}()},
\code{\link{plot_miss}()},
\code{\link{plot_power}()},
\code{\link{plot_type1}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{plotting functions}
\keyword{plot}
