% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_type1}
\alias{plot_type1}
\title{Plot Type I Error Rates}
\usage{
plot_type1(data, alpha_level = 0.05, criteria = "liberal")
}
\arguments{
\item{data}{Dataframe of results.}

\item{alpha_level}{Numeric.
Available values are \code{0.05}, \code{0.01}, \code{0.001}.}

\item{criteria}{Character string.
Available values are \code{"liberal"}, \code{"moderate"}, \code{"strict"}.}
}
\description{
Plot Type I Error Rates
}
\seealso{
Other plotting functions: 
\code{\link{plot_bias}()},
\code{\link{plot_kurt}()},
\code{\link{plot_miss}()},
\code{\link{plot_power}()},
\code{\link{plot_rmse}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{plotting functions}
\keyword{plot}
