% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results_mvn_fit.R
\docType{data}
\name{results_mvn_fit}
\alias{results_mvn_fit}
\title{Monte Carlo Simulation Results - Model Fitting (Multivariate Normal)}
\format{
A data frame with the following variables
\describe{
\item{taskid}{
Simulation task identification number.
}
\item{n}{
Sample size.
}
\item{reps}{
Monte Carlo replications.
}
\item{taudot}{
Population slope of path from \code{x} to \code{y} \eqn{\left( \dot{\tau} \right)}
}
\item{beta}{
Population slope of path from \code{m} to \code{y} \eqn{\left( \beta \right)}
}
\item{alpha}{
Population slope of path from \code{x} to \code{m} \eqn{\left( \alpha \right)}
}
\item{alphabeta}{
Population indirect effect of \code{x} on \code{y} through \code{m} \eqn{\left( \alpha \beta \right)}
}
\item{sigma2x}{
Population variance of \code{x} \eqn{\left( \sigma_{x}^{2} \right)}
}
\item{sigma2epsilonm}{
Population error variance of \code{m} \eqn{\left( \sigma_{\varepsilon_{m}}^{2} \right)}
}
\item{sigma2epsilony}{
Population error variance of \code{y} \eqn{\left( \sigma_{\varepsilon_{y}}^{2} \right)}
}
\item{mux}{
Population mean of \code{x} \eqn{\left( \mu_x \right)}.
}
\item{deltam}{
Population intercept of \code{m} \eqn{\left( \delta_m \right)}.
}
\item{deltay}{
Population intercept of \code{y} \eqn{\left( \delta_y \right)}.
}
\item{deltayhat}{
Mean of estimated intercept of \code{y} \eqn{\left( \hat{\delta}_y \right)}.
}
\item{taudothat}{
Mean of estimated slope of path from \code{x} to \code{y} \eqn{\left( \hat{\dot{\tau}} \right)}
}
\item{betahat}{
Mean of estimated slope of path from \code{m} to \code{y} \eqn{\left( \hat{\beta} \right)}
}
\item{deltamhat}{
Mean of estimated intercept of \code{m} \eqn{\left( \hat{\delta_m} \right)}.
}
\item{alphahat}{
Mean of estimated slope of path from \code{x} to \code{m} \eqn{\left( \hat{\alpha} \right)}
}
\item{alphahatbetahat}{
Mean of estimated indirect effect of \code{x} on \code{y} through \code{m} \eqn{\left( \hat{\alpha} \hat{\beta} \right)}
}
\item{alphahatprimebetahatprime}{
Mean of estimated standardized indirect effect of \code{x} on \code{y} through \code{m} \eqn{\left( \hat{\alpha}^{\prime} \hat{\beta}^{\prime} \right)}
}
\item{sigma2xhat}{
Mean of estimated variance of \code{x} \eqn{\left( \hat{\sigma}_{x}^{2} \right)}
}
\item{sigma2epsilonmhat}{
Mean of estimated error variance of \code{m} \eqn{\left( \hat{\sigma}_{\varepsilon_{m}}^{2} \right)}
}
\item{sigma2epsilonyhat}{
Mean of estimated error variance of \code{y} \eqn{\left( \hat{\sigma}_{\varepsilon_{y}}^{2} \right)}
}
\item{muxhat}{
Mean of estimated mean of \code{x} \eqn{\left( \hat{\mu}_x \right)}.
}
\item{sehatdeltayhat}{
Mean of estimated standard error of \eqn{\hat{\delta}_{y}}.
}
\item{sehattaudothat}{
Mean of estimated standard error of \eqn{\hat{\dot{\tau}}}.
}
\item{sehatbetahat}{
Mean of estimated standard error of \eqn{\hat{\beta}}.
}
\item{sehatdeltamhat}{
Mean of estimated standard error of \eqn{\hat{\delta}_{m}}.
}
\item{sehatalphahat}{
Mean of estimated standard error of \eqn{\hat{\alpha}}.
}
\item{theta}{
Population parameter \eqn{\alpha \beta}.
}
\item{deltayhat_var}{
Variance of estimated intercept of \code{y} \eqn{\left( \hat{\delta}_y \right)}.
}
\item{taudothat_var}{
Variance of estimated slope of path from \code{x} to \code{y} \eqn{\left( \hat{\dot{\tau}} \right)}
}
\item{betahat_var}{
Variance of estimated slope of path from \code{m} to \code{y} \eqn{\left( \hat{\beta} \right)}
}
\item{deltamhat_var}{
Variance of estimated intercept of \code{m} \eqn{\left( \hat{\delta_m} \right)}.
}
\item{alphahat_var}{
Variance of estimated slope of path from \code{x} to \code{m} \eqn{\left( \hat{\alpha} \right)}
}
\item{alphahatbetahat_var}{
Variance of estimated indirect effect of \code{x} on \code{y} through \code{m} \eqn{\left( \hat{\alpha} \hat{\beta} \right)}
}
\item{alphahatprimebetahatprime_var}{
Variance of estimated standardized indirect effect of \code{x} on \code{y} through \code{m} \eqn{\left( \hat{\alpha}^{\prime} \hat{\beta}^{\prime} \right)}
}
\item{deltayhat_sd}{
Standard deviation of estimated intercept of \code{y} \eqn{\left( \hat{\delta}_y \right)}.
}
\item{taudothat_sd}{
Standard deviation of estimated slope of path from \code{x} to \code{y} \eqn{\left( \hat{\dot{\tau}} \right)}
}
\item{betahat_sd}{
Standard deviation of estimated slope of path from \code{m} to \code{y} \eqn{\left( \hat{\beta} \right)}
}
\item{deltamhat_sd}{
Standard deviation of estimated intercept of \code{m} \eqn{\left( \hat{\delta_m} \right)}.
}
\item{alphahat_sd}{
Standard deviation of estimated slope of path from \code{x} to \code{m} \eqn{\left( \hat{\alpha} \right)}
}
\item{alphahatbetahat_sd}{
Standard deviation of estimated indirect effect of \code{x} on \code{y} through \code{m} \eqn{\left( \hat{\alpha} \hat{\beta} \right)}
}
\item{alphahatprimebetahatprime_sd}{
Standard deviation of estimated standardized indirect effect of \code{x} on \code{y} through \code{m} \eqn{\left( \hat{\alpha}^{\prime} \hat{\beta}^{\prime} \right)}
}
\item{deltayhat_skew}{
Skewness of estimated intercept of \code{y} \eqn{\left( \hat{\delta}_y \right)}.
}
\item{taudothat_skew}{
Skewness of estimated slope of path from \code{x} to \code{y} \eqn{\left( \hat{\dot{\tau}} \right)}
}
\item{betahat_skew}{
Skewness of estimated slope of path from \code{m} to \code{y} \eqn{\left( \hat{\beta} \right)}
}
\item{deltamhat_skew}{
Skewness of estimated intercept of \code{m} \eqn{\left( \hat{\delta_m} \right)}.
}
\item{alphahat_skew}{
Skewness of estimated slope of path from \code{x} to \code{m} \eqn{\left( \hat{\alpha} \right)}
}
\item{alphahatbetahat_skew}{
Skewness of estimated indirect effect of \code{x} on \code{y} through \code{m} \eqn{\left( \hat{\alpha} \hat{\beta} \right)}
}
\item{alphahatprimebetahatprime_skew}{
Skewness of estimated standardized indirect effect of \code{x} on \code{y} through \code{m} \eqn{\left( \hat{\alpha}^{\prime} \hat{\beta}^{\prime} \right)}
}
\item{deltayhat_kurt}{
Excess kurtosis of estimated intercept of \code{y} \eqn{\left( \hat{\delta}_y \right)}.
}
\item{taudothat_kurt}{
Excess kurtosis of estimated slope of path from \code{x} to \code{y} \eqn{\left( \hat{\dot{\tau}} \right)}
}
\item{betahat_kurt}{
Excess kurtosis of estimated slope of path from \code{m} to \code{y} \eqn{\left( \hat{\beta} \right)}
}
\item{deltamhat_kurt}{
Excess kurtosis of estimated intercept of \code{m} \eqn{\left( \hat{\delta_m} \right)}.
}
\item{alphahat_kurt}{
Excess kurtosis of estimated slope of path from \code{x} to \code{m} \eqn{\left( \hat{\alpha} \right)}
}
\item{alphahatbetahat_kurt}{
Excess kurtosis of estimated indirect effect of \code{x} on \code{y} through \code{m} \eqn{\left( \hat{\alpha} \hat{\beta} \right)}
}
\item{alphahatprimebetahatprime_kurt}{
Excess kurtosis of estimated standardized indirect effect of \code{x} on \code{y} through \code{m} \eqn{\left( \hat{\alpha}^{\prime} \hat{\beta}^{\prime} \right)}
}
\item{deltayhat_bias}{
Bias of estimated intercept of \code{y} \eqn{\left( \hat{\delta}_y \right)}.
}
\item{taudothat_bias}{
Bias of estimated slope of path from \code{x} to \code{y} \eqn{\left( \hat{\dot{\tau}} \right)}
}
\item{betahat_bias}{
Bias of estimated slope of path from \code{m} to \code{y} \eqn{\left( \hat{\beta} \right)}
}
\item{deltamhat_bias}{
Bias of estimated intercept of \code{m} \eqn{\left( \hat{\delta_m} \right)}.
}
\item{alphahat_bias}{
Bias of estimated slope of path from \code{x} to \code{m} \eqn{\left( \hat{\alpha} \right)}
}
\item{alphahatbetahat_bias}{
Bias of estimated indirect effect of \code{x} on \code{y} through \code{m} \eqn{\left( \hat{\alpha} \hat{\beta} \right)}
}
\item{alphahatprimebetahatprime_bias}{
Bias of estimated standardized indirect effect of \code{x} on \code{y} through \code{m} \eqn{\left( \hat{\alpha}^{\prime} \hat{\beta}^{\prime} \right)}
}
\item{deltayhat_mse}{
Mean square error of estimated intercept of \code{y} \eqn{\left( \hat{\delta}_y \right)}.
}
\item{taudothat_mse}{
Mean square error of estimated slope of path from \code{x} to \code{y} \eqn{\left( \hat{\dot{\tau}} \right)}
}
\item{betahat_mse}{
Mean square error of estimated slope of path from \code{m} to \code{y} \eqn{\left( \hat{\beta} \right)}
}
\item{deltamhat_mse}{
Mean square error of estimated intercept of \code{m} \eqn{\left( \hat{\delta_m} \right)}.
}
\item{alphahat_mse}{
Mean square error of estimated slope of path from \code{x} to \code{m} \eqn{\left( \hat{\alpha} \right)}
}
\item{alphahatbetahat_mse}{
Mean square error of estimated indirect effect of \code{x} on \code{y} through \code{m} \eqn{\left( \hat{\alpha} \hat{\beta} \right)}
}
\item{alphahatprimebetahatprime_mse}{
Mean square error of estimated standardized indirect effect of \code{x} on \code{y} through \code{m} \eqn{\left( \hat{\alpha}^{\prime} \hat{\beta}^{\prime} \right)}
}
\item{deltayhat_rmse}{
Root mean square error of estimated intercept of \code{y} \eqn{\left( \hat{\delta}_y \right)}.
}
\item{taudothat_rmse}{
Root mean square error of estimated slope of path from \code{x} to \code{y} \eqn{\left( \hat{\dot{\tau}} \right)}
}
\item{betahat_rmse}{
Root mean square error of estimated slope of path from \code{m} to \code{y} \eqn{\left( \hat{\beta} \right)}
}
\item{deltamhat_rmse}{
Root mean square error of estimated intercept of \code{m} \eqn{\left( \hat{\delta_m} \right)}.
}
\item{alphahat_rmse}{
Root mean square error of estimated slope of path from \code{x} to \code{m} \eqn{\left( \hat{\alpha} \right)}
}
\item{alphahatbetahat_rmse}{
Root mean square error of estimated indirect effect of \code{x} on \code{y} through \code{m} \eqn{\left( \hat{\alpha} \hat{\beta} \right)}
}
\item{alphahatprimebetahatprime_rmse}{
Root mean square error of estimated standardized indirect effect of \code{x} on \code{y} through \code{m} \eqn{\left( \hat{\alpha}^{\prime} \hat{\beta}^{\prime} \right)}
}
\item{std}{
Standardized vs. unstandardize indirect effect.
}
\item{Method}{
Method used. Fit in this case.
}
\item{n_label}{
Sample size labels.
}
\item{alpha_label}{
\eqn{\alpha} labels.
}
\item{beta_label}{
\eqn{\beta} labels.
}
\item{taudot_label}{
\eqn{\dot{\tau}} labels.
}
\item{theta_label}{
\eqn{\theta} labels.
}
}
}
\usage{
results_mvn_fit
}
\description{
Monte Carlo Simulation Results - Model Fitting (Multivariate Normal)
}
\details{
The simple mediation model is given by
\deqn{
    y_i = \delta_y + \dot{\tau} x_i + \beta m_i + \varepsilon_{y_{i}}
  }

\deqn{
    m_i = \delta_m + \alpha x_i + \varepsilon_{m_{i}}
  }

The parameters for the mean structure are
\deqn{
    \boldsymbol{\theta}_{\text{mean structure}} = \left\{ \mu_x, \delta_m, \delta_y \right\} .
  }

The parameters for the covariance strcuture are
\deqn{
    \boldsymbol{\theta}_{\text{covariance structure}} = \left\{ \dot{\tau}, \beta, \alpha, \sigma_{x}^{2},
    \sigma_{\varepsilon_{m}}^{2}, \sigma_{\varepsilon_{y}}^{2} \right\} .
  }
}
\examples{
data(results_mvn_fit, package = "jeksterslabRmedsimple")
head(results_mvn_fit)
str(results_mvn_fit)
}
\seealso{
Other results functions: 
\code{\link{results_mvn_ci}},
\code{\link{results_mvn_std_ci}}
}
\concept{results functions}
\keyword{results}
