% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results_mvn_mar_std_fit.sem.R
\docType{data}
\name{results_mvn_mar_std_fit.sem}
\alias{results_mvn_mar_std_fit.sem}
\title{Results: Fit Standardized Simple Mediation Model for Data Generated from a Multivariate Normal Distribution with Data Missing Completely at Random - Structural Equation Modeling}
\format{
A data frame with the following variables
\describe{
\item{taskid}{
Simulation task identification number.
}
\item{n}{
Sample size.
}
\item{reps}{
Monte Carlo replications.
}
\item{taudot}{
Population slope of path from \code{x} to \code{y} \eqn{\left( \dot{\tau} \right)}.
}
\item{beta}{
Population slope of path from \code{m} to \code{y} \eqn{\left( \beta \right)}.
}
\item{alpha}{
Population slope of path from \code{x} to \code{m} \eqn{\left( \alpha \right)}.
}
\item{alphabeta}{
Population indirect effect of \code{x} on \code{y} through \code{m} \eqn{\left( \alpha \beta \right)}.
}
\item{sigma2x}{
Population variance of \code{x} \eqn{\left( \sigma_{x}^{2} \right)}.
}
\item{sigma2epsilonm}{
Population error variance of \code{m} \eqn{\left( \sigma_{\varepsilon_{m}}^{2} \right)}.
}
\item{sigma2epsilony}{
Population error variance of \code{y} \eqn{\left( \sigma_{\varepsilon_{y}}^{2} \right)}.
}
\item{mux}{
Population mean of \code{x} \eqn{\left( \mu_x \right)}.
}
\item{deltam}{
Population intercept of \code{m} \eqn{\left( \delta_m \right)}.
}
\item{deltay}{
Population intercept of \code{y} \eqn{\left( \delta_y \right)}.
}
\item{lambdaxhat}{
Mean of estimated factor loading \code{xlatent ~ x} \eqn{ \left( \lambda_x \right)}. Numerically equivalent to the standard deviation of \code{x}.
}
\item{lambdamhat}{
Mean of estimated factor loading \code{mlatent ~ m} \eqn{ \left( \lambda_m \right)}. Numerically equivalent to the standard deviation of \code{m}.
}
\item{lambdayhat}{
Mean of estimated factor loading \code{ylatent ~ y} \eqn{ \left( \lambda_y \right)}. Numerically equivalent to the standard deviation of \code{y}.
}
\item{taudothatprime}{
Mean of estimated standardized slope of path from \code{x} to \code{y} \eqn{\left( \hat{\dot{\tau}}^{\prime} \right)}.
}
\item{betahatprime}{
Mean of estimated standardized slope of path from \code{m} to \code{y} \eqn{\left( \hat{\beta}^{\prime} \right)}.
}
\item{alphahatprime}{
Mean of estimated standardized slope of path from \code{x} to \code{m} \eqn{\left( \hat{\alpha}^{\prime} \right)}.
}
\item{sigma2hatepsilonylatenthat}{
Mean of estimated error variance of \code{y} \eqn{\left( \hat{\sigma}_{\varepsilon_{y_{\mathrm{latent}}}}^{2} \right)}.
}
\item{sigma2hatepsilonmlatenthat}{
Mean of estimated error variance of \code{m} \eqn{\left( \hat{\sigma}_{\varepsilon_{m_{\mathrm{latent}}}}^{2} \right)}.
}
\item{muyhat}{
Mean of estimated mean of \code{y} \eqn{\left( \hat{\mu}_y \right)}.
}
\item{mumhat}{
Mean of estimated mean of \code{m} \eqn{\left( \hat{\mu}_m \right)}.
}
\item{muxhat}{
Mean of estimated mean of \code{x} \eqn{\left( \hat{\mu}_x \right)}.
}
\item{alphahatprimebetahatprime}{
Mean of estimated standardized indirect effect of \code{x} on \code{y} through \code{m} \eqn{\left( \hat{\alpha}^{\prime} \hat{\beta}^{\prime} \right)}.
}
\item{sehatlambdaxhat}{
Mean of estimated standard error of estimated factor loading \code{xlatent ~ x} \eqn{ \left( \lambda_x \right)}.
}
\item{sehatlambdamhat}{
Mean of estimated standard error of estimated factor loading \code{mlatent ~ m} \eqn{ \left( \lambda_m \right)}.
}
\item{sehatlambdayhat}{
Mean of estimated standard error of estimated factor loading \code{ylatent ~ y} \eqn{ \left( \lambda_y \right)}.
}
\item{sehattaudothatprime}{
Mean of estimated standard error of estimated standardized slope of path from \code{x} to \code{y} \eqn{\left( \hat{\dot{\tau}}^{\prime} \right)}.
}
\item{sehatbetahatprime}{
Mean of estimated standard error of estimated standardized slope of path from \code{m} to \code{y} \eqn{\left( \hat{\beta}^{\prime} \right)}.
}
\item{sehatalphahatprime}{
Mean of estimated standard error of estimated standardized slope of path from \code{x} to \code{m} \eqn{\left( \hat{\alpha}^{\prime} \right)}.
}
\item{sehatsigma2hatepsilonylatenthat}{
Mean of estimated standard error of estimated error variance of \code{y} \eqn{\left( \hat{\sigma}_{\varepsilon_{y_{\mathrm{latent}}}}^{2} \right)}.
}
\item{sehatsigma2hatepsilonmlatenthat}{
Mean of estimated standard error of estimated error variance of \code{m} \eqn{\left( \hat{\sigma}_{\varepsilon_{m_{\mathrm{latent}}}}^{2} \right)}.
}
\item{sehatmuyhat}{
Mean of standard error of estimated mean of \code{y} \eqn{\left( \hat{\mu}_y \right)}.
}
\item{sehatmumhat}{
Mean of standard error of estimated mean of \code{m} \eqn{\left( \hat{\mu}_m \right)}.
}
\item{sehatmuxhat}{
Mean of standard error of estimated mean of \code{x} \eqn{\left( \hat{\mu}_x \right)}.
}
\item{theta}{
Population parameter \eqn{\alpha^{\prime} \beta^{\prime}}.
}
\item{taudothatprime_var}{
Variance of estimated standardized slope of path from \code{x} to \code{y} \eqn{\left( \hat{\dot{\tau}}^{\prime} \right)}.
}
\item{betahatprime_var}{
Variance of estimated standardized slope of path from \code{m} to \code{y} \eqn{\left( \hat{\beta}^{\prime} \right)}.
}
\item{alphahatprime_var}{
Variance of estimated standardized slope of path from \code{x} to \code{m} \eqn{\left( \hat{\alpha}^{\prime} \right)}.
}
\item{alphahatprimebetahatprime_var}{
Variance of estimated standardized indirect effect of \code{x} on \code{y} through \code{m} \eqn{\left( \hat{\alpha}^{\prime} \hat{\beta}^{\prime} \right)}.
}
\item{taudothatprime_sd}{
Standard deviation of estimated standardized slope of path from \code{x} to \code{y} \eqn{\left( \hat{\dot{\tau}}^{\prime} \right)}.
}
\item{betahatprime_sd}{
Standard deviation of estimated standardized slope of path from \code{m} to \code{y} \eqn{\left( \hat{\beta}^{\prime} \right)}.
}
\item{alphahatprime_sd}{
Standard deviation of estimated standardized slope of path from \code{x} to \code{m} \eqn{\left( \hat{\alpha}^{\prime} \right)}.
}
\item{alphahatprimebetahatprime_sd}{
Standard deviation of estimated standardized indirect effect of \code{x} on \code{y} through \code{m} \eqn{\left( \hat{\alpha}^{\prime} \hat{\beta}^{\prime} \right)}.
}
\item{taudothatprime_skew}{
Skewness of estimated standardized slope of path from \code{x} to \code{y} \eqn{\left( \hat{\dot{\tau}}^{\prime} \right)}.
}
\item{betahatprime_skew}{
Skewness of estimated standardized slope of path from \code{m} to \code{y} \eqn{\left( \hat{\beta}^{\prime} \right)}.
}
\item{alphahatprime_skew}{
Skewness of estimated standardized slope of path from \code{x} to \code{m} \eqn{\left( \hat{\alpha}^{\prime} \right)}.
}
\item{alphahatprimebetahatprime_skew}{
Skewness of estimated standardized indirect effect of \code{x} on \code{y} through \code{m} \eqn{\left( \hat{\alpha}^{\prime} \hat{\beta}^{\prime} \right)}.
}
\item{taudothatprime_kurt}{
Excess kurtosis of estimated standardized slope of path from \code{x} to \code{y} \eqn{\left( \hat{\dot{\tau}}^{\prime} \right)}.
}
\item{betahatprime_kurt}{
Excess kurtosis of estimated standardized slope of path from \code{m} to \code{y} \eqn{\left( \hat{\beta}^{\prime} \right)}.
}
\item{alphahatprime_kurt}{
Excess kurtosis of estimated standardized slope of path from \code{x} to \code{m} \eqn{\left( \hat{\alpha}^{\prime} \right)}.
}
\item{alphahatprimebetahatprime_kurt}{
Excess kurtosis of estimated standardized indirect effect of \code{x} on \code{y} through \code{m} \eqn{\left( \hat{\alpha}^{\prime} \hat{\beta}^{\prime} \right)}.
}
\item{taudothatprime_bias}{
Bias of estimated standardized slope of path from \code{x} to \code{y} \eqn{\left( \hat{\dot{\tau}}^{\prime} \right)}.
}
\item{betahatprime_bias}{
Bias of estimated standardized slope of path from \code{m} to \code{y} \eqn{\left( \hat{\beta}^{\prime} \right)}.
}
\item{alphahatprime_bias}{
Bias of estimated standardized slope of path from \code{x} to \code{m} \eqn{\left( \hat{\alpha}^{\prime} \right)}.
}
\item{alphahatprimebetahatprime_bias}{
Bias of estimated standardized indirect effect of \code{x} on \code{y} through \code{m} \eqn{\left( \hat{\alpha}^{\prime} \hat{\beta}^{\prime} \right)}.
}
\item{taudothatprime_mse}{
Mean square error of estimated standardized slope of path from \code{x} to \code{y} \eqn{\left( \hat{\dot{\tau}}^{\prime} \right)}.
}
\item{betahatprime_mse}{
Mean square error of estimated standardized slope of path from \code{m} to \code{y} \eqn{\left( \hat{\beta}^{\prime} \right)}.
}
\item{alphahatprime_mse}{
Mean square error of estimated standardized slope of path from \code{x} to \code{m} \eqn{\left( \hat{\alpha}^{\prime} \right)}.
}
\item{alphahatprimebetahatprime_mse}{
Mean square error of estimated standardized indirect effect of \code{x} on \code{y} through \code{m} \eqn{\left( \hat{\alpha}^{\prime} \hat{\beta}^{\prime} \right)}.
}
\item{taudothatprime_rmse}{
Root mean square error of estimated standardized slope of path from \code{x} to \code{y} \eqn{\left( \hat{\dot{\tau}}^{\prime} \right)}.
}
\item{betahatprime_rmse}{
Root mean square error of estimated standardized slope of path from \code{m} to \code{y} \eqn{\left( \hat{\beta}^{\prime} \right)}.
}
\item{alphahatprime_rmse}{
Root mean square error of estimated standardized slope of path from \code{x} to \code{m} \eqn{\left( \hat{\alpha}^{\prime} \right)}.
}
\item{alphahatprimebetahatprime_rmse}{
Root mean square error of estimated standardized indirect effect of \code{x} on \code{y} through \code{m} \eqn{\left( \hat{\alpha}^{\prime} \hat{\beta}^{\prime} \right)}.
}
\item{missing}{
Type of missingness.
}
\item{std}{
Standardized vs. unstandardize indirect effect.
}
\item{Method}{
Method used. Fit in this case.
}
\item{n_label}{
Sample size labels.
}
\item{alpha_label}{
\eqn{\alpha} labels.
}
\item{beta_label}{
\eqn{\beta} labels.
}
\item{taudot_label}{
\eqn{\dot{\tau}} labels.
}
\item{theta_label}{
\eqn{\theta} labels.
}
}
}
\usage{
results_mvn_mar_std_fit.sem
}
\description{
Results: Fit Standardized Simple Mediation Model for Data Generated from a Multivariate Normal Distribution with Data Missing Completely at Random - Structural Equation Modeling
}
\details{
The standardized simple mediation model is given by the following measurement model and regression model.

Measurement model
\deqn{x_{\mathrm{latent}} = \lambda_x x}
\deqn{m_{\mathrm{latent}} = \lambda_m m}
\deqn{y_{\mathrm{latent}} = \lambda_y y}

Regression model
\deqn{y_{\mathrm{latent}} = \dot{\tau}^{\prime} x_{\mathrm{latent}} + \beta^{\prime} m_{\mathrm{latent}} + \varepsilon_{y_{\mathrm{latent}}}}
\deqn{m_{\mathrm{latent}} = \alpha^{\prime} x_{\mathrm{latent}} + \varepsilon_{m_{\mathrm{latent}}}}
\itemize{
\item The measurement errors in the measurement model are fixed to \code{0}
\item The variance of \eqn{x_{\mathrm{latent}}} \eqn{\left( \sigma_{x_{\mathrm{latent}}}^{2} \right)} is fixed to \code{1}
\item The error variance \eqn{\varepsilon_{y_{\mathrm{latent}}}} \eqn{\left( \sigma_{\varepsilon_{y_{\mathrm{latent}}}}^{2} \right)}
is constrained to \eqn{1 - \dot{\tau}^{\prime 2} - \beta^{\prime 2} -  2 \dot{\tau}^{\prime} \beta^{\prime} \alpha^{\prime}}
\item The error variance \eqn{\varepsilon_{m_{\mathrm{latent}}}} \eqn{\left( \sigma_{\varepsilon_{m_{\mathrm{latent}}}}^{2} \right)}
is constrained to \eqn{1 - \alpha^{\prime 2}}
}
}
\examples{
data(results_mvn_mar_std_fit.sem, package = "jeksterslabRmedsimple")
head(results_mvn_mar_std_fit.sem)
str(results_mvn_mar_std_fit.sem)
}
\seealso{
Other results functions: 
\code{\link{results_mvn_ci}},
\code{\link{results_mvn_fit.ols}},
\code{\link{results_mvn_mar_ci}},
\code{\link{results_mvn_mar_fit.sem}},
\code{\link{results_mvn_mcar_ci}},
\code{\link{results_mvn_mcar_fit.sem}},
\code{\link{results_mvn_mcar_std_fit.sem}},
\code{\link{results_mvn_mnar_ci}},
\code{\link{results_mvn_mnar_fit.sem}},
\code{\link{results_mvn_std_ci}},
\code{\link{results_mvn_std_fit.sem}}
}
\concept{results functions}
\keyword{results}
