% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results_mvn_mnar_nb.fiml_ci.R
\docType{data}
\name{results_mvn_mnar_nb.fiml_ci}
\alias{results_mvn_mnar_nb.fiml_ci}
\title{Monte Carlo Simulation Results - Nonparametric Bootstrap Confidence Intervals (Multivariate Normal with Data Missing Not at Random)}
\format{
A data frame with the following variables
\describe{
\item{taskid}{
Simulation task identification number.
}
\item{n}{
Sample size.
}
\item{simreps}{
Monte Carlo replications.
}
\item{taudot}{
Population slope of path from \code{x} to \code{y} \eqn{\left( \dot{\tau} \right)}
}
\item{beta}{
Population slope of path from \code{m} to \code{y} \eqn{\left( \beta \right)}
}
\item{alpha}{
Population slope of path from \code{x} to \code{m} \eqn{\left( \alpha \right)}
}
\item{alphabeta}{
Population indirect effect of \code{x} on \code{y} through \code{m} \eqn{\left( \alpha \beta \right)}
}
\item{sigma2x}{
Population variance of \code{x} \eqn{\left( \sigma_{x}^{2} \right)}
}
\item{sigma2epsilonm}{
Population error variance of \code{m} \eqn{\left( \sigma_{\varepsilon_{m}}^{2} \right)}
}
\item{sigma2epsilony}{
Population error variance of \code{y} \eqn{\left( \sigma_{\varepsilon_{y}}^{2} \right)}
}
\item{mux}{
Population mean of \code{x} \eqn{\left( \mu_x \right)}.
}
\item{deltam}{
Population intercept of \code{m} \eqn{\left( \delta_m \right)}.
}
\item{deltay}{
Population intercept of \code{y} \eqn{\left( \delta_y \right)}.
}
\item{est}{
Mean of the estimate of the indirect effect \eqn{\left( \hat{\alpha} \hat{\beta} \right)}.
}
\item{se}{
Mean of the estimate of standard error of the indirect effect \eqn{\left( \hat{\alpha} \hat{\beta} \right)}.
}
\item{reps}{
Monte Carlo method of bootstrap replications.
}
\item{ci_0.05}{
Mean of the lower limit confidence interval for the 99.9\% confidence interval.
}
\item{ci_0.5}{
Mean of the lower limit confidence interval for the 99\% confidence interval.
}
\item{ci_2.5}{
Mean of the lower limit confidence interval for the 95\% confidence interval.
}
\item{ci_97.5}{
Mean of the upper limit confidence interval for the 95\% confidence interval.
}
\item{ci_99.5}{
Mean of the upper limit confidence interval for the 99\% confidence interval.
}
\item{ci_99.95}{
Mean of the upper limit confidence interval for the 99.9\% confidence interval.
}
\item{zero_hit_99.9}{
Mean zero hit for the 99.9\% confidence interval.
}
\item{zero_hit_99}{
Mean zero hit for the 99\% confidence interval.
}
\item{zero_hit_95}{
Mean zero hit for the 95\% confidence interval.
}
\item{len_99.9}{
Mean confidence interval length for the 99.9\% confidence interval.
}
\item{len_99}{
Mean confidence interval length for the 99\% confidence interval.
}
\item{len_95}{
Mean confidence interval length for the 95\% confidence interval.
}
\item{shape_99.9}{
Mean confidence interval shape for the 99.9\% confidence interval.
}
\item{shape_99}{
Mean confidence interval shape for the 99\% confidence interval.
}
\item{shape_95}{
Mean confidence interval shape for the 95\% confidence interval.
}
\item{theta_hit_99.9}{
Mean theta hit for the 99.9\% confidence interval.
}
\item{theta_hit_99}{
Mean theta hit for the 99\% confidence interval.
}
\item{theta_hit_95}{
Mean theta hit for the 95\% confidence interval.
}
\item{theta_miss_99.9}{
Mean theta miss for the 99.9\% confidence interval.
}
\item{theta_miss_99}{
Mean theta miss for the 99\% confidence interval.
}
\item{theta_miss_95}{
Mean theta miss for the 95\% confidence interval.
}
\item{theta}{
Population parameter \eqn{\alpha \beta}.
}
\item{power_99.9}{
Mean power for the 99.9\% confidence interval.
}
\item{power_99}{
Mean power for the 99\% confidence interval.
}
\item{power_95}{
Mean power for the 95\% confidence interval.
}
\item{liberal_ll_99.9}{
Lower limit of the liberal criteria for the 99.9\% confidence interval.
}
\item{liberal_ul_99.9}{
Upper limit of the liberal criteria for the 99.9\% confidence interval.
}
\item{moderate_ll_99.9}{
Lower limit of the moderate criteria for the 99.9\% confidence interval.
}
\item{moderate_ul_99.9}{
Upper limit of the moderate criteria for the 99.9\% confidence interval.
}
\item{strict_ll_99.9}{
Lower limit of the strict criteria for the 99.9\% confidence interval.
}
\item{strict_ul_99.9}{
Upper limit of the strict criteria for the 99.9\% confidence interval.
}
\item{liberal_ll_99}{
Lower limit of the liberal criteria for the 99\% confidence interval.
}
\item{liberal_ul_99}{
Upper limit of the liberal criteria for the 99\% confidence interval.
}
\item{moderate_ll_99}{
Lower limit of the moderate criteria for the 99\% confidence interval.
}
\item{moderate_ul_99}{
Upper limit of the moderate criteria for the 99\% confidence interval.
}
\item{strict_ll_99}{
Lower limit of the strict criteria for the 99\% confidence interval.
}
\item{strict_ul_99}{
Upper limit of the strict criteria for the 99\% confidence interval.
}
\item{liberal_ll_95}{
Lower limit of the liberal criteria for the 95\% confidence interval.
}
\item{liberal_ul_95}{
Upper limit of the liberal criteria for the 95\% confidence interval.
}
\item{moderate_ll_95}{
Lower limit of the moderate criteria for the 95\% confidence interval.
}
\item{moderate_ul_95}{
Upper limit of the moderate criteria for the 95\% confidence interval.
}
\item{strict_ll_95}{
Lower limit of the strict criteria for the 95\% confidence interval.
}
\item{strict_ul_95}{
Upper limit of the strict criteria for the 95\% confidence interval.
}
\item{serlin_ll_95}{
Lower limit of the Serlin criteria for the 95\% confidence interval.
}
\item{serlin_ul_95}{
Upper limit of the Serlin criteria for the 95\% confidence interval.
}
\item{liberal_99.9}{
Logical. 1 if miss rate is inside the liberal robustness criteria for 99.9\% confidence interval.
}
\item{liberal_99}{
Logical. 1 if miss rate is inside the liberal robustness criteria for 99\% confidence interval.
}
\item{liberal_95}{
Logical. 1 if miss rate is inside the liberal robustness criteria for 95\% confidence interval.
}
\item{moderate_99.9}{
Logical. 1 if miss rate is inside the moderate robustness criteria for 99.9\% confidence interval.
}
\item{moderate_99}{
Logical. 1 if miss rate is inside the moderate robustness criteria for 99\% confidence interval.
}
\item{moderate_95}{
Logical. 1 if miss rate is inside the moderate robustness criteria for 95\% confidence interval.
}
\item{strict_99.9}{
Logical. 1 if miss rate is inside the strict robustness criteria for 99.9\% confidence interval.
}
\item{strict_99}{
Logical. 1 if miss rate is inside the strict robustness criteria for 99\% confidence interval.
}
\item{strict_95}{
Logical. 1 if miss rate is inside the strict robustness criteria for 95\% confidence interval.
}
\item{serlin_95}{
Logical. 1 if miss rate is inside the Serlin robustness criteria for 95\% confidence interval.
}
\item{missing}{
Type of missingness.
}
\item{std}{
Standardized vs. unstandardize indirect effect.
}
\item{Method}{
Method used. Fit in this case.
}
\item{n_label}{
Sample size labels.
}
\item{alpha_label}{
\eqn{\alpha} labels.
}
\item{beta_label}{
\eqn{\beta} labels.
}
\item{taudot_label}{
\eqn{\dot{\tau}} labels.
}
\item{theta_label}{
\eqn{\theta} labels.
}
}
}
\usage{
results_mvn_mnar_nb.fiml_ci
}
\description{
Monte Carlo Simulation Results - Nonparametric Bootstrap Confidence Intervals (Multivariate Normal with Data Missing Not at Random)
}
\details{
The simple mediation model is given by
\deqn{
    y_i = \delta_y + \dot{\tau} x_i + \beta m_i + \varepsilon_{y_{i}}
  }

\deqn{
    m_i = \delta_m + \alpha x_i + \varepsilon_{m_{i}}
  }

The parameters for the mean structure are
\deqn{
    \boldsymbol{\theta}_{\text{mean structure}} = \left\{ \mu_x, \delta_m, \delta_y \right\} .
  }

The parameters for the covariance strcuture are
\deqn{
    \boldsymbol{\theta}_{\text{covariance structure}} = \left\{ \dot{\tau}, \beta, \alpha, \sigma_{x}^{2},
    \sigma_{\varepsilon_{m}}^{2}, \sigma_{\varepsilon_{y}}^{2} \right\} .
  }
}
\examples{
data(results_mvn_mnar_nb.fiml_ci, package = "jeksterslabRmedsimple")
head(results_mvn_mnar_nb.fiml_ci)
str(results_mvn_mnar_nb.fiml_ci)
}
\seealso{
Other results: 
\code{\link{results_mvn_fit.ols}},
\code{\link{results_mvn_fit.sem}},
\code{\link{results_mvn_mar_fit.sem}},
\code{\link{results_mvn_mar_mc.mvn_ci}},
\code{\link{results_mvn_mar_nb.fiml_ci}},
\code{\link{results_mvn_mcar_fit.sem}},
\code{\link{results_mvn_mcar_mc.mvn_ci}},
\code{\link{results_mvn_mcar_nb.fiml_ci}},
\code{\link{results_mvn_mnar_fit.sem}},
\code{\link{results_mvn_mnar_mc.mvn_ci}},
\code{\link{results_mvn_nb_ci}},
\code{\link{results_mvn_ols_mc.mvn_ci}},
\code{\link{results_mvn_pb.mvn_ci}},
\code{\link{results_mvn_sem_mc.mvn_ci}},
\code{\link{results_vm_mod_fit.ols}},
\code{\link{results_vm_mod_fit.sem.mlr}},
\code{\link{results_vm_mod_nb_ci}},
\code{\link{results_vm_mod_ols_mc.mvn_ci}},
\code{\link{results_vm_mod_pb.mvn_ci}},
\code{\link{results_vm_mod_sem_mc.mvn_ci}},
\code{\link{results_vm_sev_fit.ols}},
\code{\link{results_vm_sev_fit.sem.mlr}},
\code{\link{results_vm_sev_nb_ci}},
\code{\link{results_vm_sev_ols_mc.mvn_ci}},
\code{\link{results_vm_sev_pb.mvn_ci}},
\code{\link{results_vm_sev_sem_mc.mvn_ci}}
}
\concept{results}
\keyword{results}
