% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ram.R
\name{S}
\alias{S}
\title{S Matrix}
\usage{
S(sigma2x, sigma2epsilonm, sigma2epsilony)
}
\arguments{
\item{sigma2x}{Numeric.
Variance of \code{x} \eqn{\left( \sigma_{x}^{2} \right)}.}

\item{sigma2epsilonm}{Numeric.
Residual variance of \eqn{\varepsilon_{m_{i}}} \eqn{\left( \sigma_{\varepsilon_m}^{2} \right)}.}

\item{sigma2epsilony}{Numeric.
Residual variance of \eqn{\varepsilon_{y_{i}}} \eqn{\left( \sigma_{\varepsilon_y}^{2} \right)}.}
}
\description{
Matrix of variance of \code{x} and residual variances from the simple mediation model.
}
\details{
The simple mediation model is given by
\deqn{
    y_i = \delta_y + \dot{\tau} x_i + \beta m_i + \varepsilon_{y_{i}}
  }

\deqn{
    m_i = \delta_m + \alpha x_i + \varepsilon_{m_{i}}
  }

The parameters for the mean structure are
\deqn{
    \boldsymbol{\theta}_{\text{mean structure}} = \left\{ \mu_x, \delta_m, \delta_y \right\} .
  }

The parameters for the covariance strcuture are
\deqn{
    \boldsymbol{\theta}_{\text{covariance structure}} = \left\{ \dot{\tau}, \beta, \alpha, \sigma_{x}^{2},
    \sigma_{\varepsilon_{m}}^{2}, \sigma_{\varepsilon_{y}}^{2} \right\} .
  }
}
\examples{
S(
  sigma2x = 1.293469,
  sigma2epsilonm = 0.9296691,
  sigma2epsilony = 0.9310597
)
}
\seealso{
Other reticular action model functions: 
\code{\link{A}()},
\code{\link{Mfrommu}()},
\code{\link{M}()},
\code{\link{Sfromsigma2}()},
\code{\link{Sigmathetafromsigma2}()},
\code{\link{Sigmatheta}()},
\code{\link{mutheta}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{reticular action model functions}
\keyword{ram}
