% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ram.R
\name{S.std}
\alias{S.std}
\title{Standardized S Matrix}
\usage{
S.std(sigma2epsilonm, sigma2epsilony)
}
\arguments{
\item{sigma2epsilonm}{Numeric.
Error variance of \eqn{\varepsilon_{m_{latent_{i}}}} \eqn{\left( \sigma_{\varepsilon_{m_{latent}}}^{2} \right)}.}

\item{sigma2epsilony}{Numeric.
Error variance of \eqn{\varepsilon_{y_{latent_{i}}}} \eqn{\left( \sigma_{\varepsilon_{y_{latent}}}^{2} \right)}.}
}
\description{
Matrix of error variances from the standardized simple mediation model.
}
\examples{
S.std(sigma2epsilonm = 0.8624347, sigma2epsilony = 0.7227811)
}
\seealso{
Other reticular action model functions: 
\code{\link{A.std}()},
\code{\link{A}()},
\code{\link{Mfrommu}()},
\code{\link{M}()},
\code{\link{Sfromsigma2}()},
\code{\link{Sigmatheta.std}()},
\code{\link{Sigmathetafromsigma2}()},
\code{\link{Sigmatheta}()},
\code{\link{S}()},
\code{\link{mutheta}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{reticular action model functions}
\keyword{ram}
