% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvn.R
\name{setparamsmvn}
\alias{setparamsmvn}
\title{Set Parameters for Data Generated from a Multivariate Normal Distribution}
\usage{
setparamsmvn(taskid)
}
\arguments{
\item{taskid}{Numeric.
Task ID.}
}
\description{
Set Parameters for Data Generated from a Multivariate Normal Distribution
}
\details{
The simple mediation model is given by
\deqn{
    y_i = \delta_y + \dot{\tau} x_i + \beta m_i + \varepsilon_{y_{i}}
  }

\deqn{
    m_i = \delta_m + \alpha x_i + \varepsilon_{m_{i}}
  }

The parameters for the mean structure are
\deqn{
    \boldsymbol{\theta}_{\text{mean structure}} = \left\{ \mu_x, \delta_m, \delta_y \right\} .
  }

The parameters for the covariance strcuture are
\deqn{
    \boldsymbol{\theta}_{\text{covariance structure}} = \left\{ \dot{\tau}, \beta, \alpha, \sigma_{x}^{2},
    \sigma_{\varepsilon_{m}}^{2}, \sigma_{\varepsilon_{y}}^{2} \right\} .
  }
}
\examples{
setparamsmvn(
  taskid = 1
)
}
\seealso{
Other multivariate normal data functions: 
\code{\link{mvn_dat_mar_simulation}()},
\code{\link{mvn_dat_mar_task}()},
\code{\link{mvn_dat_mar}()},
\code{\link{mvn_dat_mcar_simulation}()},
\code{\link{mvn_dat_mcar_task}()},
\code{\link{mvn_dat_mcar}()},
\code{\link{mvn_dat_mnar_simulation}()},
\code{\link{mvn_dat_mnar_task}()},
\code{\link{mvn_dat_mnar}()},
\code{\link{mvn_dat_simulation}()},
\code{\link{mvn_dat_task}()},
\code{\link{mvn_dat}()},
\code{\link{mvn}()},
\code{\link{useparamsmvn}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{multivariate normal data functions}
\keyword{mvn}
