% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ram.R
\name{Sigmatheta}
\alias{Sigmatheta}
\title{Model-Implied Variance-Covariance Matrix}
\usage{
Sigmatheta(taudot, beta, alpha, sigma2x, sigma2epsilonm, sigma2epsilony)
}
\arguments{
\item{taudot}{Numeric.
Slope of path from \code{x} to \code{y} \eqn{\left( \dot{\tau} \right)}.}

\item{beta}{Numeric.
Slope of path from \code{m} to \code{y} \eqn{\left( \beta \right)} .}

\item{alpha}{Numeric.
Slope of path from \code{x} to \code{m} \eqn{\left( \alpha \right)} .}

\item{sigma2x}{Numeric.
Variance of \code{x} \eqn{\left( \sigma_{x}^{2} \right)}.}

\item{sigma2epsilonm}{Numeric.
Error variance of \eqn{\varepsilon_{m_{i}}} \eqn{\left( \sigma_{\varepsilon_m}^{2} \right)}.}

\item{sigma2epsilony}{Numeric.
Error variance of \eqn{\varepsilon_{y_{i}}} \eqn{\left( \sigma_{\varepsilon_y}^{2} \right)}.}
}
\description{
Model-implied variance-covariance matrix from the simple mediation model.
}
\details{
The simple mediation model is given by
\deqn{
    y_i = \delta_y + \dot{\tau} x_i + \beta m_i + \varepsilon_{y_{i}}
  }

\deqn{
    m_i = \delta_m + \alpha x_i + \varepsilon_{m_{i}}
  }

The parameters for the mean structure are
\deqn{
    \boldsymbol{\theta}_{\text{mean structure}} = \left\{ \mu_x, \delta_m, \delta_y \right\} .
  }

The parameters for the covariance strcuture are
\deqn{
    \boldsymbol{\theta}_{\text{covariance structure}} = \left\{ \dot{\tau}, \beta, \alpha, \sigma_{x}^{2},
    \sigma_{\varepsilon_{m}}^{2}, \sigma_{\varepsilon_{y}}^{2} \right\} .
  }
}
\examples{
Sigmatheta(
  taudot = 0.207648,
  beta = 0.451039,
  alpha = 0.338593,
  sigma2x = 1.293469,
  sigma2epsilonm = 0.9296691,
  sigma2epsilony = 0.9310597
)
cov(jeksterslabRdatarepo::thirst)
}
\seealso{
Other reticular action model functions: 
\code{\link{A.std}()},
\code{\link{A}()},
\code{\link{Mfrommu}()},
\code{\link{M}()},
\code{\link{S.std}()},
\code{\link{Sfromsigma2}()},
\code{\link{Sigmatheta.std}()},
\code{\link{Sigmathetafromsigma2}()},
\code{\link{S}()},
\code{\link{mutheta}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{reticular action model functions}
\keyword{ram}
