% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ram.R
\name{Sigmatheta.std}
\alias{Sigmatheta.std}
\title{Standardized Model-Implied Variance-Covariance Matrix}
\usage{
Sigmatheta.std(
  taudotprime,
  betaprime,
  alphaprime,
  lambdax,
  lambdam,
  lambday,
  sigma2epsilonm,
  sigma2epsilony
)
}
\arguments{
\item{taudotprime}{Numeric.
Standardized slope of path from \code{x} to \code{y} \eqn{\left( \dot{\tau}^{\prime} \right)}.}

\item{betaprime}{Numeric.
Standardized slope of path from \code{m} to \code{y} \eqn{\left( \beta^{\prime} \right)} .}

\item{alphaprime}{Numeric.
Standardized slope of path from \code{x} to \code{m} \eqn{\left( \alpha^{\prime} \right)} .}

\item{lambdax}{Numeric.
Factor loading \code{xlatent ~ x} \eqn{ \left( \lambda_x \right)} .
Numerically equivalent to the standard deviation of \code{x}.}

\item{lambdam}{Numeric.
Factor loading \code{mlatent ~ m} \eqn{ \left( \lambda_m \right)} .
Numerically equivalent to the standard deviation of \code{m}.}

\item{lambday}{Numeric.
Factor loading \code{ylatent ~ y} \eqn{ \left( \lambda_y \right)} .
Numerically equivalent to the standard deviation of \code{y}.}

\item{sigma2epsilonm}{Numeric.
Error variance of \eqn{\varepsilon_{m_{latent_{i}}}} \eqn{\left( \sigma_{\varepsilon_{m_{latent}}}^{2} \right)}.}

\item{sigma2epsilony}{Numeric.
Error variance of \eqn{\varepsilon_{y_{latent_{i}}}} \eqn{\left( \sigma_{\varepsilon_{y_{latent}}}^{2} \right)}.}
}
\description{
Model-implied variance-covariance matrix from the standardized simple mediation model.
}
\examples{
Sigmatheta.std(
  taudotprime = 0.2080748,
  betaprime = 0.4126006,
  alphaprime = 0.3708979,
  lambdax = 1.137308,
  lambdam = 1.038248,
  lambday = 1.134973,
  sigma2epsilonm = 0.8624347,
  sigma2epsilony = 0.7227811
)
cov(jeksterslabRdatarepo::thirst)
}
\seealso{
Other reticular action model functions: 
\code{\link{A.std}()},
\code{\link{A}()},
\code{\link{Mfrommu}()},
\code{\link{M}()},
\code{\link{S.std}()},
\code{\link{Sfromsigma2}()},
\code{\link{Sigmathetafromsigma2}()},
\code{\link{Sigmatheta}()},
\code{\link{S}()},
\code{\link{mutheta}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{reticular action model functions}
\keyword{ram}
