% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ram.R
\name{Sigmathetafromsigma2}
\alias{Sigmathetafromsigma2}
\title{Model-Implied Variance-Covariance Matrix}
\usage{
Sigmathetafromsigma2(taudot, beta, alpha, sigma2x, sigma2m, sigma2y)
}
\arguments{
\item{taudot}{Numeric.
Slope of path from \code{x} to \code{y} \eqn{\left( \dot{\tau} \right)}.}

\item{beta}{Numeric.
Slope of path from \code{m} to \code{y} \eqn{\left( \beta \right)} .}

\item{alpha}{Numeric.
Slope of path from \code{x} to \code{m} \eqn{\left( \alpha \right)} .}

\item{sigma2x}{Numeric.
Variance of \code{x} \eqn{\left( \sigma_{x}^{2} \right)}.}

\item{sigma2m}{Numeric.
Variance of \code{m} \eqn{\left( \sigma_{m}^{2} \right)} .}

\item{sigma2y}{Numeric.
Variance of \code{y} \eqn{\left( \sigma_{y}^{2} \right)} .}
}
\description{
Model-implied variance-covariance matrix from the simple mediation model.
}
\details{
The simple mediation model is given by
\deqn{
    y_i = \delta_y + \dot{\tau} x_i + \beta m_i + \varepsilon_{y_{i}}
  }

\deqn{
    m_i = \delta_m + \alpha x_i + \varepsilon_{m_{i}}
  }

The parameters for the mean structure are
\deqn{
    \boldsymbol{\theta}_{\text{mean structure}} = \left\{ \mu_x, \delta_m, \delta_y \right\} .
  }

The parameters for the covariance strcuture are
\deqn{
    \boldsymbol{\theta}_{\text{covariance structure}} = \left\{ \dot{\tau}, \beta, \alpha, \sigma_{x}^{2},
    \sigma_{\varepsilon_{m}}^{2}, \sigma_{\varepsilon_{y}}^{2} \right\} .
  }
}
\examples{
Sigmathetafromsigma2(
  taudot = 0.207648,
  beta = 0.451039,
  alpha = 0.338593,
  sigma2x = 1.2934694,
  sigma2m = 1.0779592,
  sigma2y = 1.2881633
)
}
\seealso{
Other reticular action model functions: 
\code{\link{A}()},
\code{\link{Mfrommu}()},
\code{\link{M}()},
\code{\link{Sfromsigma2}()},
\code{\link{Sigmatheta}()},
\code{\link{S}()},
\code{\link{mutheta}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{reticular action model functions}
\keyword{ram}
