% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval.R
\name{theta_hit}
\alias{theta_hit}
\title{Confidence Interval - Theta Hit}
\usage{
theta_hit(lo, theta, up)
}
\arguments{
\item{lo}{Numeric.
Lower limit of the estimated confidence interval
\eqn{\left( \hat{\theta}_{\mathrm{lo}} \right)}.}

\item{theta}{Numeric.
Population parameter
\eqn{\left( \theta \right)}.}

\item{up}{Numeric.
Upper limit of the estimated confidence interval
\eqn{\left( \hat{\theta}_{\mathrm{up}} \right)}.}
}
\value{
Returns
\code{TRUE} if \code{theta} \eqn{\left( \theta \right)} is between the interval
\code{lo}
\eqn{\left( \hat{\theta}_{\mathrm{lo}} \right)}
to
\code{up}
\eqn{\left( \hat{\theta}_{\mathrm{up}} \right)}.
Returns
\code{FALSE} if \code{theta} \eqn{\left( \theta \right)} is outside the interval
\code{lo}
\eqn{\left( \hat{\theta}_{\mathrm{lo}} \right)}
to
\code{up}
\eqn{\left( \hat{\theta}_{\mathrm{up}} \right)}.
}
\description{
Confidence Interval - Theta Hit
}
\examples{
# FALSE
theta_hit(lo = 1, theta = 0, up = 2)
# TRUE
theta_hit(lo = -1, theta = 0, up = 1)
}
\seealso{
Other confidence intervals functions: 
\code{\link{bcaci}()},
\code{\link{bcci}()},
\code{\link{evalci}()},
\code{\link{len}()},
\code{\link{pcci}()},
\code{\link{shape}()},
\code{\link{zero_hit}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{confidence intervals functions}
\keyword{ci}
