% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp.R
\name{useparamsexp}
\alias{useparamsexp}
\title{Use Parameters (Exponential)}
\usage{
useparamsexp(taskid)
}
\arguments{
\item{taskid}{Numeric.
Task ID.}
}
\description{
Use Parameters (Exponential)
}
\details{
The simple mediation model is given by
\deqn{
    y_i = \delta_y + \dot{\tau} x_i + \beta m_i + \varepsilon_{y_{i}}
  }

\deqn{
    m_i = \delta_m + \alpha x_i + \varepsilon_{m_{i}}
  }

The parameters for the mean structure are
\deqn{
    \boldsymbol{\theta}_{\text{mean structure}} = \left\{ \mu_x, \delta_m, \delta_y \right\} .
  }

The parameters for the covariance structure are
\deqn{
    \boldsymbol{\theta}_{\text{covariance structure}} = \left\{ \dot{\tau}, \beta, \alpha, \sigma_{x}^{2},
    \sigma_{\varepsilon_{m}}^{2}, \sigma_{\varepsilon_{y}}^{2} \right\} .
  }
}
\examples{
useparamsexp(taskid = 1)
}
\seealso{
Other exponential X data functions: 
\code{\link{exp_dat_simulation}()},
\code{\link{exp_dat_task}()},
\code{\link{exp_dat}()},
\code{\link{genexp}()},
\code{\link{setparamsexp}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{exponential X data functions}
\keyword{exp}
