% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vm.R
\name{vm}
\alias{vm}
\title{Generate Nonnormal Data Using the Vale and Maurelli (1983) Approach}
\usage{
vm(n, mu, Sigma, skewness, kurtosis)
}
\arguments{
\item{n}{Integer.
Samples size.}

\item{mu}{Numeric vector.
Mean vector.}

\item{Sigma}{Numeric matrix.
Variance-covariance matrix.}

\item{skewness}{Numeric vector. Skewness.}

\item{kurtosis}{Numeric vector. Kurtosis.}
}
\description{
Generate Nonnormal Data Using the Vale and Maurelli (1983) Approach
}
\examples{
n <- 100
mu <- c(0, 0)
Sigma <- matrix(data = c(1, 0.50, 0.50, 1), ncol = 2)
skewness <- c(5, 2)
kurtosis <- c(3, 3)
data <- vm(n = n, mu = mu, Sigma = Sigma, skewness = skewness, kurtosis = kurtosis)
colMeans(data)
cov(data)
apply(X = data, MARGIN = 2, FUN = jeksterslabRdist::skew)
apply(X = data, MARGIN = 2, FUN = jeksterslabRdist::kurt)
data
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{multivariate nonnormal data functions}
\keyword{vm}
