% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vm_mod_complete_dat.R
\name{vm_mod_dat_simulation_task_summary}
\alias{vm_mod_dat_simulation_task_summary}
\title{Generate Nonnormal Data Using the Vale and Maurelli (1983) Approach
following a Simple Mediation Model (Skewness = 2, Kurtosis = 7)
(Summary - Task)}
\usage{
vm_mod_dat_simulation_task_summary(taskid, dir = getwd())
}
\arguments{
\item{taskid}{Numeric.
Task ID.}

\item{dir}{Character string.
Directory where results of \verb{vm_mod_dat_*} are stored.}
}
\description{
Generate Nonnormal Data Using the Vale and Maurelli (1983) Approach
following a Simple Mediation Model (Skewness = 2, Kurtosis = 7)
(Summary - Task)
}
\seealso{
Other multivariate normal data functions: 
\code{\link{mvn_dat_simulation}()},
\code{\link{mvn_dat_task}()},
\code{\link{mvn_dat}()},
\code{\link{mvn}()},
\code{\link{setparamsmvn}()},
\code{\link{useparamsmvn}()},
\code{\link{vm_mod_dat_simulation_simulation_summary}()},
\code{\link{vm_mod_dat_simulation}()},
\code{\link{vm_mod_dat_task}()},
\code{\link{vm_mod_dat}()},
\code{\link{vm_sev_dat_simulation_simulation_summary}()},
\code{\link{vm_sev_dat_simulation_task_summary}()},
\code{\link{vm_sev_dat_simulation}()},
\code{\link{vm_sev_dat_task}()},
\code{\link{vm_sev_dat}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{multivariate normal data functions}
\keyword{vm_dat}
