% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vm_mod_complete_unstd_jack.R
\name{vm_mod_jack}
\alias{vm_mod_jack}
\title{Jackknife Estimates of Indirect Effect in a Simple Mediation Model for Data Generated Using the Vale and Maurelli (1983) Approach (Skewness = 2, Kurtosis = 7)}
\usage{
vm_mod_jack(data, taskid)
}
\arguments{
\item{data}{\code{n} by 3 matrix or data frame.
\code{data[, 1]} correspond to values for \code{x}.
\code{data[, 2]} correspond to values for \code{m}.
\code{data[, 3]} correspond to values for \code{y}.}

\item{taskid}{Numeric.
Task ID.}
}
\description{
Jackknife Estimates of Indirect Effect in a Simple Mediation Model for Data Generated Using the Vale and Maurelli (1983) Approach (Skewness = 2, Kurtosis = 7)
}
\examples{
taskid <- 1
data <- vm_mod_dat(taskid = taskid)
vm_mod_jack(data = data, taskid = taskid)
}
\seealso{
Other jackknife functions: 
\code{\link{jack}()},
\code{\link{mvn_jack_simulation}()},
\code{\link{mvn_jack_task}()},
\code{\link{mvn_jack}()},
\code{\link{mvn_std_jack_simulation}()},
\code{\link{mvn_std_jack_task}()},
\code{\link{mvn_std_jack}()},
\code{\link{vm_mod_jack_simulation}()},
\code{\link{vm_mod_jack_task}()},
\code{\link{vm_mod_std_jack_simulation}()},
\code{\link{vm_mod_std_jack_task}()},
\code{\link{vm_mod_std_jack}()},
\code{\link{vm_sev_jack_simulation}()},
\code{\link{vm_sev_jack_task}()},
\code{\link{vm_sev_jack}()},
\code{\link{vm_sev_std_jack_simulation}()},
\code{\link{vm_sev_std_jack_task}()},
\code{\link{vm_sev_std_jack}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{jackknife functions}
\keyword{jack}
