% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vm_mod_complete_unstd_pb.vm.R
\name{vm_mod_pb.vm}
\alias{vm_mod_pb.vm}
\title{Parametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated Using the Vale and Maurelli (1983) Approach (Skewness = 2, Kurtosis = 7)}
\usage{
vm_mod_pb.vm(data, taskid)
}
\arguments{
\item{data}{\code{n} by 3 matrix or data frame.
\code{data[, 1]} correspond to values for \code{x}.
\code{data[, 2]} correspond to values for \code{m}.
\code{data[, 3]} correspond to values for \code{y}.}

\item{taskid}{Numeric.
Task ID.}
}
\description{
Parametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated Using the Vale and Maurelli (1983) Approach (Skewness = 2, Kurtosis = 7)
}
\examples{
taskid <- 1
data <- vm_mod_dat(taskid = taskid)
vm_mod_pb.vm(data = data, taskid = taskid)
}
\seealso{
Other parametric functions: 
\code{\link{mvn_mar_10_pb.mvn_bcci_simulation}()},
\code{\link{mvn_mar_10_pb.mvn_bcci_task}()},
\code{\link{mvn_mar_10_pb.mvn_pcci_simulation}()},
\code{\link{mvn_mar_10_pb.mvn_pcci_task}()},
\code{\link{mvn_mar_10_pb.mvn_simulation}()},
\code{\link{mvn_mar_10_pb.mvn_task}()},
\code{\link{mvn_mar_10_pb.mvn}()},
\code{\link{mvn_mar_20_pb.mvn_bcci_simulation}()},
\code{\link{mvn_mar_20_pb.mvn_bcci_task}()},
\code{\link{mvn_mar_20_pb.mvn_pcci_simulation}()},
\code{\link{mvn_mar_20_pb.mvn_pcci_task}()},
\code{\link{mvn_mar_20_pb.mvn_simulation}()},
\code{\link{mvn_mar_20_pb.mvn_task}()},
\code{\link{mvn_mar_20_pb.mvn}()},
\code{\link{mvn_mar_30_pb.mvn_bcci_simulation}()},
\code{\link{mvn_mar_30_pb.mvn_bcci_task}()},
\code{\link{mvn_mar_30_pb.mvn_pcci_simulation}()},
\code{\link{mvn_mar_30_pb.mvn_pcci_task}()},
\code{\link{mvn_mar_30_pb.mvn_simulation}()},
\code{\link{mvn_mar_30_pb.mvn_task}()},
\code{\link{mvn_mar_30_pb.mvn}()},
\code{\link{mvn_mcar_10_pb.mvn_bcci_simulation}()},
\code{\link{mvn_mcar_10_pb.mvn_bcci_task}()},
\code{\link{mvn_mcar_10_pb.mvn_pcci_simulation}()},
\code{\link{mvn_mcar_10_pb.mvn_pcci_task}()},
\code{\link{mvn_mcar_10_pb.mvn_simulation}()},
\code{\link{mvn_mcar_10_pb.mvn_task}()},
\code{\link{mvn_mcar_10_pb.mvn}()},
\code{\link{mvn_mcar_20_pb.mvn_bcci_simulation}()},
\code{\link{mvn_mcar_20_pb.mvn_bcci_task}()},
\code{\link{mvn_mcar_20_pb.mvn_pcci_simulation}()},
\code{\link{mvn_mcar_20_pb.mvn_pcci_task}()},
\code{\link{mvn_mcar_20_pb.mvn_simulation}()},
\code{\link{mvn_mcar_20_pb.mvn_task}()},
\code{\link{mvn_mcar_20_pb.mvn}()},
\code{\link{mvn_mcar_30_pb.mvn_bcci_simulation}()},
\code{\link{mvn_mcar_30_pb.mvn_bcci_task}()},
\code{\link{mvn_mcar_30_pb.mvn_pcci_simulation}()},
\code{\link{mvn_mcar_30_pb.mvn_pcci_task}()},
\code{\link{mvn_mcar_30_pb.mvn_simulation}()},
\code{\link{mvn_mcar_30_pb.mvn_task}()},
\code{\link{mvn_mcar_30_pb.mvn}()},
\code{\link{mvn_mnar_10_pb.mvn_bcci_simulation}()},
\code{\link{mvn_mnar_10_pb.mvn_bcci_task}()},
\code{\link{mvn_mnar_10_pb.mvn_pcci_simulation}()},
\code{\link{mvn_mnar_10_pb.mvn_pcci_task}()},
\code{\link{mvn_mnar_10_pb.mvn_simulation}()},
\code{\link{mvn_mnar_10_pb.mvn_task}()},
\code{\link{mvn_mnar_10_pb.mvn}()},
\code{\link{mvn_mnar_20_pb.mvn_bcci_simulation}()},
\code{\link{mvn_mnar_20_pb.mvn_bcci_task}()},
\code{\link{mvn_mnar_20_pb.mvn_pcci_simulation}()},
\code{\link{mvn_mnar_20_pb.mvn_pcci_task}()},
\code{\link{mvn_mnar_20_pb.mvn_simulation}()},
\code{\link{mvn_mnar_20_pb.mvn_task}()},
\code{\link{mvn_mnar_20_pb.mvn}()},
\code{\link{mvn_mnar_30_pb.mvn_bcci_simulation}()},
\code{\link{mvn_mnar_30_pb.mvn_bcci_task}()},
\code{\link{mvn_mnar_30_pb.mvn_pcci_simulation}()},
\code{\link{mvn_mnar_30_pb.mvn_pcci_task}()},
\code{\link{mvn_mnar_30_pb.mvn_simulation}()},
\code{\link{mvn_mnar_30_pb.mvn_task}()},
\code{\link{mvn_mnar_30_pb.mvn}()},
\code{\link{mvn_pb.mvn_bcaci_simulation}()},
\code{\link{mvn_pb.mvn_bcaci_task}()},
\code{\link{mvn_pb.mvn_bcci_simulation}()},
\code{\link{mvn_pb.mvn_bcci_task}()},
\code{\link{mvn_pb.mvn_pcci_simulation}()},
\code{\link{mvn_pb.mvn_pcci_task}()},
\code{\link{mvn_pb.mvn_simulation}()},
\code{\link{mvn_pb.mvn_task}()},
\code{\link{mvn_pb.mvn}()},
\code{\link{mvn_std_pb.mvn_bcaci_simulation}()},
\code{\link{mvn_std_pb.mvn_bcaci_task}()},
\code{\link{mvn_std_pb.mvn_bcci_simulation}()},
\code{\link{mvn_std_pb.mvn_bcci_task}()},
\code{\link{mvn_std_pb.mvn_pcci_simulation}()},
\code{\link{mvn_std_pb.mvn_pcci_task}()},
\code{\link{mvn_std_pb.mvn_simulation}()},
\code{\link{mvn_std_pb.mvn_task}()},
\code{\link{mvn_std_pb.mvn}()},
\code{\link{pb.mvn}()},
\code{\link{pb.vm}()},
\code{\link{vm_mod_pb.mvn_bcaci_simulation}()},
\code{\link{vm_mod_pb.mvn_bcaci_task}()},
\code{\link{vm_mod_pb.mvn_bcci_simulation}()},
\code{\link{vm_mod_pb.mvn_bcci_task}()},
\code{\link{vm_mod_pb.mvn_pcci_simulation}()},
\code{\link{vm_mod_pb.mvn_pcci_task}()},
\code{\link{vm_mod_pb.mvn_simulation}()},
\code{\link{vm_mod_pb.mvn_task}()},
\code{\link{vm_mod_pb.mvn}()},
\code{\link{vm_mod_pb.vm_bcaci_simulation}()},
\code{\link{vm_mod_pb.vm_bcaci_task}()},
\code{\link{vm_mod_pb.vm_bcci_simulation}()},
\code{\link{vm_mod_pb.vm_bcci_task}()},
\code{\link{vm_mod_pb.vm_pcci_simulation}()},
\code{\link{vm_mod_pb.vm_pcci_task}()},
\code{\link{vm_mod_pb.vm_simulation}()},
\code{\link{vm_mod_pb.vm_task}()},
\code{\link{vm_mod_std_pb.mvn_bcaci_simulation}()},
\code{\link{vm_mod_std_pb.mvn_bcaci_task}()},
\code{\link{vm_mod_std_pb.mvn_bcci_simulation}()},
\code{\link{vm_mod_std_pb.mvn_bcci_task}()},
\code{\link{vm_mod_std_pb.mvn_pcci_simulation}()},
\code{\link{vm_mod_std_pb.mvn_pcci_task}()},
\code{\link{vm_mod_std_pb.mvn_simulation}()},
\code{\link{vm_mod_std_pb.mvn_task}()},
\code{\link{vm_mod_std_pb.mvn}()},
\code{\link{vm_sev_pb.mvn_bcaci_simulation}()},
\code{\link{vm_sev_pb.mvn_bcaci_task}()},
\code{\link{vm_sev_pb.mvn_bcci_simulation}()},
\code{\link{vm_sev_pb.mvn_bcci_task}()},
\code{\link{vm_sev_pb.mvn_pcci_simulation}()},
\code{\link{vm_sev_pb.mvn_pcci_task}()},
\code{\link{vm_sev_pb.mvn_simulation}()},
\code{\link{vm_sev_pb.mvn_task}()},
\code{\link{vm_sev_pb.mvn}()},
\code{\link{vm_sev_pb.vm_bcaci_simulation}()},
\code{\link{vm_sev_pb.vm_bcaci_task}()},
\code{\link{vm_sev_pb.vm_bcci_simulation}()},
\code{\link{vm_sev_pb.vm_bcci_task}()},
\code{\link{vm_sev_pb.vm_pcci_simulation}()},
\code{\link{vm_sev_pb.vm_pcci_task}()},
\code{\link{vm_sev_pb.vm_simulation}()},
\code{\link{vm_sev_pb.vm_task}()},
\code{\link{vm_sev_pb.vm}()},
\code{\link{vm_sev_std_pb.mvn_bcaci_simulation}()},
\code{\link{vm_sev_std_pb.mvn_bcaci_task}()},
\code{\link{vm_sev_std_pb.mvn_bcci_simulation}()},
\code{\link{vm_sev_std_pb.mvn_bcci_task}()},
\code{\link{vm_sev_std_pb.mvn_pcci_simulation}()},
\code{\link{vm_sev_std_pb.mvn_pcci_task}()},
\code{\link{vm_sev_std_pb.mvn_simulation}()},
\code{\link{vm_sev_std_pb.mvn_task}()},
\code{\link{vm_sev_std_pb.mvn}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{parametric functions}
\keyword{pb}
