% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vm_mod_complete_unstd_sem_mc.mvn.R
\name{vm_mod_sem_mc.mvn}
\alias{vm_mod_sem_mc.mvn}
\title{Monte Carlo Method Assuming Multivariate Normal Distribution for Indirect Effect in a Simple Mediation Model
for Data Generated Using the Vale and Maurelli (1983) Approach (Skewness = 2, Kurtosis = 7)}
\usage{
vm_mod_sem_mc.mvn(
  taskid,
  R = 20000L,
  alphahat,
  sehatalphahat,
  betahat,
  sehatbetahat
)
}
\arguments{
\item{taskid}{Numeric.
Task ID.}

\item{R}{Integer.
Monte Carlo replications.}

\item{alphahat}{Numeric.
Estimated slope of path from \code{x} to \code{m} \eqn{\left( \hat{\alpha} \right)} .}

\item{sehatalphahat}{Numeric.
Estimated standard error of slope of path from \code{x} to \code{m} \eqn{\left( \widehat{se}_{\hat{\alpha}} \right)} .}

\item{betahat}{Numeric.
Estimated slope of path from \code{m} to \code{y} \eqn{\left( \hat{\beta} \right)} .}

\item{sehatbetahat}{Numeric.
Estimated standard error of slope of path from \code{m} to \code{y} \eqn{\left( \widehat{se}_{\hat{\beta}} \right)} .}
}
\description{
Monte Carlo Method Assuming Multivariate Normal Distribution for Indirect Effect in a Simple Mediation Model
for Data Generated Using the Vale and Maurelli (1983) Approach (Skewness = 2, Kurtosis = 7)
}
\examples{
taskid <- 1
data <- vm_mod_dat(taskid = taskid)
fit.sem.mlr(data = data, minimal = TRUE)

fit <- vm_mod_fit.sem.mlr(data = data, taskid = taskid)
thetahatstar <- vm_mod_sem_mc.mvn(
  taskid = taskid, R = 20000L,
  alphahat = fit["alphahat"], sehatalphahat = fit["sehatalphahat"],
  betahat = fit["betahat"], sehatbetahat = fit["sehatbetahat"]
)
hist(thetahatstar)
}
\seealso{
Other monte carlo method functions: 
\code{\link{beta_ols_mc.mvn_pcci_simulation}()},
\code{\link{beta_ols_mc.mvn_pcci_task}()},
\code{\link{beta_ols_mc.mvn_simulation}()},
\code{\link{beta_ols_mc.mvn_task}()},
\code{\link{beta_ols_mc.mvn}()},
\code{\link{exp_ols_mc.mvn_pcci_simulation}()},
\code{\link{exp_ols_mc.mvn_pcci_task}()},
\code{\link{exp_ols_mc.mvn_simulation}()},
\code{\link{exp_ols_mc.mvn_task}()},
\code{\link{exp_ols_mc.mvn}()},
\code{\link{mc.mvn}()},
\code{\link{mc.t}()},
\code{\link{mc.wishart}()},
\code{\link{mvn_mar_10_mc.mvn_pcci_simulation}()},
\code{\link{mvn_mar_10_mc.mvn_pcci_task}()},
\code{\link{mvn_mar_10_mc.mvn_simulation}()},
\code{\link{mvn_mar_10_mc.mvn_task}()},
\code{\link{mvn_mar_10_mc.mvn}()},
\code{\link{mvn_mar_20_mc.mvn_pcci_simulation}()},
\code{\link{mvn_mar_20_mc.mvn_pcci_task}()},
\code{\link{mvn_mar_20_mc.mvn_simulation}()},
\code{\link{mvn_mar_20_mc.mvn_task}()},
\code{\link{mvn_mar_20_mc.mvn}()},
\code{\link{mvn_mar_30_mc.mvn_pcci_simulation}()},
\code{\link{mvn_mar_30_mc.mvn_pcci_task}()},
\code{\link{mvn_mar_30_mc.mvn_simulation}()},
\code{\link{mvn_mar_30_mc.mvn_task}()},
\code{\link{mvn_mar_30_mc.mvn}()},
\code{\link{mvn_mcar_10_mc.mvn_pcci_simulation}()},
\code{\link{mvn_mcar_10_mc.mvn_pcci_task}()},
\code{\link{mvn_mcar_10_mc.mvn_simulation}()},
\code{\link{mvn_mcar_10_mc.mvn_task}()},
\code{\link{mvn_mcar_10_mc.mvn}()},
\code{\link{mvn_mcar_20_mc.mvn_pcci_simulation}()},
\code{\link{mvn_mcar_20_mc.mvn_pcci_task}()},
\code{\link{mvn_mcar_20_mc.mvn_simulation}()},
\code{\link{mvn_mcar_20_mc.mvn_task}()},
\code{\link{mvn_mcar_20_mc.mvn}()},
\code{\link{mvn_mcar_30_mc.mvn_pcci_simulation}()},
\code{\link{mvn_mcar_30_mc.mvn_pcci_task}()},
\code{\link{mvn_mcar_30_mc.mvn_simulation}()},
\code{\link{mvn_mcar_30_mc.mvn_task}()},
\code{\link{mvn_mcar_30_mc.mvn}()},
\code{\link{mvn_mnar_10_mc.mvn_pcci_simulation}()},
\code{\link{mvn_mnar_10_mc.mvn_pcci_task}()},
\code{\link{mvn_mnar_10_mc.mvn_simulation}()},
\code{\link{mvn_mnar_10_mc.mvn_task}()},
\code{\link{mvn_mnar_10_mc.mvn}()},
\code{\link{mvn_mnar_20_mc.mvn_pcci_simulation}()},
\code{\link{mvn_mnar_20_mc.mvn_pcci_task}()},
\code{\link{mvn_mnar_20_mc.mvn_simulation}()},
\code{\link{mvn_mnar_20_mc.mvn_task}()},
\code{\link{mvn_mnar_20_mc.mvn}()},
\code{\link{mvn_mnar_30_mc.mvn_pcci_simulation}()},
\code{\link{mvn_mnar_30_mc.mvn_pcci_task}()},
\code{\link{mvn_mnar_30_mc.mvn_simulation}()},
\code{\link{mvn_mnar_30_mc.mvn_task}()},
\code{\link{mvn_mnar_30_mc.mvn}()},
\code{\link{mvn_ols_mc.mvn_pcci_simulation}()},
\code{\link{mvn_ols_mc.mvn_pcci_task}()},
\code{\link{mvn_ols_mc.mvn_simulation}()},
\code{\link{mvn_ols_mc.mvn_task}()},
\code{\link{mvn_ols_mc.mvn}()},
\code{\link{mvn_sem_mc.mvn_pcci_simulation}()},
\code{\link{mvn_sem_mc.mvn_pcci_task}()},
\code{\link{mvn_sem_mc.mvn_simulation}()},
\code{\link{mvn_sem_mc.mvn_task}()},
\code{\link{mvn_sem_mc.mvn}()},
\code{\link{mvn_std_mc.mvn.delta_pcci_simulation}()},
\code{\link{mvn_std_mc.mvn.delta_pcci_task}()},
\code{\link{mvn_std_mc.mvn.delta_simulation}()},
\code{\link{mvn_std_mc.mvn.delta_task}()},
\code{\link{mvn_std_mc.mvn.delta}()},
\code{\link{mvn_std_mc.mvn.sem_pcci_simulation}()},
\code{\link{mvn_std_mc.mvn.sem_pcci_task}()},
\code{\link{mvn_std_mc.mvn.sem_simulation}()},
\code{\link{mvn_std_mc.mvn.sem_task}()},
\code{\link{mvn_std_mc.mvn.sem}()},
\code{\link{mvn_std_mc.mvn.tb_pcci_simulation}()},
\code{\link{mvn_std_mc.mvn.tb_pcci_task}()},
\code{\link{mvn_std_mc.mvn.tb_simulation}()},
\code{\link{mvn_std_mc.mvn.tb_task}()},
\code{\link{mvn_std_mc.mvn.tb}()},
\code{\link{mvn_std_mc.wishart_pcci_simulation}()},
\code{\link{mvn_std_mc.wishart_pcci_task}()},
\code{\link{mvn_std_mc.wishart_simulation}()},
\code{\link{mvn_std_mc.wishart_task}()},
\code{\link{mvn_std_mc.wishart}()},
\code{\link{vm_mod_ols_mc.mvn_pcci_simulation}()},
\code{\link{vm_mod_ols_mc.mvn_pcci_task}()},
\code{\link{vm_mod_ols_mc.mvn_simulation}()},
\code{\link{vm_mod_ols_mc.mvn_task}()},
\code{\link{vm_mod_ols_mc.mvn}()},
\code{\link{vm_mod_sem_mc.mvn_pcci_simulation}()},
\code{\link{vm_mod_sem_mc.mvn_pcci_task}()},
\code{\link{vm_mod_sem_mc.mvn_simulation}()},
\code{\link{vm_mod_sem_mc.mvn_task}()},
\code{\link{vm_sev_ols_mc.mvn_pcci_simulation}()},
\code{\link{vm_sev_ols_mc.mvn_pcci_task}()},
\code{\link{vm_sev_ols_mc.mvn_simulation}()},
\code{\link{vm_sev_ols_mc.mvn_task}()},
\code{\link{vm_sev_ols_mc.mvn}()},
\code{\link{vm_sev_sem_mc.mvn_pcci_simulation}()},
\code{\link{vm_sev_sem_mc.mvn_pcci_task}()},
\code{\link{vm_sev_sem_mc.mvn_simulation}()},
\code{\link{vm_sev_sem_mc.mvn_task}()},
\code{\link{vm_sev_sem_mc.mvn}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{monte carlo method functions}
\keyword{mc}
