% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vm_mod_complete_std_fit.sem.mlr.R
\name{vm_mod_std_fit.sem.mlr_simulation}
\alias{vm_mod_std_fit.sem.mlr_simulation}
\title{Fit Simple Mediation Model
for Data Generated Using the Vale and Maurelli (1983) Approach (Skewness = 2, Kurtosis = 7) - Structural Equation Modeling
(Simulation)}
\usage{
vm_mod_std_fit.sem.mlr_simulation(
  dir = getwd(),
  all = TRUE,
  taskid = NULL,
  overwrite = FALSE,
  par = TRUE,
  ncores = NULL,
  blas_threads = TRUE,
  mc = TRUE,
  lb = FALSE,
  cl_eval = FALSE,
  cl_export = FALSE,
  cl_expr,
  cl_vars
)
}
\arguments{
\item{dir}{Character string.
Directory where results of \verb{mvn_dat_*} are stored.}

\item{all}{Logical.
Process all results.}

\item{taskid}{Numeric.
Task ID.}

\item{overwrite}{Logical.
Overwrite existing results.}

\item{par}{Logical.
If \code{TRUE}, use multiple cores.
If \code{FALSE}, use \code{\link[=lapply]{lapply()}}.}

\item{ncores}{Integer.
Number of cores to use if \code{par = TRUE}.
If unspecified, defaults to \code{detectCores() - 1}.}

\item{blas_threads}{Logical.
If \code{TRUE}, set BLAS threads using \code{blas_set_num_threads(threads = blas_get_num_procs())}.
If \code{FALSE}, set BLAS threads using \code{blas_set_num_threads(threads = 1)}.
If \code{par = TRUE}, \code{blas_threads} is automatically set to \code{FALSE}
to prevent conflicts in parallel processing.
This argument is useful when \code{FUN} can handle implicit parallelism
when \code{par = FALSE}, for example linear algebra operations.}

\item{mc}{Logical.
If \code{TRUE}, use \code{\link[parallel:mclapply]{parallel::mclapply()}}.
If \code{FALSE}, use \code{\link[parallel:clusterApply]{parallel::parLapply()}} or \code{\link[parallel:clusterApply]{parallel::parLapplyLB()}}.
Ignored if \code{par = FALSE}.}

\item{lb}{Logical.
If \code{TRUE} use \code{\link[parallel:clusterApply]{parallel::parLapplyLB()}}.
If \code{FALSE}, use \code{\link[parallel:clusterApply]{parallel::parLapply()}}.
Ignored if \code{par = FALSE} and \code{mc = TRUE}.}

\item{cl_eval}{Logical.
Execute \code{\link[parallel:clusterApply]{parallel::clusterEvalQ()}} using \code{cl_expr}.
Ignored if \code{mc = TRUE}.}

\item{cl_export}{Logical.
Execute \code{\link[parallel:clusterApply]{parallel::clusterExport()}} using \code{cl_vars}.
Ignored if \code{mc = TRUE}.}

\item{cl_expr}{Expression.
Expression passed to \code{\link[parallel:clusterApply]{parallel::clusterEvalQ()}}
Ignored if \code{mc = TRUE}.}

\item{cl_vars}{Character vector.
Names of objects to pass to \code{\link[parallel:clusterApply]{parallel::clusterExport()}}
Ignored if \code{mc = TRUE}.}
}
\description{
Fit Simple Mediation Model
for Data Generated Using the Vale and Maurelli (1983) Approach (Skewness = 2, Kurtosis = 7) - Structural Equation Modeling
(Simulation)
}
\seealso{
Other model fit functions: 
\code{\link{beta_fit.ols_simulation_summary}()},
\code{\link{beta_fit.ols_simulation}()},
\code{\link{beta_fit.ols_task_summary}()},
\code{\link{beta_fit.ols_task}()},
\code{\link{beta_fit.ols}()},
\code{\link{beta_fit.sem.mlr_simulation_summary}()},
\code{\link{beta_fit.sem.mlr_simulation}()},
\code{\link{beta_fit.sem.mlr_task_summary}()},
\code{\link{beta_fit.sem.mlr_task}()},
\code{\link{beta_fit.sem.mlr}()},
\code{\link{beta_std_fit.sem.mlr_simulation_summary}()},
\code{\link{beta_std_fit.sem.mlr_simulation}()},
\code{\link{beta_std_fit.sem.mlr_task_summary}()},
\code{\link{beta_std_fit.sem.mlr_task}()},
\code{\link{beta_std_fit.sem.mlr}()},
\code{\link{exp_fit.ols_simulation_summary}()},
\code{\link{exp_fit.ols_simulation}()},
\code{\link{exp_fit.ols_task_summary}()},
\code{\link{exp_fit.ols_task}()},
\code{\link{exp_fit.ols}()},
\code{\link{exp_fit.sem.mlr_simulation_summary}()},
\code{\link{exp_fit.sem.mlr_simulation}()},
\code{\link{exp_fit.sem.mlr_task_summary}()},
\code{\link{exp_fit.sem.mlr_task}()},
\code{\link{exp_fit.sem.mlr}()},
\code{\link{exp_std_fit.sem.mlr_simulation_summary}()},
\code{\link{exp_std_fit.sem.mlr_simulation}()},
\code{\link{exp_std_fit.sem.mlr_task_summary}()},
\code{\link{exp_std_fit.sem.mlr_task}()},
\code{\link{exp_std_fit.sem.mlr}()},
\code{\link{fit.cov}()},
\code{\link{fit.ols}()},
\code{\link{fit.sem.mlr}()},
\code{\link{fit.sem}()},
\code{\link{mvn_fit.ols_simulation_summary}()},
\code{\link{mvn_fit.ols_simulation}()},
\code{\link{mvn_fit.ols_task_summary}()},
\code{\link{mvn_fit.ols_task}()},
\code{\link{mvn_fit.ols}()},
\code{\link{mvn_fit.sem_simulation_summary}()},
\code{\link{mvn_fit.sem_simulation}()},
\code{\link{mvn_fit.sem_task_summary}()},
\code{\link{mvn_fit.sem_task}()},
\code{\link{mvn_fit.sem}()},
\code{\link{mvn_mar_10_fit.sem_simulation_summary}()},
\code{\link{mvn_mar_10_fit.sem_simulation}()},
\code{\link{mvn_mar_10_fit.sem_task_summary}()},
\code{\link{mvn_mar_10_fit.sem_task}()},
\code{\link{mvn_mar_10_fit.sem}()},
\code{\link{mvn_mar_20_fit.sem_simulation_summary}()},
\code{\link{mvn_mar_20_fit.sem_simulation}()},
\code{\link{mvn_mar_20_fit.sem_task_summary}()},
\code{\link{mvn_mar_20_fit.sem_task}()},
\code{\link{mvn_mar_20_fit.sem}()},
\code{\link{mvn_mar_30_fit.sem_simulation_summary}()},
\code{\link{mvn_mar_30_fit.sem_simulation}()},
\code{\link{mvn_mar_30_fit.sem_task_summary}()},
\code{\link{mvn_mar_30_fit.sem_task}()},
\code{\link{mvn_mar_30_fit.sem}()},
\code{\link{mvn_mcar_10_fit.sem_simulation_summary}()},
\code{\link{mvn_mcar_10_fit.sem_simulation}()},
\code{\link{mvn_mcar_10_fit.sem_task_summary}()},
\code{\link{mvn_mcar_10_fit.sem_task}()},
\code{\link{mvn_mcar_10_fit.sem}()},
\code{\link{mvn_mcar_20_fit.sem_simulation_summary}()},
\code{\link{mvn_mcar_20_fit.sem_simulation}()},
\code{\link{mvn_mcar_20_fit.sem_task_summary}()},
\code{\link{mvn_mcar_20_fit.sem_task}()},
\code{\link{mvn_mcar_20_fit.sem}()},
\code{\link{mvn_mcar_30_fit.sem_simulation_summary}()},
\code{\link{mvn_mcar_30_fit.sem_simulation}()},
\code{\link{mvn_mcar_30_fit.sem_task_summary}()},
\code{\link{mvn_mcar_30_fit.sem_task}()},
\code{\link{mvn_mcar_30_fit.sem}()},
\code{\link{mvn_mnar_10_fit.sem_simulation_summary}()},
\code{\link{mvn_mnar_10_fit.sem_simulation}()},
\code{\link{mvn_mnar_10_fit.sem_task_summary}()},
\code{\link{mvn_mnar_10_fit.sem_task}()},
\code{\link{mvn_mnar_10_fit.sem}()},
\code{\link{mvn_mnar_20_fit.sem_simulation_summary}()},
\code{\link{mvn_mnar_20_fit.sem_simulation}()},
\code{\link{mvn_mnar_20_fit.sem_task_summary}()},
\code{\link{mvn_mnar_20_fit.sem_task}()},
\code{\link{mvn_mnar_20_fit.sem}()},
\code{\link{mvn_mnar_30_fit.sem_simulation_summary}()},
\code{\link{mvn_mnar_30_fit.sem_simulation}()},
\code{\link{mvn_mnar_30_fit.sem_task_summary}()},
\code{\link{mvn_mnar_30_fit.sem_task}()},
\code{\link{mvn_mnar_30_fit.sem}()},
\code{\link{mvn_std_fit.sem_simulation_summary}()},
\code{\link{mvn_std_fit.sem_simulation}()},
\code{\link{mvn_std_fit.sem_task_summary}()},
\code{\link{mvn_std_fit.sem_task}()},
\code{\link{mvn_std_fit.sem}()},
\code{\link{vm_mod_fit.ols_simulation_summary}()},
\code{\link{vm_mod_fit.ols_simulation}()},
\code{\link{vm_mod_fit.ols_task_summary}()},
\code{\link{vm_mod_fit.ols_task}()},
\code{\link{vm_mod_fit.ols}()},
\code{\link{vm_mod_fit.sem.mlr_simulation_summary}()},
\code{\link{vm_mod_fit.sem.mlr_simulation}()},
\code{\link{vm_mod_fit.sem.mlr_task_summary}()},
\code{\link{vm_mod_fit.sem.mlr_task}()},
\code{\link{vm_mod_fit.sem.mlr}()},
\code{\link{vm_mod_std_fit.sem.mlr_simulation_summary}()},
\code{\link{vm_mod_std_fit.sem.mlr_task_summary}()},
\code{\link{vm_mod_std_fit.sem.mlr_task}()},
\code{\link{vm_mod_std_fit.sem.mlr}()},
\code{\link{vm_sev_fit.ols_simulation_summary}()},
\code{\link{vm_sev_fit.ols_simulation}()},
\code{\link{vm_sev_fit.ols_task_summary}()},
\code{\link{vm_sev_fit.ols_task}()},
\code{\link{vm_sev_fit.ols}()},
\code{\link{vm_sev_fit.sem.mlr_simulation_summary}()},
\code{\link{vm_sev_fit.sem.mlr_simulation}()},
\code{\link{vm_sev_fit.sem.mlr_task_summary}()},
\code{\link{vm_sev_fit.sem.mlr_task}()},
\code{\link{vm_sev_fit.sem.mlr}()},
\code{\link{vm_sev_std_fit.sem.mlr_simulation_summary}()},
\code{\link{vm_sev_std_fit.sem.mlr_simulation}()},
\code{\link{vm_sev_std_fit.sem.mlr_task_summary}()},
\code{\link{vm_sev_std_fit.sem.mlr_task}()},
\code{\link{vm_sev_std_fit.sem.mlr}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{model fit functions}
\keyword{fit}
