% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vm_sev_complete_unstd_sem_mc.mvn.R
\name{vm_sev_sem_mc.mvn_pcci_task}
\alias{vm_sev_sem_mc.mvn_pcci_task}
\title{Monte Carlo Method Confidence Intervals Assuming Multivariate Normal Distribution for Indirect Effect in a Simple Mediation Model
for Data Generated Using the Vale and Maurelli (1983) Approach (Skewness = 3, Kurtosis = 21)
(Single Task)}
\usage{
vm_sev_sem_mc.mvn_pcci_task(taskid, dir = getwd())
}
\arguments{
\item{taskid}{Numeric.
Task ID.}

\item{dir}{Character string.
Directory where results of \verb{vm_sev_dat_*} are stored.}
}
\description{
Monte Carlo Method Confidence Intervals Assuming Multivariate Normal Distribution for Indirect Effect in a Simple Mediation Model
for Data Generated Using the Vale and Maurelli (1983) Approach (Skewness = 3, Kurtosis = 21)
(Single Task)
}
\seealso{
Other monte carlo method functions: 
\code{\link{mc.mvn}()},
\code{\link{mc.t}()},
\code{\link{mc.wishart}()},
\code{\link{mvn_mar_10_mc.mvn_pcci_simulation}()},
\code{\link{mvn_mar_10_mc.mvn_pcci_task}()},
\code{\link{mvn_mar_10_mc.mvn_simulation}()},
\code{\link{mvn_mar_10_mc.mvn_task}()},
\code{\link{mvn_mar_10_mc.mvn}()},
\code{\link{mvn_mar_20_mc.mvn_pcci_simulation}()},
\code{\link{mvn_mar_20_mc.mvn_pcci_task}()},
\code{\link{mvn_mar_20_mc.mvn_simulation}()},
\code{\link{mvn_mar_20_mc.mvn_task}()},
\code{\link{mvn_mar_20_mc.mvn}()},
\code{\link{mvn_mar_30_mc.mvn_pcci_simulation}()},
\code{\link{mvn_mar_30_mc.mvn_pcci_task}()},
\code{\link{mvn_mar_30_mc.mvn_simulation}()},
\code{\link{mvn_mar_30_mc.mvn_task}()},
\code{\link{mvn_mar_30_mc.mvn}()},
\code{\link{mvn_mcar_10_mc.mvn_pcci_simulation}()},
\code{\link{mvn_mcar_10_mc.mvn_pcci_task}()},
\code{\link{mvn_mcar_10_mc.mvn_simulation}()},
\code{\link{mvn_mcar_10_mc.mvn_task}()},
\code{\link{mvn_mcar_10_mc.mvn}()},
\code{\link{mvn_mcar_20_mc.mvn_pcci_simulation}()},
\code{\link{mvn_mcar_20_mc.mvn_pcci_task}()},
\code{\link{mvn_mcar_20_mc.mvn_simulation}()},
\code{\link{mvn_mcar_20_mc.mvn_task}()},
\code{\link{mvn_mcar_20_mc.mvn}()},
\code{\link{mvn_mcar_30_mc.mvn_pcci_simulation}()},
\code{\link{mvn_mcar_30_mc.mvn_pcci_task}()},
\code{\link{mvn_mcar_30_mc.mvn_simulation}()},
\code{\link{mvn_mcar_30_mc.mvn_task}()},
\code{\link{mvn_mcar_30_mc.mvn}()},
\code{\link{mvn_mnar_10_mc.mvn_pcci_simulation}()},
\code{\link{mvn_mnar_10_mc.mvn_pcci_task}()},
\code{\link{mvn_mnar_10_mc.mvn_simulation}()},
\code{\link{mvn_mnar_10_mc.mvn_task}()},
\code{\link{mvn_mnar_10_mc.mvn}()},
\code{\link{mvn_mnar_20_mc.mvn_pcci_simulation}()},
\code{\link{mvn_mnar_20_mc.mvn_pcci_task}()},
\code{\link{mvn_mnar_20_mc.mvn_simulation}()},
\code{\link{mvn_mnar_20_mc.mvn_task}()},
\code{\link{mvn_mnar_20_mc.mvn}()},
\code{\link{mvn_mnar_30_mc.mvn_pcci_simulation}()},
\code{\link{mvn_mnar_30_mc.mvn_pcci_task}()},
\code{\link{mvn_mnar_30_mc.mvn_simulation}()},
\code{\link{mvn_mnar_30_mc.mvn_task}()},
\code{\link{mvn_mnar_30_mc.mvn}()},
\code{\link{mvn_ols_mc.mvn_pcci_simulation}()},
\code{\link{mvn_ols_mc.mvn_pcci_task}()},
\code{\link{mvn_ols_mc.mvn_simulation}()},
\code{\link{mvn_ols_mc.mvn_task}()},
\code{\link{mvn_ols_mc.mvn}()},
\code{\link{mvn_sem_mc.mvn_pcci_simulation}()},
\code{\link{mvn_sem_mc.mvn_pcci_task}()},
\code{\link{mvn_sem_mc.mvn_simulation}()},
\code{\link{mvn_sem_mc.mvn_task}()},
\code{\link{mvn_sem_mc.mvn}()},
\code{\link{mvn_std_mc.mvn.delta_pcci_simulation}()},
\code{\link{mvn_std_mc.mvn.delta_pcci_task}()},
\code{\link{mvn_std_mc.mvn.delta_simulation}()},
\code{\link{mvn_std_mc.mvn.delta_task}()},
\code{\link{mvn_std_mc.mvn.delta}()},
\code{\link{mvn_std_mc.mvn.sem_pcci_simulation}()},
\code{\link{mvn_std_mc.mvn.sem_pcci_task}()},
\code{\link{mvn_std_mc.mvn.sem_simulation}()},
\code{\link{mvn_std_mc.mvn.sem_task}()},
\code{\link{mvn_std_mc.mvn.sem}()},
\code{\link{mvn_std_mc.mvn.tb_pcci_simulation}()},
\code{\link{mvn_std_mc.mvn.tb_pcci_task}()},
\code{\link{mvn_std_mc.mvn.tb_simulation}()},
\code{\link{mvn_std_mc.mvn.tb_task}()},
\code{\link{mvn_std_mc.mvn.tb}()},
\code{\link{mvn_std_mc.wishart_pcci_simulation}()},
\code{\link{mvn_std_mc.wishart_pcci_task}()},
\code{\link{mvn_std_mc.wishart_simulation}()},
\code{\link{mvn_std_mc.wishart_task}()},
\code{\link{mvn_std_mc.wishart}()},
\code{\link{vm_mod_ols_mc.mvn_pcci_simulation}()},
\code{\link{vm_mod_ols_mc.mvn_pcci_task}()},
\code{\link{vm_mod_ols_mc.mvn_simulation}()},
\code{\link{vm_mod_ols_mc.mvn_task}()},
\code{\link{vm_mod_ols_mc.mvn}()},
\code{\link{vm_mod_sem_mc.mvn_pcci_simulation}()},
\code{\link{vm_mod_sem_mc.mvn_pcci_task}()},
\code{\link{vm_mod_sem_mc.mvn_simulation}()},
\code{\link{vm_mod_sem_mc.mvn_task}()},
\code{\link{vm_mod_sem_mc.mvn}()},
\code{\link{vm_sev_ols_mc.mvn_pcci_simulation}()},
\code{\link{vm_sev_ols_mc.mvn_pcci_task}()},
\code{\link{vm_sev_ols_mc.mvn_simulation}()},
\code{\link{vm_sev_ols_mc.mvn_task}()},
\code{\link{vm_sev_ols_mc.mvn}()},
\code{\link{vm_sev_sem_mc.mvn_pcci_simulation}()},
\code{\link{vm_sev_sem_mc.mvn_simulation}()},
\code{\link{vm_sev_sem_mc.mvn_task}()},
\code{\link{vm_sev_sem_mc.mvn}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{monte carlo method functions}
\keyword{mc}
