% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval.R
\name{zero_hit}
\alias{zero_hit}
\title{Confidence Interval - Zero Hit}
\usage{
zero_hit(lo, up)
}
\arguments{
\item{lo}{Numeric.
Lower limit of the estimated confidence interval
\eqn{\left( \hat{\theta}_{\mathrm{lo}} \right)}.}

\item{up}{Numeric.
Upper limit of the estimated confidence interval
\eqn{\left( \hat{\theta}_{\mathrm{up}} \right)}.}
}
\value{
Returns
\code{TRUE} if zero is between the interval
\code{lo}
\eqn{\left( \hat{\theta}_{\mathrm{lo}} \right)}
to
\code{up}
\eqn{\left( \hat{\theta}_{\mathrm{up}} \right)}.
Returns
\code{FALSE} if zero is outside the interval
\code{lo}
\eqn{\left( \hat{\theta}_{\mathrm{lo}} \right)}
to
\code{up}
\eqn{\left( \hat{\theta}_{\mathrm{up}} \right)}.
}
\description{
Confidence Interval - Zero Hit
}
\examples{
# FALSE
zero_hit(lo = 1, up = 2)
# TRUE
zero_hit(lo = -1, up = 1)
}
\seealso{
Other confidence intervals functions: 
\code{\link{bcaci}()},
\code{\link{bcci}()},
\code{\link{evalci}()},
\code{\link{jackknife}()},
\code{\link{len}()},
\code{\link{pcci}()},
\code{\link{shape}()},
\code{\link{theta_hit}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{confidence intervals functions}
\keyword{ci}
