% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci.R
\name{bcaci}
\alias{bcaci}
\title{Bias-Corrected and Accelerated Confidence Intervals}
\usage{
bcaci(
  thetahatstar,
  thetahat,
  theta = NULL,
  data,
  std = FALSE,
  alpha = c(0.001, 0.01, 0.05)
)
}
\arguments{
\item{thetahatstar}{Numeric vector.
Sampling distribution of thetahat.}

\item{thetahat}{Numeric.
Parameter estimate.}

\item{theta}{Numeric.
Parameter.
Optional argument.}

\item{data}{\code{n} by 3 matrix or data frame.
\code{data[, 1]} correspond to values for \code{x}.
\code{data[, 2]} correspond to values for \code{m}.
\code{data[, 3]} correspond to values for \code{y}.}

\item{std}{Logical.
Standardize the indirect effect
\eqn{
    \hat{\alpha}^{\prime} \hat{\beta}^{\prime}
    = \hat{\alpha} \hat{\beta} \frac{\hat{\sigma}_x}{\hat{\sigma}_y}}.}

\item{alpha}{Numeric vector.
Alpha level.
By default \code{alpha = c(0.001, 0.01, 0.05)}.}
}
\description{
Bias-Corrected and Accelerated Confidence Intervals
}
\examples{
B <- 5000
data <- jeksterslabRdatarepo::thirst
thetahat <- .fit(data, minimal = TRUE)
thetahatstar <- .pbmvn(data = data, B = 5000, par = FALSE)
bcaci(
  thetahatstar = thetahatstar,
  thetahat = thetahat,
  theta = 0.15, # assuming that the true indirect effect is 0.15
  data = data
)
thetahat <- .fit(data, minimal = TRUE, std = TRUE)
thetahatstar <- .pbmvn(data = data, std = TRUE, B = 5000, par = FALSE)
bcaci(
  thetahatstar = thetahatstar,
  thetahat = thetahat,
  theta = 0.15, # assuming that the true indirect effect is 0.15
  data = data,
  std = TRUE
)
}
\seealso{
Other confidence intervals functions: 
\code{\link{bcci}()},
\code{\link{evalci}()},
\code{\link{jackknife}()},
\code{\link{len}()},
\code{\link{pcci}()},
\code{\link{shape}()},
\code{\link{theta_hit}()},
\code{\link{zero_hit}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{confidence intervals functions}
\keyword{ci}
