% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.cov.R
\name{fit.cov}
\alias{fit.cov}
\title{Fit Saturated Model - Structural Equation Modeling}
\usage{
fit.cov(data)
}
\arguments{
\item{data}{\code{n} by 3 matrix or data frame.
\code{data[, 1]} correspond to values for \code{x}.
\code{data[, 2]} correspond to values for \code{m}.
\code{data[, 3]} correspond to values for \code{y}.}
}
\description{
Fits covariances and means of \eqn{x}, \eqn{m}, and \eqn{y} using structural equation modeling.
}
\examples{
library(lavaan)
taskid <- 1

# Complete Data ----------------------------------------------------
original <- mvn_dat(taskid = taskid)
cov(original)
colMeans(original)

# Missing completely at random -------------------------------------
data <- mvn_dat_mcar(data = original, taskid = taskid)
out <- fit.cov(data = data)
summary(out)

# Missing at random ------------------------------------------------
data <- mvn_dat_mar(data = original, taskid = taskid)
out <- fit.cov(data = data)
summary(out)
}
\seealso{
Other model fit functions: 
\code{\link{fit.ols}()},
\code{\link{fit.sem}()},
\code{\link{mvn_fit.ols_simulation_summary}()},
\code{\link{mvn_fit.ols_simulation}()},
\code{\link{mvn_fit.ols_task_summary}()},
\code{\link{mvn_fit.ols_task}()},
\code{\link{mvn_fit.ols}()},
\code{\link{mvn_fit.sem_simulation_summary}()},
\code{\link{mvn_fit.sem_simulation}()},
\code{\link{mvn_fit.sem_task_summary}()},
\code{\link{mvn_fit.sem_task}()},
\code{\link{mvn_fit.sem}()},
\code{\link{mvn_mar_10_fit.sem_simulation_summary}()},
\code{\link{mvn_mar_10_fit.sem_simulation}()},
\code{\link{mvn_mar_10_fit.sem_task_summary}()},
\code{\link{mvn_mar_10_fit.sem_task}()},
\code{\link{mvn_mar_10_fit.sem}()},
\code{\link{mvn_mar_20_fit.sem_simulation_summary}()},
\code{\link{mvn_mar_20_fit.sem_simulation}()},
\code{\link{mvn_mar_20_fit.sem_task_summary}()},
\code{\link{mvn_mar_20_fit.sem_task}()},
\code{\link{mvn_mar_20_fit.sem}()},
\code{\link{mvn_mar_30_fit.sem_simulation_summary}()},
\code{\link{mvn_mar_30_fit.sem_simulation}()},
\code{\link{mvn_mar_30_fit.sem_task_summary}()},
\code{\link{mvn_mar_30_fit.sem_task}()},
\code{\link{mvn_mar_30_fit.sem}()},
\code{\link{mvn_mar_50_fit.sem_simulation_summary}()},
\code{\link{mvn_mar_50_fit.sem_simulation}()},
\code{\link{mvn_mar_50_fit.sem_task_summary}()},
\code{\link{mvn_mar_50_fit.sem_task}()},
\code{\link{mvn_mar_50_fit.sem}()},
\code{\link{mvn_mar_fit.sem_simulation_summary}()},
\code{\link{mvn_mar_fit.sem_simulation}()},
\code{\link{mvn_mar_fit.sem_task_summary}()},
\code{\link{mvn_mar_fit.sem_task}()},
\code{\link{mvn_mar_fit.sem}()},
\code{\link{mvn_mar_std_fit.sem_simulation_summary}()},
\code{\link{mvn_mar_std_fit.sem_simulation}()},
\code{\link{mvn_mar_std_fit.sem_task_summary}()},
\code{\link{mvn_mar_std_fit.sem_task}()},
\code{\link{mvn_mar_std_fit.sem}()},
\code{\link{mvn_mcar_10_fit.sem_simulation_summary}()},
\code{\link{mvn_mcar_10_fit.sem_simulation}()},
\code{\link{mvn_mcar_10_fit.sem_task_summary}()},
\code{\link{mvn_mcar_10_fit.sem_task}()},
\code{\link{mvn_mcar_10_fit.sem}()},
\code{\link{mvn_mcar_20_fit.sem_simulation_summary}()},
\code{\link{mvn_mcar_20_fit.sem_simulation}()},
\code{\link{mvn_mcar_20_fit.sem_task_summary}()},
\code{\link{mvn_mcar_20_fit.sem_task}()},
\code{\link{mvn_mcar_20_fit.sem}()},
\code{\link{mvn_mcar_30_fit.sem_simulation_summary}()},
\code{\link{mvn_mcar_30_fit.sem_simulation}()},
\code{\link{mvn_mcar_30_fit.sem_task_summary}()},
\code{\link{mvn_mcar_30_fit.sem_task}()},
\code{\link{mvn_mcar_30_fit.sem}()},
\code{\link{mvn_mcar_50_fit.sem_simulation_summary}()},
\code{\link{mvn_mcar_50_fit.sem_simulation}()},
\code{\link{mvn_mcar_50_fit.sem_task_summary}()},
\code{\link{mvn_mcar_50_fit.sem_task}()},
\code{\link{mvn_mcar_50_fit.sem}()},
\code{\link{mvn_mcar_fit.sem_simulation_summary}()},
\code{\link{mvn_mcar_fit.sem_simulation}()},
\code{\link{mvn_mcar_fit.sem_task_summary}()},
\code{\link{mvn_mcar_fit.sem_task}()},
\code{\link{mvn_mcar_fit.sem}()},
\code{\link{mvn_mcar_std_fit.sem_simulation_summary}()},
\code{\link{mvn_mcar_std_fit.sem_simulation}()},
\code{\link{mvn_mcar_std_fit.sem_task_summary}()},
\code{\link{mvn_mcar_std_fit.sem_task}()},
\code{\link{mvn_mcar_std_fit.sem}()},
\code{\link{mvn_mnar_10_fit.sem_simulation_summary}()},
\code{\link{mvn_mnar_10_fit.sem_simulation}()},
\code{\link{mvn_mnar_10_fit.sem_task_summary}()},
\code{\link{mvn_mnar_10_fit.sem_task}()},
\code{\link{mvn_mnar_10_fit.sem}()},
\code{\link{mvn_mnar_20_fit.sem_simulation_summary}()},
\code{\link{mvn_mnar_20_fit.sem_simulation}()},
\code{\link{mvn_mnar_20_fit.sem_task_summary}()},
\code{\link{mvn_mnar_20_fit.sem_task}()},
\code{\link{mvn_mnar_20_fit.sem}()},
\code{\link{mvn_mnar_30_fit.sem_simulation_summary}()},
\code{\link{mvn_mnar_30_fit.sem_simulation}()},
\code{\link{mvn_mnar_30_fit.sem_task_summary}()},
\code{\link{mvn_mnar_30_fit.sem_task}()},
\code{\link{mvn_mnar_30_fit.sem}()},
\code{\link{mvn_mnar_50_fit.sem_simulation_summary}()},
\code{\link{mvn_mnar_50_fit.sem_simulation}()},
\code{\link{mvn_mnar_50_fit.sem_task_summary}()},
\code{\link{mvn_mnar_50_fit.sem_task}()},
\code{\link{mvn_mnar_50_fit.sem}()},
\code{\link{mvn_mnar_fit.sem_simulation_summary}()},
\code{\link{mvn_mnar_fit.sem_simulation}()},
\code{\link{mvn_mnar_fit.sem_task_summary}()},
\code{\link{mvn_mnar_fit.sem_task}()},
\code{\link{mvn_mnar_fit.sem}()},
\code{\link{mvn_mnar_std_fit.sem_simulation_summary}()},
\code{\link{mvn_mnar_std_fit.sem_simulation}()},
\code{\link{mvn_mnar_std_fit.sem_task_summary}()},
\code{\link{mvn_mnar_std_fit.sem_task}()},
\code{\link{mvn_mnar_std_fit.sem}()},
\code{\link{mvn_std_fit.sem_simulation_summary}()},
\code{\link{mvn_std_fit.sem_simulation}()},
\code{\link{mvn_std_fit.sem_task_summary}()},
\code{\link{mvn_std_fit.sem_task}()},
\code{\link{mvn_std_fit.sem}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{model fit functions}
\keyword{fit}
