% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label.R
\name{label}
\alias{label}
\title{Generate Labels}
\usage{
label(out, method, model, std, missing = "NON")
}
\arguments{
\item{out}{Object.
Output of \verb{*_summary}.}

\item{method}{Character string.
Name of the method used.}

\item{model}{Character string.
Model used.}

\item{std}{Logical.
If \code{TRUE}, standardized indirect effect.
If \code{FALSE}, unstandardized indirect effect.}

\item{missing}{Character string.
Type of missingness.
\code{missing = "NON"} for complete data.
\code{missing = "MCAR"} for missing completely at random.
\code{missingess = "MAR"} for missing at random.}
}
\description{
Generate Labels
}
\seealso{
Other process results functions: 
\code{\link{process}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{process results functions}
\keyword{process}
