% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvn.R
\name{mvn}
\alias{mvn}
\title{Generate Data from a Multivariate Normal Distribution}
\usage{
mvn(n, mu, Sigma)
}
\arguments{
\item{n}{Integer.
Samples size.}

\item{mu}{Numeric vector.
Mean vector.}

\item{Sigma}{Numeric matrix.
Variance-covariance matrix.}
}
\description{
Generate Data from a Multivariate Normal Distribution
}
\examples{
n <- 100
mu <- c(0, 0)
Sigma <- matrix(data = c(1, 0.50, 0.50, 1), ncol = 2)
mvn(n = n, mu = mu, Sigma = Sigma)
}
\seealso{
Other multivariate normal data functions: 
\code{\link{mvn_dat_simulation}()},
\code{\link{mvn_dat_task}()},
\code{\link{mvn_dat}()},
\code{\link{setparamsmvn}()},
\code{\link{useparamsmvn}()},
\code{\link{vm_mod_dat_simulation_simulation_summary}()},
\code{\link{vm_mod_dat_simulation_task_summary}()},
\code{\link{vm_mod_dat_simulation}()},
\code{\link{vm_mod_dat_task}()},
\code{\link{vm_mod_dat}()},
\code{\link{vm_sev_dat_simulation_simulation_summary}()},
\code{\link{vm_sev_dat_simulation_task_summary}()},
\code{\link{vm_sev_dat_simulation}()},
\code{\link{vm_sev_dat_task}()},
\code{\link{vm_sev_dat}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{multivariate normal data functions}
\keyword{mvn}
