% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvn.R
\name{mvn}
\alias{mvn}
\title{Generate Data from a Multivariate Normal Distribution}
\usage{
mvn(n, mu, Sigma)
}
\arguments{
\item{n}{Integer.
Samples size.}

\item{mu}{Numeric vector.
Mean vector.}

\item{Sigma}{Numeric matrix.
Variance-covariance matrix.}
}
\description{
Generate Data from a Multivariate Normal Distribution
}
\examples{
n <- 100
mu <- c(0, 0)
Sigma <- matrix(data = c(1, 0.50, 0.50, 1), ncol = 2)
mvn(n = n, mu = mu, Sigma = Sigma)
}
\seealso{
Other multivariate normal data functions: 
\code{\link{mvn_dat_mar_10_simulation}()},
\code{\link{mvn_dat_mar_10_task}()},
\code{\link{mvn_dat_mar_10}()},
\code{\link{mvn_dat_mar_20_simulation}()},
\code{\link{mvn_dat_mar_20_task}()},
\code{\link{mvn_dat_mar_20}()},
\code{\link{mvn_dat_mar_30_simulation}()},
\code{\link{mvn_dat_mar_30_task}()},
\code{\link{mvn_dat_mar_30}()},
\code{\link{mvn_dat_mar_50_simulation}()},
\code{\link{mvn_dat_mar_50_task}()},
\code{\link{mvn_dat_mar_50}()},
\code{\link{mvn_dat_mar_simulation}()},
\code{\link{mvn_dat_mar_task}()},
\code{\link{mvn_dat_mar}()},
\code{\link{mvn_dat_mcar_10_simulation}()},
\code{\link{mvn_dat_mcar_10_task}()},
\code{\link{mvn_dat_mcar_10}()},
\code{\link{mvn_dat_mcar_20_simulation}()},
\code{\link{mvn_dat_mcar_20_task}()},
\code{\link{mvn_dat_mcar_20}()},
\code{\link{mvn_dat_mcar_30_simulation}()},
\code{\link{mvn_dat_mcar_30_task}()},
\code{\link{mvn_dat_mcar_30}()},
\code{\link{mvn_dat_mcar_50_simulation}()},
\code{\link{mvn_dat_mcar_50_task}()},
\code{\link{mvn_dat_mcar_50}()},
\code{\link{mvn_dat_mcar_simulation}()},
\code{\link{mvn_dat_mcar_task}()},
\code{\link{mvn_dat_mcar}()},
\code{\link{mvn_dat_mnar_10_simulation}()},
\code{\link{mvn_dat_mnar_10_task}()},
\code{\link{mvn_dat_mnar_10}()},
\code{\link{mvn_dat_mnar_20_simulation}()},
\code{\link{mvn_dat_mnar_20_task}()},
\code{\link{mvn_dat_mnar_20}()},
\code{\link{mvn_dat_mnar_30_simulation}()},
\code{\link{mvn_dat_mnar_30_task}()},
\code{\link{mvn_dat_mnar_30}()},
\code{\link{mvn_dat_mnar_50_simulation}()},
\code{\link{mvn_dat_mnar_50_task}()},
\code{\link{mvn_dat_mnar_50}()},
\code{\link{mvn_dat_mnar_simulation}()},
\code{\link{mvn_dat_mnar_task}()},
\code{\link{mvn_dat_mnar}()},
\code{\link{mvn_dat_simulation}()},
\code{\link{mvn_dat_task}()},
\code{\link{mvn_dat}()},
\code{\link{setparamsmvn}()},
\code{\link{useparamsmvn}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{multivariate normal data functions}
\keyword{mvn}
