% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvn_complete_dat.R
\name{mvn_dat}
\alias{mvn_dat}
\title{Generate Data from a Multivariate Normal Distribution
following a Simple Mediation Model}
\usage{
mvn_dat(taskid)
}
\arguments{
\item{taskid}{Numeric.
Task ID.}
}
\description{
Generate Data from a Multivariate Normal Distribution
following a Simple Mediation Model
}
\examples{
data <- mvn_dat(taskid = 1)
fit.ols(data, minimal = TRUE)
fit.ols(data, minimal = TRUE, std = TRUE)
fit.ols(data, minimal = FALSE)
}
\seealso{
Other multivariate normal data functions: 
\code{\link{mvn_dat_simulation}()},
\code{\link{mvn_dat_task}()},
\code{\link{mvn}()},
\code{\link{setparamsmvn}()},
\code{\link{useparamsmvn}()},
\code{\link{vm_mod_dat_simulation_simulation_summary}()},
\code{\link{vm_mod_dat_simulation_task_summary}()},
\code{\link{vm_mod_dat_simulation}()},
\code{\link{vm_mod_dat_task}()},
\code{\link{vm_mod_dat}()},
\code{\link{vm_sev_dat_simulation_simulation_summary}()},
\code{\link{vm_sev_dat_simulation_task_summary}()},
\code{\link{vm_sev_dat_simulation}()},
\code{\link{vm_sev_dat_task}()},
\code{\link{vm_sev_dat}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{multivariate normal data functions}
\keyword{mvn_dat}
