% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvn.R
\name{mvn_dat}
\alias{mvn_dat}
\title{Generate Data from a Multivariate Normal Distribution}
\usage{
mvn_dat(taskid)
}
\arguments{
\item{taskid}{Numeric.
Task ID.}
}
\description{
Generate Data from a Multivariate Normal Distribution
}
\examples{
data <- mvn_dat(taskid = 1)
.fit(data, minimal = TRUE)
.fit(data, minimal = TRUE, std = TRUE)
.fit(data, minimal = FALSE)
}
\seealso{
Other multivariate normal data functions: 
\code{\link{mvn_dat_simulation}()},
\code{\link{mvn_dat_task}()},
\code{\link{mvn}()},
\code{\link{setparamsmvn}()},
\code{\link{useparamsmvn}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{multivariate normal data functions}
\keyword{mvn}
