% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvn_mar.R
\name{mvn_dat_mar}
\alias{mvn_dat_mar}
\title{Generate Data with MAR from Data generated from a Multivariate Normal Distribution - Simple Mediation Model}
\usage{
mvn_dat_mar(data, taskid)
}
\arguments{
\item{data}{\code{n} by 3 matrix or data frame.
\code{data[, 1]} correspond to values for \code{x}.
\code{data[, 2]} correspond to values for \code{m}.
\code{data[, 3]} correspond to values for \code{y}.}

\item{taskid}{Numeric.
Task ID.}
}
\description{
Generate Data with MAR from Data generated from a Multivariate Normal Distribution - Simple Mediation Model
}
\seealso{
Other multivariate normal data functions: 
\code{\link{mvn_dat_mar_simulation}()},
\code{\link{mvn_dat_mar_task}()},
\code{\link{mvn_dat_mcar_simulation}()},
\code{\link{mvn_dat_mcar_task}()},
\code{\link{mvn_dat_mcar}()},
\code{\link{mvn_dat_simulation}()},
\code{\link{mvn_dat_task}()},
\code{\link{mvn_dat}()},
\code{\link{mvn}()},
\code{\link{setparamsmvn}()},
\code{\link{useparamsmvn}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{multivariate normal data functions}
\keyword{mvn}
