% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{process}
\alias{process}
\title{Process Task Results}
\usage{
process(taskid, out)
}
\arguments{
\item{out}{Object.
Output of \verb{*_task}.}
}
\description{
Process Task Results
}
\seealso{
Other process results functions: 
\code{\link{label}()}
}
\author{
Ivan Jacob Agaloos Pesign
}
\concept{process results functions}
\keyword{process}
